% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmodR-analysis-type.R
\docType{methods}
\name{analysis-accessors}
\alias{analysis-accessors}
\alias{getPositions}
\alias{getModifications}
\alias{getPlotType,analysis-method}
\alias{getPositions,analysis-method}
\alias{getModifications,analysis-method}
\title{Accessor for \code{analysis} class objects}
\usage{
getPlotType(object)

getPositions(object)

getModifications(object)

\S4method{getPlotType}{analysis}(object)

\S4method{getPositions}{analysis}(object)

\S4method{getModifications}{analysis}(object)
}
\arguments{
\item{object}{analysis object}
}
\value{
character defining the plot type for this analysis class

a list of data as a lists of list(replicate) of DataFrame(transcript)

a list of data as a lists of list(replicate) of DataFrame(transcript)
}
\examples{
\donttest{
getPlotType(analysis)
getPositions(analysis)
getModifications(analysis)
}
}
