% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RNAmodR-analysis-type.R
\docType{class}
\name{analysis}
\alias{analysis}
\title{analysis}
\description{
The class is virtual and has to be extended from, eg. \code{analysis_default}
 
The analysis class is the main function for detecting modifications. It works
usually in a three step manner.
\itemize{
  \item{"1."}{\code{convertReadsToPositions}: This function aggregates the
  information of reads per bam file, eg. per replicate and stores them for
  subsequent analysis}
  \item{"2."}{\code{parseMod}: This function searches for a defined pattern. 
  For this \code{mod} class objects will be created on the fly based on the 
  names of the modification one wants to detect and which can understand the
  data aggregated by \code{convertReadsToPositions}.}
  \item{"3."}{\code{mergePositionsOfReplicates}: This function aggregates the
  base data further for storage. This involves usually generating a per
  position mean and sd or equivalent.}
}
}
