% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RNAmodR-analysis-default.R
\docType{class}
\name{analysis_default}
\alias{analysis_default}
\title{the default analysis class for RNAmodR}
\description{
\code{analysis_default}: the class can be used for analyzing the 5'-end of
reads to detect specific post-transcriptional modifications.
}
\examples{
ad <- new("analysis_default")
}
