% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR-utils-preparations.R
\name{appendFastaFiles}
\alias{appendFastaFiles}
\alias{appendGFF}
\alias{appendGFF}
\title{Appending multiple fasta or gff3 files}
\usage{
appendFastaFiles(..., ident, msg)

appendGFF(..., ident, msg)
}
\arguments{
\item{...}{any character will interpreted as file name used for appending}

\item{ident}{a string used for naming the output file}

\item{msg}{a message to be displayed}
}
\value{
file name of the saved file
}
\description{
append one or more fasta or gff files
}
\examples{
\donttest{
appendFastaFiles(fafile, 
                 fafile2, 
                 ident = "test", 
                 msg = "Test fasta file saved: ")
appendGFF(gfffile, 
          gfffile2, 
          ident = "test", 
          msg = "Test gff3 file saved: ")
}
\donttest{
appendGFF("file1.gff",
          "file2.gff",
          "test",
          "done")
}
}
