% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTxMethods.R
\name{BuildTxTest}
\alias{BuildTxTest}
\title{Build a test transcriptome.}
\usage{
BuildTxTest()
}
\description{
Build a test transcriptome. See 'Details'.
}
\details{
The function builds a transcriptome for testing purposes.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}

