% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{CalculateGC}
\alias{CalculateGC}
\title{Calculate GC content within window around loci from a \code{txLoc}
object.}
\usage{
CalculateGC(locus, flank = 10)
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{flank}{An integer scalar; see 'Details'.}
}
\description{
Calculate GC content within window around loci from a \code{txLoc}
object. See 'Details'.
}
\details{
The function calculates the GC content within a region around every
transcript locus from a \code{txLoc} object. The window is defined
by extending the position of every transcript locus upstream and
downstream by \code{flank} nucleotides (if possible).
}

