% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{CheckClass}
\alias{CheckClass}
\title{Check if object has class classType.}
\usage{
CheckClass(object, classType = NULL, classType2 = NULL)
}
\arguments{
\item{object}{An R object.}

\item{classType}{String of class type for toplevel.}

\item{classType2}{String of class type for level 2.}
}
\value{
Logical. \code{TRUE} if object is of class classType.
}
\description{
Check if object has class classType.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
