% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{CheckClassTxLocConsistency}
\alias{CheckClassTxLocConsistency}
\title{Check if entries of two \code{txLoc} objects are consistent.}
\usage{
CheckClassTxLocConsistency(obj1, obj2)
}
\arguments{
\item{obj1}{A \code{txLoc} object.}

\item{obj2}{A \code{txLoc} object.}
}
\value{
A logical scalar.
}
\description{
Check if entries of two \code{txLoc} objects are consistent.
See 'Details'.
}
\details{
The function checks if entries from two \code{txLoc} objects
are based on the same reference genome _and_ contain entries
for the same transcript regions
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
