% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{CheckClassTxLocRef}
\alias{CheckClassTxLocRef}
\title{Check if entries of two \code{txLoc} objects are based on the
same reference genome.}
\usage{
CheckClassTxLocRef(obj1, obj2)
}
\arguments{
\item{obj1}{A \code{txLoc} object.}

\item{obj2}{A \code{txLoc} object.}
}
\value{
A logical scalar.
}
\description{
Check if entries of two \code{txLoc} objects are based on the
same reference genome.
}
\keyword{internal}

