% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/class-RNAmod-mod-D.R,
%   R/class-RNAmod-mod-m3C.R, R/class-RNAmod-mod-m7G.R
\docType{methods}
\name{checkForModification}
\alias{checkForModification}
\alias{checkForModification,mod_D,numeric,numeric,list,list-method}
\alias{checkForModification,mod_m3C,numeric,numeric,list,list-method}
\alias{checkForModification,mod_m7G,numeric,numeric,list,list-method}
\title{checkForModification}
\usage{
checkForModification(object, location, globalLocations, data, modClasses)

\S4method{checkForModification}{mod_D,numeric,numeric,list,list}(object,
  location, globalLocations, data, modClasses)

\S4method{checkForModification}{mod_m3C,numeric,numeric,list,list}(object,
  location, globalLocations, data, modClasses)

\S4method{checkForModification}{mod_m7G,numeric,numeric,list,list}(object,
  location, globalLocations, data, modClasses)
}
\description{
\code{mod_D}

\code{mod_m3C}

\code{mod_m7G}
}
