% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTxMethods.R
\name{CheckPkgDependencies}
\alias{CheckPkgDependencies}
\title{Check package dependencies.}
\usage{
CheckPkgDependencies(genomeVersion = "hg38")
}
\arguments{
\item{genomeVersion}{A character string; refers to a specific
reference genome assembly version; default is \code{"hg38"}.}
}
\value{
A logical scalar. See 'Details'.
}
\description{
Check package dependencies. See 'Details'.
}
\details{
This is a low-level function that is being called from
\code{BuildTx}. The function checks that the R packages
providing the relevant gene annotations and reference genome
for reference genome \code{genomeVersion} are available
in the current R environment. The function returns \code{TRUE}
if the necessary R packages are installed.
}
\examples{
\dontrun{
CheckPkgDependencies("hg38");
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
