% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTxMethods.R
\name{CollapseTxBySec}
\alias{CollapseTxBySec}
\title{Collapse entries from a \code{GetTxBySec} return object.}
\usage{
CollapseTxBySec(txBySec, geneXID, verbose = FALSE)
}
\arguments{
\item{txBySec}{A \code{list} of \code{GRangesList} objects;
output of function \code{GetTxBySec}.}

\item{geneXID}{A \code{data.frame}; output of function
\code{GetGeneIds} providing gene IDs from different gene
annotations.}

\item{verbose}{A logical scalar; if \code{TRUE} print
additional output; default it \code{FALSE}}
}
\value{
A \code{list} of \code{GRangesList} objects
}
\description{
Collapse entries from a \code{GetTxBySec} return object.
See 'Details'.
This is a low-level function that is being called from
\code{BuildTx}.
}
\details{
Entries are collapsed using the following method:
\itemize{
  \item If multiple entries exist for one Entrez ID, the
transcript with the longest CDS is kept.
  \item For every CDS with a unique RefSeq ID, the
5'/3'UTRs closest to the CDS (non-overlapping) is kept; in
the case of a distance tie, the longest UTR is kept.
  \item For every 5'UTR with a unique RefSeq ID, the
corresponding upstream promoter is chosen.
}
}
\keyword{internal}

