% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ModifierSet-comparison.R
\docType{methods}
\name{compare}
\alias{compare}
\alias{compareByCoord}
\alias{plotCompare}
\alias{plotCompareByCoord}
\alias{compareByCoord,ModifierSet,GRanges-method}
\alias{compareByCoord,ModifierSet,GRangesList-method}
\alias{plotCompareByCoord,ModifierSet,GRanges-method}
\alias{plotCompareByCoord,ModifierSet,GRangesList-method}
\title{compare}
\usage{
compare(x, name, from = 1L, to = 30L, ...)

compareByCoord(x, coord, ...)

plotCompare(x, name, from = 1L, to = 30L, normalize, ...)

plotCompareByCoord(x, coord, normalize, ...)

\S4method{compareByCoord}{ModifierSet,GRanges}(x, coord, ...)

\S4method{compareByCoord}{ModifierSet,GRangesList}(x, coord, ...)

\S4method{plotCompareByCoord}{ModifierSet,GRanges}(x, coord, normalize,
  ...)

\S4method{plotCompareByCoord}{ModifierSet,GRangesList}(x, coord, normalize,
  ...)
}
\arguments{
\item{x}{a \code{Modifier} or \code{ModifierSet} object.}

\item{...}{optional parameters:
\itemize{
\item{\code{name}}{Limit results to one specific gene or transcript}
\item{...}{passed on to \code{\link{subsetByCoord}}}
}}

\item{coord}{coordinates of position to subset to. Either a \code{GRanges} or
a \code{GRangesList} object. For both types the Parent column is expected to
match the gene or transcript name.}

\item{normalize}{either a single logical or character value. If it is a 
character, it must match one of the names in the \code{ModifierSet}.}
}
\value{
for \code{compareByCoord} as \code{\link{DataFrameList}} and for
\code{plotCompareByCoord} a \code{ggplot} object, which can be modified 
further.
}
\description{
title
}
