% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/class-RNAmodR-data-type.R
\docType{methods}
\name{convertReadsToData}
\alias{convertReadsToData}
\alias{convertReadsToData,data,character,character,GRanges,ScanBamParam-method}
\title{Convert reads to data for search for modifications}
\usage{
convertReadsToData(x, files, conditions, gff, param)

\S4method{convertReadsToData}{data,character,character,GRanges,ScanBamParam}(x,
  files, conditions, gff, param)
}
\arguments{
\item{files}{files which to analyze}

\item{gff}{a GRanges object for the genome}

\item{param}{ScanBamParam to use for loading the bam files}

\item{object}{data object}
}
\value{
a modifified \code{data} class object containing now the read 
data required for modification detection.
}
\description{
load the input files, eg. bam files andpre-processes the information to be 
used in the modification detection.
}
