% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{counts}
\alias{counts}
\title{Number of counts for example data}
\format{An object of class \code{numeric} of length 1.}
\usage{
counts
}
\description{
example data - total number of counts
}
\keyword{RNAmod}
\keyword{counts}
\keyword{datasets}
