% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-coverage.R
\docType{class}
\name{CoverageSequenceData}
\alias{CoverageSequenceData}
\alias{CoverageSequenceData-class}
\title{CoverageSequenceData}
\usage{
CoverageSequenceData(bamfiles, annotation, sequences, seqinfo, ...)
}
\description{
title
}
