% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmodR-data-type.R
\docType{methods}
\name{data-accessors}
\alias{data-accessors}
\alias{getPositions}
\alias{getModifications}
\alias{getDataLabel}
\alias{getDataFormat}
\alias{getPlotType,data-method}
\alias{getPlotType}
\alias{getPositions}
\alias{getModifications}
\alias{getPositions,data-method}
\alias{getModifications,data-method}
\alias{getDataLabel,data-method}
\alias{getDataFormat,data-method}
\title{Accessor for \code{data} class objects}
\usage{
getPlotType(object)

getPositions(object)

getModifications(object)

getDataLabel(object)

getDataFormat(object)

\S4method{getPlotType}{data}(object)

\S4method{getPositions}{data}(object)

\S4method{getModifications}{data}(object)

\S4method{getDataLabel}{data}(object)

\S4method{getDataFormat}{data}(object)
}
\arguments{
\item{object}{an data object}
}
\value{
character defining the plot type for this data class

a list of data as a lists of list(replicate) of DataFrame(transcript)

a list of data as a lists of list(replicate) of DataFrame(transcript)

the description of the type of position data, which can be used eg. 
as a label for printing

a function for formating the label of the position data
}
\description{
The accessor function to \code{data} class objects can be used to access
the data saved in slots of the object. See examples for available functions.
}
\examples{
\donttest{
getPlotType(data)
getPositions(data)
getModifications(data)
}
}
