% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTxMethods.R
\name{DedupeBasedOnNearestRef}
\alias{DedupeBasedOnNearestRef}
\title{De-duplicate \code{GRangesList} entries based on distance.}
\usage{
DedupeBasedOnNearestRef(query, ref, showPb = FALSE)
}
\arguments{
\item{query}{A \code{GRangesList} object; the query features}

\item{ref}{A \code{GRangesList} object; the reference features}

\item{showPb}{A logical scalar; if \code{TRUE} show a progress
bar.}
}
\value{
A \code{GRangesList} object.
}
\description{
De-duplicate \code{GRangesList} entries based on distance.
See 'Details'.
This is a low-level function that is being called from
\code{CollapseTxBySec}.
}
\details{
For two \code{GRangesList} objects select entries from
\code{query} that have an entry in \code{ref} and collapse
multiple entries with the same name from \code{query} based
on the minimum distance to the single entry from \code{ref}.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}

