% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{DownsampleTxLoc}
\alias{DownsampleTxLoc}
\title{Downsample a \code{txLoc} object.}
\usage{
DownsampleTxLoc(txLoc1, txLoc2, seed = NULL)
}
\arguments{
\item{txLoc1}{A \code{txLoc} object; this is the \code{txLoc} object that
will be downsampled.}

\item{txLoc2}{A \code{txLoc} object; this is the \code{txLoc} object that
will be used as a target for the downsampling.}

\item{seed}{A single value, interpreted as an \code{integer}, or \code{NULL};
this is to ensure reproducibility when subsampling \code{txLoc2} sites;
default is \code{NULL}.}
}
\value{
A \code{txLoc} object.
}
\description{
Downsample a \code{txLoc1} object based on the number of sites per region
from a \code{txLoc2} object.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
