% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosData-end-pos.R
\docType{class}
\name{EndPosData}
\alias{EndPosData}
\alias{End5PosData-class}
\alias{End3PosData-class}
\alias{EndPosData-class}
\alias{End5PosData}
\alias{End3PosData}
\title{EndPosData}
\usage{
End5PosData(bamfiles, fasta, gff, ...)

End3PosData(bamfiles, fasta, gff, ...)

EndPosData(bamfiles, fasta, gff, ...)
}
\description{
title
}
