% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-end-pos.R
\docType{class}
\name{EndSequenceData}
\alias{EndSequenceData}
\alias{End5SequenceData-class}
\alias{End3SequenceData-class}
\alias{EndSequenceData-class}
\alias{End5SequenceData}
\alias{End3SequenceData}
\title{EndSequenceData}
\usage{
End5SequenceData(bamfiles, fasta, gff, ...)

End3SequenceData(bamfiles, fasta, gff, ...)

EndSequenceData(bamfiles, fasta, gff, ...)
}
\description{
title
}
