% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{EstimateCIrFomBS}
\alias{EstimateCIrFomBS}
\title{Calculate 95% confidence interval from data using empirical
bootstrap.}
\usage{
EstimateCIrFomBS(x, breaks, nBS = 5000)
}
\arguments{
\item{x}{A data vector.}

\item{breaks}{Vector of integer breaks for binning x.}

\item{nBS}{Number of boostrap samples. Default is 5000.}
}
\value{
List of upper and lower 95% confidence interval
bounds for every bin value.
}
\description{
Calculate 95% confidence interval from data using empirical
bootstrap.
}
\keyword{internal}

