% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{FilterTxLoc}
\alias{FilterTxLoc}
\title{Filter sections of a \code{txLoc} object.}
\usage{
FilterTxLoc(txLoc, filter = NULL)
}
\arguments{
\item{txLoc}{A \code{txLoc} object.}

\item{filter}{A \code{character} vector; only keep transcript regions
that match entries from \code{filter}; if \code{NULL}, return the original
\code{txLoc} object.}
}
\value{
A \code{txLoc} object.
}
\description{
Filter sections of a \code{txLoc} object.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
