% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GenerateNull}
\alias{GenerateNull}
\title{Generate a null distribution of transcript sites.}
\usage{
GenerateNull(locus, id = NULL, method = c("ntAbund", "perm"), nt = "C",
  showPb = TRUE)
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{id}{A character string; identifier for null sites; if
\code{NULL} then id = \code{"null"}; default is \code{NULL}.}

\item{method}{A character string; specifies the method used
to generate null distribution; if \code{method == "ntAbund"}
the position of all nucleotides specified by \code{nt} will
be used as null sites; if \code{method == "perm"} then null
sites will be generated by uniform-randomly shuffling candidate
positions from \code{locus} within the corresponding transcript
region; default is \code{"ntAbund"}.}

\item{nt}{A single character; if \code{method == "ntAbund"},
use \code{nt} to derive distribution of null sites.}

\item{showPb}{A logical scalar; if \code{TRUE} show a progress
bar; default is \code{TRUE}.}
}
\value{
A \code{txLoc} object. See 'Details'.
}
\description{
Generate a null distribution of transcript sites. See 'Details'.
}
\details{
The function generates a null distribution of single-nucleotide
sites in transcript sections. Two different methods can be
employed:
\enumerate{
\item If \code{method == "ntAbund"} sites are generated based
on the position of all nucleotides \code{nt} in those transcript
sections that also contain a site in \code{locus}.
\item If \code{method == "perm"} sites are generated based on
sites in \code{locus} by permuting the position of those sites
uniformly across the associated transcript section.
}
It is import to emphasise that any downstream enrichment
analysis may depend critically on the choice of the null
distribution. For example, a position permution-based null
distribution may not be a valid null distribution, if the
position distribution is highly non-uniform across a transcript
section. This is the case e.g. for the spatial distribution of
cytosines across the 5'UTR, CDS and/or 3'UTR. In this case, a
better null distribution would be to consider all cytosines
in transcript regions that contain also a site in \code{locus}.
The return object is a \code{txLoc} of transcript loci in
different transcript sections.
}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR");
sites <- ReadBED(bedFile);
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38");
negSites <- GenerateSNMNull(posSites, method = "permutation");
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

