% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GenerateSNMNull}
\alias{GenerateSNMNull}
\title{Generate a null distribution of transcript sites based on features from locus.}
\usage{
GenerateSNMNull(locus, id = NULL, method = "nuclAbundance",
  nucleotide = "C")
}
\arguments{
\item{locus}{A txLoc object.}

\item{id}{Identifier for null sites. If NULL then id = \code{"null"}.
Default is NULL.}

\item{method}{Method used to generate null distribution.
If \code{method == "nucAbundance"} then the position of all nucleotides
specified by \code{nucleotide} will be used as null sites.
If \code{method == "permutation"} then null sites will be generated
by uniform-randomly shuffling candidate positions from locus within
the corresponding transcript region.
Default is \code{"nucAbundance"}.}

\item{nucleotide}{If \code{method == "nucAbundance"}, use \code{nucleotide}
to derive distribution of null sites.}
}
\value{
A \code{txLoc} object. Note that genome coordinates will not be
available for null sites.
}
\description{
Generate a null distribution of transcript sites based on features from locus.
}
\examples{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR");
sites <- ReadBED(bedFile);
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38");
negSites <- GenerateSNMNull(posSites, method = "permutation");

}

