% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{GetColPal}
\alias{GetColPal}
\title{Return specific colour palette.}
\usage{
GetColPal(pal = c("apple", "google"), n = NULL, alpha = 1)
}
\arguments{
\item{pal}{A character string.}

\item{n}{An integer scalar.}

\item{alpha}{A real scalar.}
}
\value{
A character vector.
}
\description{
Return specific colour palette.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}

