% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetDistanceStartStop}
\alias{GetDistanceStartStop}
\title{Get relative distance of sites to start/stop codon. [NOT WORKING]}
\usage{
GetDistanceStartStop(txLoc, refGenome = "hg38")
}
\arguments{
\item{txLoc}{A \code{txLoc} object.}

\item{refGenome}{A character string; specifies a specific
reference genome assembly version based on which the matching
transcriptome is loaded; default is \code{"hg38"}.}
}
\value{
A \code{GRanges} object. See 'Details'.
}
\description{
Get relative distances of sites from a \code{txLoc} object to the start/stop
codons from a reference transcriptome. See 'Details'.
}
\details{
The function calculates relative distances of sites from a \code{txLoc}
object to the corresponding transcripts start and stop codons.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
