% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{GetDistNearestStartStop}
\alias{GetDistNearestStartStop}
\title{Get distance of sites to the nearest start/stop codon.}
\usage{
GetDistNearestStartStop(txLoc)
}
\arguments{
\item{txLoc}{A \code{txLoc} object.}
}
\value{
A \code{list}. See 'Details'.
}
\description{
Get distance of sites from a \code{txLoc} to the nearest start/stop codon.
}
\details{
The function converts transcript region coordinates from a \code{txLoc} to 
transcript coordinates, and returns distances of sites to the nearest 
start/stop codon. A transcript is defined as the concatenation of
the following transcript regions: 5'UTR, CDS, 3'UTR. The return object is
a \code{list} of distances of sites from \code{txLoc} to the nearest start 
and stop codons.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
