% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetEEJunct}
\alias{GetEEJunct}
\title{Get exon-exon junctions.}
\usage{
GetEEJunct(refGenome = "hg38", filter = "CDS")
}
\arguments{
\item{refGenome}{A character string; specifies a specific
reference genome assembly version based on which the matching
transcriptome is loaded; default is \code{"hg38"}.}

\item{filter}{A character vector; only consider transcript
sections specified in \code{filter}; default is 
\code{"CDS"}.}
}
\value{
A \code{txLoc} object. See 'Details'.
}
\description{
Get exon-exon junctions from transcriptome. See 'Details'.
}
\details{
The function extracts exon-exon junction positions from a
reference transcriptome specified by \code{refGenome}, and 
returns a \code{txLoc} object of exon-exon junctions.
The position of an exon-exon junction is defined as the 
position of the 3'-last nucleotide of an exon followed by
an intron.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
