% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmod.R
\docType{methods}
\name{getExperimentData}
\alias{getExperimentData}
\alias{getExperimentData,RNAmod-method}
\title{getExperimentData}
\usage{
getExperimentData(.Object, number)

\S4method{getExperimentData}{RNAmod}(.Object, number)
}
\arguments{
\item{.Object}{a RNAmod object}

\item{number}{optional number to return results for one experiment only}
}
\value{
a list containg the information on the experiment or a data.frame 
with all the information
}
\examples{
data("RNAmod_test", package = "RNAmod")

# All experiments
getSampleData(experiment)
# One experiment as list
getSampleData(experiment, 1)
# Two experiments as data.frame
getSampleData(experiment, c(1,2))
}
