% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTxMethods.R
\name{GetGeneIds}
\alias{GetGeneIds}
\title{Get different gene IDs.}
\usage{
GetGeneIds(txdb)
}
\arguments{
\item{txdb}{A \code{TxDb} object.}
}
\value{
A \code{data.frame}. See 'Details'.
}
\description{
Get different gene IDs for mapping between different gene
annotations (Ensembl, Entrez, etc.). See 'Details'.
This is a low-level function that is being called from
\code{BuildTx}.
}
\details{
The function extracts the genome assembly version from
the \code{TxDb} object, and loads the suitable genome wide
annotation package. For example, if \code{TxDb} is based
on \code{"hg38"}, the function loads \code{org.Hs.eg.db}.
Various IDs (Ensembl, Unigene, gene symbols) are extracted
from the annotation package by mapping RefSeq IDs from
\code{TxDb} via \code{mapIds} from the \code{AnnotationDbi}
package. Please note that column entries are not necessarily
unique: For example, two different Ensembl IDs might have the
same Entrez ID. The return object is a \code{data.frame} with
the following columns:
\itemize{
  \item REFSEQ: RefSeq Accession number; these entries are unique
  and refer to the mRNA transcript
  \item ENTREZID: Entrez gene ID
  \item SYMBOL: Gene symbol
  \item ENSEMBL: Ensembl gene ID
  \item GENENAME: Gene name
}
Note that mapping between gene IDs is a many-to-many mapping process.
For example, different transcript RefSeq IDs can belong to the same
gene. Which transcripts are associated with which gene depends in
turn on the gene reference system (RefSeq, Ensembl, etc.). On top of
that, the same transcript may have >1 location in the reference
genome: NM_000451 (EntrezID 6473) is annotated on chromosomes X and Y,
NM_001001418 (EntrezID 414060) has two loci on chr17.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
