% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmod.R
\docType{methods}
\name{getGff}
\alias{getGff}
\alias{getGff,RNAmod,numeric,character-method}
\title{getGff}
\usage{
getGff(.Object, number, modification)

\S4method{getGff}{RNAmod,numeric,character}(.Object, number, modification)
}
\arguments{
\item{.Object}{RNAmod.}

\item{number}{numeric.}

\item{modification}{character.}
}
\description{
getGff
}
