% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmodR.R
\docType{methods}
\name{getGffResult}
\alias{getGffResult}
\alias{getGffResult,RNAmodR,numeric-method}
\title{getGffResult}
\usage{
getGffResult(.Object, number, genomicCoordinates = FALSE)

\S4method{getGffResult}{RNAmodR,numeric}(.Object, number,
  genomicCoordinates = FALSE)
}
\arguments{
\item{.Object}{RNAmod.}

\item{number}{numeric.}

\item{genomicCoordinates}{logical value, whether to return results with local
transcript coordinates or genomic coordinates. (default = FALSE)}

\item{modification}{character.}
}
\value{
the found modifications as a GRanges object in the gff3 format.
}
\description{
With this function the results of the parseForModification are
returned, if they are saved on disk. Any combination of modifications can be
requested, but only found results for the given modification are returned.
The results contain the local coordinates per transscript.
}
\examples{
\donttest{
getGffResult(mod,1,"m7G")
}
}
