% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{GetId,txLoc-method}
\alias{GetId,txLoc-method}
\title{Method "GetId" for S4 object txLoc.}
\usage{
\S4method{GetId}{txLoc}(x)
}
\arguments{
\item{x}{A \code{txLoc} object.}
}
\value{
A string.
}
\description{
Get the id string of a \code{txLoc} object.
}
\keyword{internal}

