% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{GetId}
\alias{GetId}
\title{Generic method "GetId" for S4 object txLoc.}
\usage{
GetId(x)
}
\arguments{
\item{x}{A \code{txLoc} object.}
}
\description{
Generic method "GetId" for S4 object txLoc.
}
\keyword{internal}

