% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{GetLoci,txLoc-method}
\alias{GetLoci,txLoc-method}
\title{Method "GetLoci" for S4 object txLoc.}
\usage{
\S4method{GetLoci}{txLoc}(x)
}
\value{
A list of dataframes with loci for
every transcript section.
}
\description{
Get the loci of a \code{txLoc} object
as a list of dataframes.
}
\keyword{internal}

