% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{GetLoci}
\alias{GetLoci}
\alias{GetLoci,txLoc-method}
\title{Method "GetLoci" for S4 object txLoc.}
\usage{
GetLoci(x)

\S4method{GetLoci}{txLoc}(x)
}
\arguments{
\item{x}{A \code{txLoc} object.}
}
\value{
A list of dataframes with loci for
every transcript section.
}
\description{
Get the loci of a \code{txLoc} object
as a list of dataframes.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}

