% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetLocus.MapFromTranscripts}
\alias{GetLocus.MapFromTranscripts}
\title{Construct \code{txLoc}-suitable \code{dataframe} object from
mapping transcript loci to genomic loci.}
\usage{
GetLocus.MapFromTranscripts(gr, ref, seq, section, geneXID)
}
\arguments{
\item{gr}{A \code{GRanges} object; specifies transcript loci.}

\item{ref}{A \code{GRanges} object; the reference transcripts.}

\item{seq}{A \code{DNAStringSet}; the reference transcript
sequences.}

\item{section}{A character scalar; specifies the transcript
section.}

\item{geneXID}{A \code{data.frame}; output of function
\code{GetGeneIds} providing gene IDs from different gene
annotations.}
}
\value{
A \code{dataframe} object.
}
\description{
Construct \code{txLoc}-suitable \code{dataframe} object from
mapping transcript loci to genomic loci.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}

