% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetMFE}
\alias{GetMFE}
\title{Fold sequences.}
\usage{
GetMFE(data, colSeq, colId = NULL)
}
\arguments{
\item{data}{A \code{dataframe} object. See 'Details'.}

\item{colSeq}{An integer scalar; specifies the column in
\code{data} containing the sequences.}

\item{colId}{An integer scalar; specifies the column in
\code{data} containing sequence IDs; if \code{NULL} IDs
are generated automatically; default is \code{NULL}.}
}
\value{
A \code{dataframe} object. See 'Details'.
}
\description{
Fold sequences. See 'Details'.
}
\details{
The function takes a \code{dataframe}, extracts sequences from a
column specified by \code{colSeq}, and predicts secondary structures
using RNAfold \url{http://rna.tbi.univie.ac.at/}.
An optional column containing sequence IDs may be specified by
\code{colId}.
The function returns a \code{dataframe} with three columns:
\enumerate{
\item Column 1: Sequence ID
\item Column 2: Length of the sequence (in nt)
\item Column 3: Mean free energy (MFE)
}
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
