% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmodR-plot.R
\docType{methods}
\name{getModPlot}
\alias{getModPlot}
\alias{getModPlotFromSummarizedExperiment}
\alias{saveModPlot}
\alias{saveModPlotFromSummarizedExperiment}
\alias{getModPlot,RNAmodR,numeric,character,character-method}
\alias{getModPlot}
\alias{saveModPlots}
\alias{saveModPlotFromSummarizedExperiment}
\alias{getModPlotFromSummarizedExperiment,SummarizedExperiment,GRanges,FaFile,character,character-method}
\alias{saveModPlot,RNAmodR,numeric,character,character-method}
\alias{saveModPlot,RNAmodR,numeric,character,missing-method}
\alias{saveModPlotFromSummarizedExperiment,SummarizedExperiment,GRanges,FaFile,character,character,character-method}
\title{Visualization of modifications}
\usage{
getModPlot(.Object, number, modifications, gene, focus = FALSE)

getModPlotFromSummarizedExperiment(se, gff, fasta, modifications, gene,
  focus = FALSE)

saveModPlot(.Object, number, modifications, genes, focus = FALSE,
  filetype = "pdf")

saveModPlotFromSummarizedExperiment(se, gff, fasta, modifications, genes,
  folder, focus = FALSE, filetype = "pdf")

\S4method{getModPlot}{RNAmodR,numeric,character,character}(.Object, number,
  modifications, gene, focus = FALSE)


  \S4method{getModPlotFromSummarizedExperiment}{SummarizedExperiment,GRanges,FaFile,character,character}(se,
  gff, fasta, modifications, gene, focus = FALSE)

\S4method{saveModPlot}{RNAmodR,numeric,character,character}(.Object, number,
  modifications, genes, focus = FALSE, filetype = "pdf")

\S4method{saveModPlot}{RNAmodR,numeric,character,missing}(.Object, number,
  modifications, focus, filetype)


  \S4method{saveModPlotFromSummarizedExperiment}{SummarizedExperiment,GRanges,FaFile,character,character,character}(se,
  gff, fasta, modifications, genes, folder, focus = FALSE, filetype = "pdf")
}
\arguments{
\item{.Object}{a RNAmod object.}

\item{number}{an experiemnt numnber.}

\item{modifications}{name of modification to be used for analysis.}

\item{gene}{a single gene name used by \code{getModPlot}}

\item{focus}{optional: logical, whether to plot only the region next to the 
modification}

\item{se}{a SummarizedExperiment containg the experimental data.}

\item{gff}{a GRanges object with the genomic annotation data}

\item{fasta}{a FaFile object with the genomic sequences}

\item{genes}{gene names used by \code{saveModPlots}}

\item{filetype}{file type used for \code{saveModPlots}}

\item{folder}{folder for saving the output used by \code{saveModPlots}}
}
\value{
\code{getModPlot}: a \code{TableGrob} of layer plot and any mod plot

\code{saveModPlot}: TRUE if successful
}
\description{
\code{getModPlot} and \code{saveModPlot} plot the results by combining
position data and modifications with transcript layout data.
}
\examples{
\donttest{
getModPlot(mod,1,"m7G","RDN18-1")
saveModPlot(mod,1,"m7G",c("RDN18-1","tS\\\\(CGA\\\\)P"))
}
}
