% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{GetNumberOfLoci}
\alias{GetNumberOfLoci}
\title{Method \code{GetNumberOfLoci} for S4 object \code{txLoc}.}
\usage{
GetNumberOfLoci(object)
}
\arguments{
\item{object}{A \code{txLoc} object.}
}
\value{
A named integer vector with the number of sites per
transcript section.
}
\description{
Get the number of loci of a \code{txLoc} object in every transcript section.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
