% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RNAmod-type.R
\docType{methods}
\name{getPlotType,mod-method}
\alias{getPlotType,mod-method}
\title{getPlotType}
\usage{
\S4method{getPlotType}{mod}(object)
}
\value{
character defining the plot type for this modification
}
\description{
getPlotType
}
\examples{
\donttest{
getPlotType(mod)
}
}
