% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{GetRef,txLoc-method}
\alias{GetRef,txLoc-method}
\title{Method "GetRef" for S4 object txLoc.}
\usage{
\S4method{GetRef}{txLoc}(x)
}
\value{
A string.
}
\description{
Get the reference genome string of a \code{txLoc} object.
}
\keyword{internal}

