% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{GetRef}
\alias{GetRef}
\title{Generic method "GetRef" for S4 object txLoc.}
\usage{
GetRef(x)
}
\arguments{
\item{x}{A \code{txLoc} object.}
}
\description{
Generic method "GetRef" for S4 object txLoc.
}
\keyword{internal}

