% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{GetRegions}
\alias{GetRegions}
\title{Method \code{GetRegions} for S4 object \code{txLoc}.}
\usage{
GetRegions(object)
}
\arguments{
\item{object}{A \code{txLoc} object.}
}
\value{
A \code{character} vector.
}
\description{
Character vector of the transcript regions of a \code{txLoc} object.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
