% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{GetRelativeDistance}
\alias{GetRelativeDistance}
\title{Calculate mutual distances between entries for every transcript
region from two txLoc objects.}
\usage{
GetRelativeDistance(loc1, loc2, filter = NULL, method = c("ss", "mm", "se",
  "es", "ee"))
}
\arguments{
\item{loc1}{A txLoc object.}

\item{loc2}{A txLoc object.}

\item{filter}{Only consider loci in transcript regions specified in filter. Default is NULL.}

\item{method}{Method to calculate relative distances. Possible arguments
are "ss" (start-start), "mm" (midpoint-midpoint), "se" (start-end), "es"
(end-start), "ee" (end-end).}
}
\value{
List of upper and lower 95% confidence interval
bounds for every bin value.
}
\description{
Calculate mutual distances between entries for every transcript
region from two txLoc objects.
}

