% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetRelDist}
\alias{GetRelDist}
\title{Calculate mutual distances between entries for every transcript
region from two txLoc objects.}
\usage{
GetRelDist(loc1, loc2, filter = NULL, method = c("nearest", "all"),
  refPoint = c("ss", "mm", "se", "es", "ee"))
}
\arguments{
\item{loc1}{A \code{txLoc} object.}

\item{loc2}{A \code{txLoc} object.}

\item{filter}{A character vector; only consider transcript sections
specified in \code{filter}; if \code{NULL} consider all sections.}

\item{method}{A character strings; specifies the distance
calculation method; possible arguments are "all", "nearest";
default is "nearest".}

\item{refPoint}{A character string; specifies the reference point
for calculating relative distances; possible arguments are "ss"
(start-start), "mm" (midpoint-midpoint), "se" (start-end), "es"
(end-start), "ee" (end-end); default is "ss".}
}
\value{
List of upper and lower 95% confidence interval
bounds for every bin value.
}
\description{
Calculate mutual distances between entries for every transcript
region from two txLoc objects.
}
\details{
The function calculates relative distances between loci from
\code{loc1} and \code{loc2} within the same transcript section.
Positive distances refer to loc1 > loc2, negative distances to
loc1 < loc2. By default only the smallest distance for every
locus from \code{loc1} is kept (\code{method = "nearest"}).
By default relative distances correspond to distances between
start coordinates (\code{refPoint = "ss"}).
}

