% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{GetRelDistNearest}
\alias{GetRelDistNearest}
\title{Return list of nearest distances between entries from two
\code{GRangesList} objects.}
\usage{
GetRelDistNearest(gr1, gr2)
}
\arguments{
\item{gr1}{A \code{GRangesList} object.}

\item{gr2}{A \code{GRangesList} object.}
}
\value{
A list of integer vectors. See 'Details'.
}
\description{
Return list of nearest distances between entries from two
\code{GRangesList} objects. See 'Details'.
}
\details{
The function uses \code{GenomicRanges::distanceToNearest}
to calculate nearest distances between entries from two
\code{GRangesList} objects. The return object is a list
of distances, where every list element corresponds to
a \code{GRangesList} element.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}

