% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetRelDistSpliceSite}
\alias{GetRelDistSpliceSite}
\title{Calculate relative distances between given loci and splice sites.}
\usage{
GetRelDistSpliceSite(locus, ss)
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{ss}{A \code{GRangesList} object.}
}
\value{
An integer vector.
}
\description{
Calculate relative distances between given loci and splice sites.
}

