% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetRelDistSS}
\alias{GetRelDistSS}
\title{Calculate relative distances between loci from \code{txLoc}
object to splice sites.}
\usage{
GetRelDistSS(locus, ss, flank = 1000)
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{ss}{A list of two \code{GRangesList} objects.}

\item{flank}{An integer scalar; specifies the absolute maximum
relative distance used as a cutoff; default is 1000.}
}
\value{
An integer vector.
}
\description{
Calculate relative distances between loci from \code{txLoc}
and splice sites. See 'Details'.
This is a low-level function that is being called from
\code{PlotRelDistSSDistribution} and
\code{PlotRelDistSSEnrichment}.
}
\details{
The function calculates relative distances between 5' splice
sites from \code{ss} and loci from \code{locus}. Distances
are taken to be negative (positive) if a locus from
\code{txLoc} is upstream (downstream) of a splice site.
Distances flank <= d <= flank are considered. The fun
}
\keyword{internal}

