% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetSpliceSites}
\alias{GetSpliceSites}
\title{Get 5'/3' splice sites from transcriptome.}
\usage{
GetSpliceSites(refGenome = "hg38", writeBED = FALSE)
}
\arguments{
\item{refGenome}{A character string; specifies a specific
reference genome assembly version based on which the matching
transcriptome is loaded; default is \code{"hg38"}.}

\item{writeBED}{A logical scalar; if \code{TRUE} write 5'/3'
splice sites to two BED files; default is \code{FALSE}.}
}
\value{
A list of two \code{GRanges} objects. See 'Details'.
}
\description{
Get 5'/3' splice sites from transcriptome. See 'Details'.
}
\details{
This function extracts splice site positions from a reference
transcriptome, and returns a list of two \code{GRanges} objects
of all 5' and 3' splice sites.
}

