% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetSplicingSites}
\alias{GetSplicingSites}
\title{Get splicing sites from reference transcriptome}
\usage{
GetSplicingSites(refGenome = "hg38", filter = c("CDS", "5'UTR"))
}
\arguments{
\item{refGenome}{A character string; specifies a specific
reference genome assembly version based on which the matching
transcriptome is loaded; default is \code{"hg38"}.}

\item{filter}{A character vector; only consider transcript
sections specified in \code{filter}; default is
\code{c("CDS", "5'UTR")}.}
}
\value{
A \code{GRanges} object. See 'Details'.
}
\description{
Get splicing sites from transcriptome. See 'Details'.
}
\details{
The function extracts splicing sites from a reference
trancriptome specified by \code{refGenome}, and returns a
\code{GRanges} object. Splicing sites are identified as
either donor (splicing site at the 5' end of the intron)
or acceptor (splicing site at the 3' end of the intron)
site.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
