% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTxMethods.R
\name{GetTxBySec}
\alias{GetTxBySec}
\title{Get transcript sections.}
\usage{
GetTxBySec(txdb, sections = c("Promoter", "5'UTR", "CDS", "3'UTR", "Intron"),
  promUpstream = 1000, promDownstream = 0, verbose = FALSE)
}
\arguments{
\item{txdb}{A \code{TxDb} object.}

\item{sections}{A character vector; specifies which transcript
sections should be extracted from \code{txdb}; default is
\code{c("Promoter", "5'UTR", "CDS", "3'UTR", "Intron")}.}

\item{promUpstream}{An integer; specifies the number of
nucleotides to be included upstream of 5'UTR start as part of
the promoter.}

\item{promDownstream}{An integer; specifies the number of
nucleotides to be included downstream of 5'UTR start as part
of the promoter.}

\item{verbose}{A logical scalar; if \code{TRUE}, print additional
information; default is \code{FALSE}}
}
\value{
A \code{list} of \code{GRangesList} objects
}
\description{
Get transcript sections. See 'Details'.
This is a low-level function that is being called from
\code{BuildTx}.
}
\details{
The function returns a \code{list} of \code{GRangesList}
objects. The number and names of list entries match the
number and names of \code{sections}. Entries in every
\code{GRangesList} are labelled according to their
transcript RefSeq ID. Note that multiple entries with
identical RefSeq ID exist.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}

