% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTxMethods.R
\name{GetTxDb}
\alias{GetTxDb}
\title{Download a new or load an existing reference transcriptome.}
\usage{
GetTxDb(genomeVersion = "hg38", standardChrOnly = TRUE, force = FALSE,
  verbose = FALSE)
}
\arguments{
\item{genomeVersion}{A character string; refers to a specific
reference genome assembly version; default is \code{"hg38"}.}

\item{standardChrOnly}{A logical scalar; if \code{TRUE} and a
new db is created, keep only fully assembled chromosomes;
default is \code{TRUE}.}

\item{force}{A logical scalar; if \code{TRUE} force download
and overwrite existing db file; default is \code{FALSE}.}

\item{verbose}{A logical scalar; print additional information
about db; default is \code{FALSE}.}
}
\value{
A \code{TxDb} object. See 'Details'.
}
\description{
Download a new or load an existing reference transcriptome.
See 'Details'.
This is a low-level function that is being called from
\code{BuildTx}.
}
\details{
If no local transcriptome db file exists or \code{forceDownload}
is \code{TRUE}, the function generates an reference transcriptome
by calling \code{makeTxDbFromUCSC} from the \code{GenomicFeatures}
package; the return object is a \code{TxDb} object and is saved
locally as an sqlite db.
Details on the \code{txdb} return object can be obtained from
\code{print(x)}, where \code{x} is the name of the \code{txdb}
return object.
The user may generate a custom transcriptome (e.g. based on a
different gene annotation) using \code{makeTxDb} or
\code{makeTxDb}-derived functions from the \code{GenomicFeatures}
package. Please see the \code{GenomicFeatures} reference manual.
In order to use a custom transcriptome, the \code{TxDb} object
needs to be saved locally as an sqlite db, e.g.
\code{saveDb(x), "txdb_y.sqlite")}, where \code{x} is the
\code{TxDb} object and \code{y} denotes the reference genome
version as in \code{genomeVersion}.
}
\examples{
\dontrun{
txdb <- GetTxDb();
print(txdb);
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}

