% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{GetVersion,txLoc-method}
\alias{GetVersion,txLoc-method}
\title{Method "GetVersion" for S4 object txLoc.}
\usage{
\S4method{GetVersion}{txLoc}(x)
}
\value{
A string.
}
\description{
Get the version string of a \code{txLoc} object.
}
\keyword{internal}

