% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{GetVersion}
\alias{GetVersion}
\title{Generic method "GetVersion" for S4 object txLoc.}
\usage{
GetVersion(x)
}
\arguments{
\item{x}{A \code{txLoc} object.}
}
\description{
Generic method "GetVersion" for S4 object txLoc.
}
\keyword{internal}

