% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR-data.R
\docType{data}
\name{gff}
\alias{gff}
\title{GRanges object for example data}
\format{An object of class \code{GRanges} of length 32082.}
\usage{
gff
}
\description{
example data - GRanges object
}
\keyword{RNAmod}
\keyword{datasets}
\keyword{gff}
