% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmodR-ident-D.R, R/class-RNAmodR-ident-m3C.R,
%   R/class-RNAmodR-ident-m7G.R, R/class-RNAmodR-ident-ribometh.R
\docType{methods}
\name{identifyModificationPerTranscript}
\alias{identifyModificationPerTranscript}
\alias{identifyModificationsPerTranscript,RNAmodRident_D,numeric,numeric,list-method}
\alias{identifyModificationsPerTranscript,RNAmodRident_m3C,numeric,numeric,list-method}
\alias{identifyModificationsPerTranscript,RNAmodRident_m7G,numeric,numeric,list-method}
\alias{identifyModificationsPerTranscript,RNAmodRident_RiboMeth,numeric,numeric,list-method}
\title{Returns the data for each transcript}
\usage{
identifyModificationsPerTranscript(x, location, locations, data)


  \S4method{identifyModificationsPerTranscript}{RNAmodRident_D,numeric,numeric,list}(x,
  location, locations, data)


  \S4method{identifyModificationsPerTranscript}{RNAmodRident_m3C,numeric,numeric,list}(x,
  location, locations, data)


  \S4method{identifyModificationsPerTranscript}{RNAmodRident_m7G,numeric,numeric,list}(x,
  location, locations, data)


  \S4method{identifyModificationsPerTranscript}{RNAmodRident_RiboMeth,numeric,numeric,list}(x,
  location, locations, data)
}
\description{
\code{RNAmodRident_D}

\code{RNAmodRident_m3C}

\code{RNAmodRident_m7G}

\code{RNAmodRident_RiboMeth}
}
