% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{Info}
\alias{Info}
\alias{Info,txLoc-method}
\title{Method "Info" for S4 object txLoc.}
\usage{
Info(x)

\S4method{Info}{txLoc}(x)
}
\arguments{
\item{x}{A \code{txLoc} object.}
}
\description{
Print general information of a \code{txLoc} object.
}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR");
sites <- ReadBED(bedFile);
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38");
Info(posSites);
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

