% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_SummarizedExperiment}
\alias{is_SummarizedExperiment}
\alias{assert_is_SummarizedExperiment}
\alias{assert_all_are_SummarizedExperiment}
\title{assertive: SummarizedExperiment}
\usage{
is_SummarizedExperiment(x, .xname = get_name_in_parents(x))

assert_is_SummarizedExperiment(x, severity = getOption("assertive.severity",
  "stop"))

assert_all_are_SummarizedExperiment(x,
  severity = getOption("assertive.severity", "stop"))
}
\arguments{
\item{x}{object to test}

\item{.xname}{name of the object in parent env}

\item{severity}{assertions only: severity of failing assertion}
}
\value{
logical

invisible logical

invisible logical
}
\description{
\code{is_SummarizedExperiment}/\code{assert_is_SummarizedExperiment}: Check 
if input is a SummarizedExperiment object
\code{assert_all_are_SummarizedExperiment}: Check if input is a list of 
SummarizedExperiment objects
}
