% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_TxDb}
\alias{is_TxDb}
\title{assertive: TxDb}
\usage{
is_TxDb(x, .xname = assertive.base::get_name_in_parent(x))
}
\arguments{
\item{x}{object to test}

\item{.xname}{name of the object in parent env}

\item{severity}{assertions only: severity of failing assertion}
}
\value{
logical
}
\description{
\code{is_TxDb} Check 
if input is a TxDb object
}
