% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{IsEmptyChar}
\alias{IsEmptyChar}
\title{Check if all entries in a character vector are empty.}
\usage{
IsEmptyChar(v)
}
\arguments{
\item{v}{A character vector.}
}
\value{
A logical scalar.
}
\description{
Check if all entries in a character vector are empty.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}

