% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{LoadRefTx}
\alias{LoadRefTx}
\title{Load reference transcriptome.}
\usage{
LoadRefTx(refGenome = "hg38", geneXID, seqBySec, txBySec)
}
\arguments{
\item{refGenome}{A character string; specifies a specific
reference genome assembly version based on which a transcriptome
is loaded; default is \code{"hg38"}.}

\item{geneXID}{A \code{data.frame}; this is the return object
generated via R's pseudo call-by-reference.}

\item{seqBySec}{A \code{list} of \code{DNAStringSet objects};
this is the return object generated via R's pseudo
call-by-reference.}

\item{txBySec}{A \code{list} of \code{GRangesList} objects;
this is the return object generated via R's pseudo
call-by-reference.}
}
\description{
Load reference transcriptome.
}
\keyword{internal}

