% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/class-RNAmodR-mod-D.R,
%   R/class-RNAmodR-mod-m3C.R, R/class-RNAmodR-mod-m7G.R
\docType{methods}
\name{maskPositionData}
\alias{maskPositionData}
\alias{maskPositionData,mod_D,numeric,numeric-method}
\alias{maskPositionData,mod_m3C,numeric,numeric-method}
\alias{maskPositionData,mod_m7G,numeric,numeric-method}
\title{maskPositionData}
\usage{
maskPositionData(object, data, modLocations)

\S4method{maskPositionData}{mod_D,numeric,numeric}(object, data, modLocations)

\S4method{maskPositionData}{mod_m3C,numeric,numeric}(object, data, modLocations)

\S4method{maskPositionData}{mod_m7G,numeric,numeric}(object, data, modLocations)
}
\description{
\code{mod_D}

\code{mod_m3C}

\code{mod_m7G}
}
