% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/RNAmod-utils.R
\docType{methods}
\name{matchFastaToGff}
\alias{matchFastaToGff}
\alias{matchFastaToGff,character,character-method}
\title{matchFastaToGff}
\usage{
matchFastaToGff(inputFasta, inputGFF)

\S4method{matchFastaToGff}{character,character}(inputFasta, inputGFF)
}
\arguments{
\item{inputFasta}{fasta file to be used for renaming}

\item{inputGFF}{gff file containing gene, mRNA and CDS annotation data}
}
\value{
TRUE
}
\description{
title
}
