% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{MatchFlank}
\alias{MatchFlank}
\title{Match `flank` function argument}
\usage{
MatchFlank(flank)
}
\arguments{
\item{flank}{An \code{integer} scalar or an \code{nteger} vector of length 2.}
}
\value{
An \code{integer} vector of length 2.
}
\description{
Match `flank` function argument
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
