% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/class-RNAmodR-data-type.R
\docType{methods}
\name{mergeDataOfReplicates}
\alias{mergeDataOfReplicates}
\alias{mergeDataOfReplicates,data-method}
\title{Merge data of replicates}
\usage{
mergeDataOfReplicates(x)

\S4method{mergeDataOfReplicates}{data}(x)
}
\arguments{
\item{x}{a mod object}
}
\value{
a modified \code{data} class object now containing the aggregated
pre-processed data of the input. This data is used for saving.
}
\description{
merges data from replicates for each data class.
}
