% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmod-m7G.R
\docType{methods}
\name{mergeModsOfReplicates}
\alias{mergeModsOfReplicates}
\alias{mergeModsOfReplicates,mod_m7G,GRanges,FaFile,list-method}
\title{mergeModsOfReplicates}
\usage{
mergeModsOfReplicates(object, gff, seq, data)

\S4method{mergeModsOfReplicates}{mod_m7G,GRanges,FaFile,list}(object, gff, seq,
  data)
}
\arguments{
\item{object}{mod object}

\item{gff}{a GRanges object}

\item{seq}{a DNAStringSet object}

\item{res}{the results for individual replicates}
}
\description{
\code{mod_m7G}
}
