% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmod-analysis-default.R
\docType{methods}
\name{mergePositionsOfReplicates}
\alias{mergePositionsOfReplicates}
\alias{mergePositionsOfReplicates,analysis_default,GRanges,FaFile,list-method}
\title{mergePositionsOfReplicates}
\usage{
mergePositionsOfReplicates(object, gff, fafile, data)


  \S4method{mergePositionsOfReplicates}{analysis_default,GRanges,FaFile,list}(object,
  gff, fafile, data)
}
\arguments{
\item{object}{mod object}

\item{gff}{a GRanges object for the genome}

\item{fafile}{a FaFile object for the genome}

\item{data}{list of DataFrame containg the read data}
}
\description{
\code{analysis_default}
}
