% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/class-RNAmod-type.R
\docType{methods}
\name{mod-accessors}
\alias{mod-accessors}
\alias{getModType}
\alias{getPlotType,mod-method}
\alias{getModType,mod-method}
\title{Accessor for 'mod' class}
\usage{
getPlotType(object)

getModType(object)

\S4method{getPlotType}{mod}(object)

\S4method{getModType}{mod}(object)
}
\arguments{
\item{object}{mod object}
}
\value{
character defining the plot type for this modification class

character defining the modification type for this modification
class
}
\examples{
\donttest{
getPlotType(mod)
getModType(mod)
}
}
