% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RNAmod-type.R, R/class-RNAmod-D.R,
%   R/class-RNAmod-default.R, R/class-RNAmod-m3C.R, R/class-RNAmod-m7G.R
\docType{class}
\name{mod}
\alias{mod}
\alias{mod_D-class}
\alias{mod_default-class}
\alias{mod_m3C-class}
\alias{mod_m7G-class}
\title{mod}
\description{
\code{mod_D}

\code{mod_m7G}

\code{mod_m3C}

\code{mod_m7G}
}
