% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gviz-ModifiedSequenceTrack-class.R
\docType{class}
\name{ModDNASequenceTrack}
\alias{ModDNASequenceTrack}
\alias{ModDNASequenceTrack-class}
\title{ModDNASequenceTrack}
\usage{
ModDNASequenceTrack(sequence, chromosome, genome, ...)
}
\arguments{
\item{sequence}{A \code{character} vector or \code{ModDNAString} object of 
length one. The sequence to display. See 
\code{\link[Gviz:SequenceTrack-class]{SequenceTrack}}.}

\item{chromosome, genome, ...}{See 
\code{\link[Gviz:SequenceTrack-class]{SequenceTrack}}.}
}
\value{
a \code{ModDNASequenceTrack} object
}
\description{
A \code{Gviz} compatible 
\code{\link[Gviz:SequenceTrack-class]{SequenceTrack}} for showing modified 
DNA sequences.
}
\examples{
seq <- ModDNAStringSet(c(chr1 = paste0(alphabet(ModDNAString()),collapse = "")))
st <- ModDNASequenceTrack(seq)
Gviz::plotTracks(st, chromosome = "chr1",from = 1L, to = 20L)
}
