% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModExperiment-class.R
\docType{class}
\name{ModExperiment}
\alias{ModExperiment}
\alias{ModExperiment-class}
\alias{show,ModExperiment-method}
\title{ModExperiment}
\usage{
\S4method{show}{ModExperiment}(object)

\S4method{gff}{ModExperiment}(x)

\S4method{fasta}{ModExperiment}(x)

\S4method{sequences}{ModExperiment}(x)

\S4method{ranges}{ModExperiment}(x)

\S4method{bamfiles}{ModExperiment}(x)

\S4method{modifications}{ModExperiment}(x)
}
\description{
title
}
