% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifierSet-export.R
\docType{methods}
\name{ModifierSet-export}
\alias{ModifierSet-export}
\alias{SummarizedExperiment,ModifierSet-method}
\title{Exporting data to other classes and files}
\usage{
\S4method{SummarizedExperiment}{ModifierSet}(assays, ...)
}
\arguments{
\item{assays}{a \code{\link[=ModifierSet-class]{ModifierSet}} object}

\item{...}{Optional arguments:
\itemize{
\item{modificationsOnly} {\code{TRUE} or \code{FALSE}: Should only data on
positions with a modification found be included in the 
\code{RangedSummarizedExperiment}? (default = 
\code{modificationsOnly = TRUE})}
\item{flanking} {If \code{modificationsOnly = FALSE} how many neighboring
positions should be included in the \code{RangedSummarizedExperiment}?
(default = \code{flanking = 10L})}
\item{allPositions} {\code{TRUE} or \code{FALSE}: Should data on
*all* positions of the transcripts be included in the 
\code{RangedSummarizedExperiment}? This overrides the previous two options.
It is potentially very resource hungry. Be careful. (default = 
\code{allPositions = FALSE})}
}}
}
\description{
title
}
