% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-Inosine-class.R,
%   R/Modifier-Inosine-viz.R
\docType{methods}
\name{ModInosine-internals}
\alias{ModInosine-internals}
\alias{.dataTracks,ModInosine,GRanges,GRanges,XString-method}
\title{ModInosine internal functions}
\usage{
\S4method{.dataTracks}{ModInosine,GRanges,GRanges,XString}(x, data,
  seqdata, sequence, args)
}
\arguments{
\item{x, data, seqdata, sequence, args}{internally used arguments}
}
\description{
These functions are not intended for general use, but are used for 
additional package development.
}
