% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-RiboMethSeq-class.R,
%   R/Modifier-RiboMethSeq-viz.R
\docType{class}
\name{ModRiboMethSeq}
\alias{ModRiboMethSeq}
\alias{RiboMethSeq}
\alias{ModRiboMethSeq-class}
\alias{ModRiboMethSeq,character-method}
\alias{ModRiboMethSeq,BamFileList-method}
\alias{ModRiboMethSeq,SequenceData-method}
\alias{modify,ModRiboMethSeq-method}
\alias{visualizeData,ModRiboMethSeq-method}
\title{ModRiboMethSeq}
\usage{
\S4method{ModRiboMethSeq}{character}(x, fasta, gff, modifications = NULL,
  ...)

\S4method{ModRiboMethSeq}{BamFileList}(x, fasta, gff,
  modifications = NULL, ...)

\S4method{ModRiboMethSeq}{SequenceData}(x, modifications = NULL, ...)

\S4method{modify}{ModRiboMethSeq}(x, ...)

\S4method{visualizeData}{ModRiboMethSeq}(x, i, type = c("ends", "scoreA",
  "scoreB", "scoreRMS"), start, end)
}
\description{
title

score MAX as described by publiccations from the motorin lab are not 
implemented since an unambigeous description is not available from the 
literature.
}
