% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gviz-ModifiedSequenceTrack-class.R
\docType{class}
\name{ModRNASequenceTrack}
\alias{ModRNASequenceTrack}
\alias{ModRNASequenceTrack-class}
\title{ModRNASequenceTrack}
\usage{
ModRNASequenceTrack(sequence, chromosome, genome, ...)
}
\arguments{
\item{sequence}{A \code{character} vector or \code{ModRNAString} object of 
length one. The sequence to display. See 
\code{\link[Gviz:SequenceTrack-class]{SequenceTrack}}.}

\item{chromosome, genome, ...}{See 
\code{\link[Gviz:SequenceTrack-class]{SequenceTrack}}.}
}
\description{
A \code{Gviz} compatible 
\code{\link[Gviz:SequenceTrack-class]{SequenceTrack}} for showing modified 
RNA sequences.
}
\examples{


}
