% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmod-parse.R
\docType{methods}
\name{parseForModifications}
\alias{parseForModifications}
\alias{parseForModifications}
\title{parseForModifications}
\usage{
parseForModifications(.Object, number, modifications)

\S4method{parseForModifications}{RNAmod,numeric,character}(.Object, number,
  modifications)
}
\arguments{
\item{.Object}{a RNAmod object.}

\item{number}{a number defining an experiment.}

\item{modifications}{name of modification to be used for analysis.}
}
