% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/class-RNAmodR-data-type.R
\docType{methods}
\name{parseMod}
\alias{parseMod}
\alias{parseMod,data,GRanges,FaFile,list-method}
\title{parseMod}
\usage{
parseMod(object, gff, fafile, modClasses)

\S4method{parseMod}{data,GRanges,FaFile,list}(object, gff, fafile, modClasses)
}
\arguments{
\item{object}{mod object}

\item{gff}{a GRanges object for the genome}

\item{fafile}{a FaFile object for the genome}

\item{modClasses}{list of modification classes used for parsing}
}
\value{
a modified \code{data} class object containting pre-processed
data and found modifications
}
\description{
\code{data_5end}: parses the pre-processed data in the data slot
to detect modifications
}
