% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTxMethods.R
\name{PerformSanityCheck}
\alias{PerformSanityCheck}
\title{Perform sanity check of collapsed transcriptome.}
\usage{
PerformSanityCheck(txBySec)
}
\arguments{
\item{txBySec}{A \code{list} of \code{GRangesList} objects;
output of function \code{GetTxBySec} or \code{CollapseTxBySec}.}
}
\description{
Perform sanity check of collapsed transcriptome. See 'Details'.
This is a low-level function that is being called from
\code{BuildTx}.
}
\details{
The function checks for entries with duplicate RefSeq IDs
across a list of transcripts sections. For example, the return
object of \code{CollapseTxBySec} should contain transcript
segments for every transcript section with a unique RefSeq ID.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
