% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-pileup.R
\docType{class}
\name{PileupSequenceData}
\alias{PileupSequenceData}
\alias{PileupSequenceData-class}
\title{PileupSequenceData}
\usage{
PileupSequenceData(bamfiles, fasta, gff, ...)
}
\description{
title
}
