% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotAbundance.generic}
\alias{PlotAbundance.generic}
\title{Generic function for plotting abundances (histograms).}
\usage{
PlotAbundance.generic(data, xmin = NULL, xmax = NULL,
  binWidth = NULL, plotType = "s", lwd = 2, title = "",
  xlab = "", ylab = "Abundance", doBootstrap = TRUE, nBS = 5000,
  doLowess = TRUE)
}
\arguments{
\item{data}{A vector of integers.}

\item{xmin}{An integer scalar; default is \code{NULL}.}

\item{xmax}{An integer scalar; default is \code{NULL}.}

\item{binWidth}{An integer scalar; default is \code{NULL}.}

\item{plotType}{A single character; default is \code{"s"}.}

\item{lwd}{An integer scalar; default is 2.}

\item{title}{A character string; default is \code{""}.}

\item{xlab}{A character string; default is \code{""}.}

\item{ylab}{A character string; default is \code{""}.}

\item{doBootstrap}{A logical scalar; default is \code{TRUE}.}

\item{nBS}{An integer scalar; default is 5000.}
}
\description{
Generic function for plotting abundances (histograms).
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
