% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotEnrichment.Generic}
\alias{PlotEnrichment.Generic}
\title{Generic function to perform enrichment analysis and plot results.}
\usage{
PlotEnrichment.Generic(mat, title = "", x.las = 1, x.cex = 1,
  x.padj = 1, plotType = "l", reverseXaxis = FALSE,
  withExtendedAxisLabel = 0)
}
\arguments{
\item{mat}{The data matrix.}

\item{title}{An identifier for \code{mat}.}

\item{x.las}{Orientation of x-axis labels. Default is 1.}

\item{x.cex}{Scaling factor for x-axis labels. Default is 1.}

\item{x.padj}{Vertical adjustment of x-axis labels. Default is 1.}

\item{plotType}{Plot style. Default is "l".}

\item{reverseXaxis}{Reverse the order of values from \code{mat}.}

\item{withExtendedAxisLabel}{Print extended axis label.}
}
\value{
A list of \code{fisher.test} return objects and \code{mat}
}
\description{
Generic function to perform enrichment analysis and plot results.
Enrichment/depletion is evaluated using (multiple) Fisher's exact test(s).
Multiple hypothesis testing correction is applied following the method of
Bejamini and Hochberg.
Note: This function should not be invoked directly by the user.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}

