% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotGC}
\alias{PlotGC}
\title{Plot GC content.}
\usage{
PlotGC(loc1, loc2, flank = 10, filter = NULL, geneNorm = FALSE,
  subsample = TRUE)
}
\arguments{
\item{loc1}{A \code{txLoc} object.}

\item{loc2}{A \code{txLoc} object.}

\item{flank}{An integer scalar; see 'Details'.}

\item{filter}{A logical scalar; only consider loci in transcript
regions specified in filter; Default is \code{NULL}.}

\item{geneNorm}{A logical scalar; if \code{TRUE} normalise GC
content in window to GC content of transcript section; default
is \code{FALSE}.}

\item{subsample}{A logical scalar; if \code{TRUE} a subsample
of \code{loc2} is used instead of the full set; the subsample 
size is dynamically determined based on the total number of sites
in \code{loc1}; default is \code{TRUE}.}
}
\description{
Plot and assess GC content distributions from two site lists.
See 'Details'.
}
\details{
The function calculates the GC content within a region around every
site from two \code{txLoc} objects. The window is defined
by extending the position of every transcript locus upstream and
downstream by \code{flank} nucleotides (if possible).
The means of the resulting GC content distributions are assessed
using a two-tailed t-test. If \code{geneNorm = TRUE}, the site GC
content is normalised to the GC content of the entire transcript
section. If \code{subsample = TRUE}, only a subsample of entries
from the \emph{second} \code{txLoc} object will be used. This is
useful (and therefore the default), as \code{loc2} usually
refers to the much larger list of null sites.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

