% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotRelDistEEJDistribution}
\alias{PlotRelDistEEJDistribution}
\title{Plot distribution of relative distances to exon-exon junctions.}
\usage{
PlotRelDistEEJDistribution(dist, flank = 1000, binWidth = 20,
  doBootstrap = TRUE)
}
\arguments{
\item{dist}{A list of integer vectors.}

\item{flank}{An integer scalar; specifies the absolute maximum
relative distance used as a cutoff; default is 1000.}

\item{binWidth}{An integer scalar; specifies the spatial width
by which distances will be binned; default is 20.}

\item{doBootstrap}{A logical scalar; if \code{YES} calculate
95% CI based on empirical bootstrap of sites within transcript
region; default is \code{TRUE}.}
}
\description{
Plot distribution of relative distances to exon-exon junctions.
}

