% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotRelDistEnrichment}
\alias{PlotRelDistEnrichment}
\title{Perform and plot enrichment analysis of relative distances of
"positive" and "negative" control sites relative to "reference"
sites.}
\usage{
PlotRelDistEnrichment(locPos, locNeg, locRef, flank = 1000, binWidth = 20)
}
\arguments{
\item{locPos}{A \code{txLoc} object.}

\item{locNeg}{A \code{txLoc} object.}

\item{locRef}{A \code{txLoc} object.}

\item{flank}{An integer scalar; specifies the absolute maximum
relative distance used as a cutoff; default is 1000.}

\item{binWidth}{An integer scalar; specifies the spatial width
by which distances will be binned; default is 20.}
}
\description{
Perform and plot enrichment analysis of relative distances of
"positive" and "negative" control sites relative to "reference"
sites.
}

