% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotRelDistSSDistribution}
\alias{PlotRelDistSSDistribution}
\title{Plot relative distance distribution of loci from \code{txLoc}
object to splice sites.}
\usage{
PlotRelDistSSDistribution(locus, ss, flank = 1000, binWidth = 20)
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{ss}{A list of two \code{GRangesList} objects.}

\item{flank}{An integer scalar; specifies the absolute maximum
relative distance used as a cutoff; default is 1000.}

\item{binWidth}{An integer scalar; specifies the spatial width
by which distances will be binned; default is 20.}
}
\description{
Plot relative distance distribution of loci from \code{txLoc}
object to splice sites.
}

