% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotRelDistSSEnrichment}
\alias{PlotRelDistSSEnrichment}
\title{Plot relative distance enrichment of sites from \code{locPos}
and \code{locPos} to splice sites.}
\usage{
PlotRelDistSSEnrichment(locPos, locNeg, ss, flank = 1000, binWidth = 20)
}
\arguments{
\item{locPos}{A \code{txLoc} object; specifies the positive
sites.}

\item{locNeg}{A \code{txLoc} object; specifies the negative
sites.}

\item{ss}{A list of two \code{GRangesList} objects.}

\item{flank}{An integer scalar; specifies the absolute maximum
relative distance used as a cutoff; default is 1000.}

\item{binWidth}{An integer scalar; specifies the spatial width
by which distances will be binned; default is 20.}
}
\description{
Plot relative distance enrichment of sites from \code{locPos}
and \code{locPos} to splice sites.
}

