% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotRelStartStopEnrichment}
\alias{PlotRelStartStopEnrichment}
\title{Enrichment analysis of sites relative to start/stop codon.}
\usage{
PlotRelStartStopEnrichment(txLoc1, txLoc2, flank = 550, binWidth = 20)
}
\arguments{
\item{txLoc1}{A \code{txLoc} object.}

\item{txLoc2}{A \code{txLoc} object.}

\item{flank}{An integer scalar or an integer vector of length 2; specifies 
the absolute maximum relative distance(s) used as a cutoff; by specifying 
\code{flank} as a vector, a non-symmetric window can be  defined; e.g. 
`flank = c(1000, 2000)` corresponds to a window defined by 1000 nt 
downstream and 2000 nt upstream of the reference site. Default is 1000.}

\item{binWidth}{An \code{integer} scalar; specifies the spatial width in nt
by which distances will be binned; default is 20.}
}
\value{
\code{NULL}.
}
\description{
Perform and visualise the enrichment analysis of sites from a \code{txLoc} 
object relative to the start/stop codons from a reference transcriptome. 
See 'Details'.
}
\details{
The function calculates relative distances of sites from a \code{txLoc}
object to the corresponding transcript's start and stop codons.
Enrichment/depletion is assessed using multiple Fisher's exact tests on the
counts per distance bin relative to the counts in all other bins within the
window defined by (-\code{flank}, \code{flank}). Resulting enrichment plots
show odds-ratios (including 95\% confidence intervals) and associated
p-values as a function of relative distance bins. Negative distances
indicate sites from \code{txLoc} that are \emph{upstream} of the start/stop
codon; positive distances correspond to sites from \code{txLoc} that are 
\emph{downstream} of \code{txLocRef}. The bin width and window size can be 
adjusted with \code{flank} and \code{binWidth}.
Note that this function can be quite slow if \code{txLoc2} is based on all 
null sites. If this is the case, consider downsampling \code{txLoc2} using
\code{DownsampleTxLoc}.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
