% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotSectionDistribution}
\alias{PlotSectionDistribution}
\title{Plot piechart of the number of loci in every transcript section.}
\usage{
PlotSectionDistribution(locus, filter = NULL)
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{filter}{A character vector; only consider transcript sections
specified in \code{filter}; if \code{NULL} consider all sections.}
}
\description{
Plot piechart of the number of loci in every transcript section.
}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR");
sites <- ReadBED(bedFile);
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38");
PlotPieNumberOfLoci(posSites);
}

}

