% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotSectionEnrichment}
\alias{PlotSectionEnrichment}
\title{Perform enrichment analysis of sites per transcript region and plot results.}
\usage{
PlotSectionEnrichment(txLoc.pos, txLoc.neg, xAxisLblFmt = 2)
}
\arguments{
\item{txLoc.pos}{A \code{txLoc} object. These correspond to the positive 
sites.}

\item{txLoc.neg}{A \code{txLoc} object. These correspond to the negative
(null) sites.}

\item{xAxisLblFmt}{Plot extended axis labels. Default is 2. See ??? for details.}
}
\value{
\code{NULL}.
}
\description{
Perform enrichment analysis of the number of positive sites in 
\code{txLoc.pos} relative to the number of null sites in \code{txLoc.neg}
per transript region and plot results.
Enrichment/depletion is evaluated using (multiple) Fisher's exact test(s).
Multiple hypothesis testing correction is applied following the method of
Bejamini and Hochberg.
}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR")
sites <- ReadBED(bedFile)
pos <- SmartMap(sites, id = "m6A", refGenome = "hg38")
neg <- GenerateNull(pos, method = "nuclAbundance")
PlotSectionEnrichment(
    FilterTxLoc(pos, c("5'UTR", "CDS", "3'UTR")),
    FilterTxLoc(neg, c("5'UTR", "CDS", "3'UTR")))
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
