% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotSeqLogo}
\alias{PlotSeqLogo}
\title{Plot sequence logo.}
\usage{
PlotSeqLogo(txLoc, flank = 5, ylim = c(0, 2))
}
\arguments{
\item{txLoc}{A \code{txLoc} object.}

\item{flank}{An \code{integer} scalar; see 'Details'.}

\item{ylim}{An \code{integer} vector; specifies limits for the y-axis;
automatically determined if \code{ymin = NULL}; default is \code{c(0, 2)}.}
}
\description{
Plot sequence logo.
}
\details{
The function determines the sequence logo within a window defined by
extending sites from \code{txLoc} upstream and downstream by \code{flank}
nucleotides.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
