% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotSpatialDistribution}
\alias{PlotSpatialDistribution}
\title{Plot spatial distribution of loci from a \code{txLoc} object.}
\usage{
PlotSpatialDistribution(locus, filter = NULL, nbreaks = 100,
  absolute = FALSE, binWidth = NULL, posMax = 1000,
  doBootstrap = TRUE, ...)
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{filter}{Only plot loci in transcript regions specified in filter.}

\item{nbreaks}{Number of spatial bins. Default is 100.}

\item{absolute}{Plot spatial distribution in absolute coordinates.
Default is \code{FALSE}.}

\item{binWidth}{Spatial bin width. Overrides \code{nbreaks} if not
\code{NULL}.}

\item{posMax}{If \code{absolute == TRUE}, show spatial distribution
within a window given by \code{posMax} from the 5'/3' position of
the transcript feature. Default is 1000 nt.}

\item{doBootstrap}{Calculate 95% CI based on empirical bootstrap of
sites within transcript region. Default is \code{TRUE}.}

\item{...}{Additional parameters passed to plot.}
}
\description{
Plot spatial distribution of loci from a \code{txLoc} object within every
transcript region.
}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR")
sites <- ReadBED(bedFile)
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38")
PlotSpatialDistribution(posSites)
PlotSpatialDistribution(posSites,
                        absolute = TRUE,
                        filter = c("5'UTR", "CDS", "3'UTR"),
                        ylim = c(0, 200))
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
