% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotSpatialEnrichment}
\alias{PlotSpatialEnrichment}
\title{Perform spatial enrichment analysis and plot results.}
\usage{
PlotSpatialEnrichment(txLoc.pos, txLoc.neg, binWidth = 20,
  posMax = 1000)
}
\arguments{
\item{txLoc.pos}{A \code{txLoc} object. These correspond to the positive
sites.}

\item{txLoc.neg}{A \code{txLoc} object. These correspond to the negative
(null) sites.}

\item{binWidth}{Spatial bin width. Default is 20 nt.}

\item{posMax}{Evaluate enrichment within a window given by \code{posMax}.
Default is 1000 nt.}
}
\value{
\code{NULL}.
}
\description{
Perform enrichment analysis of the spatial distribution of positive sites in
\code{txLoc.pos} relative to the distribution of null sites in
\code{txLoc.neg} per transcript region and plot results.
Enrichment/depletion is evaluated using (multiple) Fisher's exact test(s).
Multiple hypothesis testing correction is applied following the method of
Bejamini and Hochberg.
}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR")
sites <- ReadBED(bedFile)
pos <- SmartMap(sites, id = "m6A", refGenome = "hg38")
null <- GenerateNull(posSites, method = "permutation")
PlotSpatialEnrichment(
    FilterTxLoc(pos, c("5'UTR", "CDS", "3'UTR")),
    FilterTxLoc(neg, c("5'UTR", "CDS", "3'UTR")))
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
