% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotTxLengthDistribution}
\alias{PlotTxLengthDistribution}
\title{Plot length distribution of transcript regions.}
\usage{
PlotTxLengthDistribution(txFeatures, asLog = TRUE, printMetrics = "median",
  filter = NULL, outliers = FALSE, ...)
}
\arguments{
\item{txFeatures}{List of \code{GRangesList} transcript features.}

\item{asLog}{Plot length distribution on a log scale. Default is
\code{TRUE}.}

\item{printMetrics}{Print median and mad. If \code{FALSE}, print
mean and sd. Default is \code{TRUE}.}

\item{filter}{Only plot transcript sections specified in \code{filter}.
For example, \code{filter = c("3'UTR", "5'UTR")} plots length distributions
of the 3'/5'UTRs only.}

\item{outliers}{Show outliers in boxplot. Default is \code{FALSE}.}

\item{...}{Additional parameters passed to x-axis.}
}
\description{
Plot length distribution of transcript regions from list of
GRangesList transcript features.
}

