% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotTxSecLength.bean}
\alias{PlotTxSecLength.bean}
\title{Plot length distribution of transcript sections.}
\usage{
PlotTxSecLength.bean(txBySec, asLog = TRUE, printMetrics = c("median",
  "mean"), filter = NULL, ...)
}
\arguments{
\item{txBySec}{A \code{list} of \code{GRangesList} objects;
specifies the list of transcript sections.}

\item{asLog}{A logical scalar; if \code{TRUE} plot length
distribution on a log scale; Default is \code{TRUE}.}

\item{printMetrics}{A character string; specifies which metrics
should be printed as part of the box labels; default is \code{"median"}.}

\item{filter}{A character vector; only plot transcript sections
specified in \code{filter}; if \code{NULL} plot all sections; default
is \code{NULL}.}

\item{...}{Any additional parameters passed to \code{axis}.}
}
\description{
Plot length distribution of transcript sections.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
