% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotTxSecLength}
\alias{PlotTxSecLength}
\title{Plot length distribution of transcript sections.}
\usage{
PlotTxSecLength(txBySec, asLog = TRUE, printMetrics = c("median", "mean"),
  filter = NULL, outliers = FALSE, ...)
}
\arguments{
\item{txBySec}{A \code{list} of \code{GRangesList} objects;
specifies the list of transcript sections.}

\item{asLog}{A logical scalar; if \code{TRUE} plot length
distribution on a log scale; Default is \code{TRUE}.}

\item{printMetrics}{A character string; specifies which metrics
should be printed as part of the box labels; default is
\code{"median"}.}

\item{filter}{A character vector; only plot length distributions
of transcript sections specified in \code{filter}; if \code{NULL}
plot all sections; default is \code{NULL}.}

\item{outliers}{A logical scalar; if \code{TRUE} show outliers in
boxplot; default is \code{FALSE}.}

\item{...}{Any additional parameters passed to \code{axis}.}
}
\description{
Plot length distribution of transcript sections from list of
GRangesList transcript features.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}

