% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosData-class.R
\docType{methods}
\name{PosData}
\alias{PosData}
\title{PosData}
\usage{
\S4method{parallelSlotNames}{PosData}(x)

\S4method{getListElement}{PosData}(x, i, exact = TRUE)

\S4method{setListElement}{PosData}(x, i, value)

\S4method{[}{PosData,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{gff}{PosData}(x)

\S4method{fasta}{PosData}(x)

\S4method{getSeq}{PosData}(x)

\S4method{ranges}{PosData}(x)

\S4method{bamfiles}{PosData}(x)
}
\description{
title
}
