% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosDataFrame-class.R
\docType{class}
\name{PosDataFrame}
\alias{PosDataFrame}
\alias{PosDataFrame-class}
\alias{sequences,PosDataFrame-method}
\alias{ranges,PosDataFrame-method}
\title{PosDataFrame}
\usage{
PosDataFrame(df, ranges, sequence, replicate, conditions)

\S4method{sequences}{PosDataFrame}(x)

\S4method{ranges}{PosDataFrame}(x)
}
\description{
title
}
