% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosData-ProtectedEndPosData.R
\docType{class}
\name{ProtectedEndPosData}
\alias{ProtectedEndPosData}
\alias{ProtectedEndPosData-class}
\title{ProtectedEndPosData}
\usage{
ProtectedEndPosData(bamfiles, fasta, gff, ...)
}
\description{
title
}
