% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-protected-end-pos.R
\docType{class}
\name{ProtectedEndSequenceData-class}
\alias{ProtectedEndSequenceData-class}
\alias{ProtectedEndSequenceData}
\alias{getData,ProtectedEndSequenceData,BamFileList,GRangesList,XStringSet,ScanBamParam-method}
\alias{aggregateData,ProtectedEndSequenceData-method}
\alias{getDataTrack,ProtectedEndSequenceData-method}
\title{ProtectedEndSequenceData}
\usage{
ProtectedEndSequenceData(bamfiles, annotation, sequences, seqinfo, ...)


  \S4method{getData}{ProtectedEndSequenceData,BamFileList,GRangesList,XStringSet,ScanBamParam}(x,
  bamfiles, grl, sequences, param, args)

\S4method{aggregateData}{ProtectedEndSequenceData}(x,
  condition = c("Both", "Treated", "Control"))

\S4method{getDataTrack}{ProtectedEndSequenceData}(x, name, ...)
}
\arguments{
\item{bamfiles, annotation, seqinfo, grl, sequences, param, args, ...}{See 
\code{\link[=SequenceData-class]{SequenceData}} and
\code{\link[=SequenceData-functions]{SequenceData-functions}}}

\item{x}{a \code{ProtectedEndSequenceData}}

\item{condition}{For \code{\link{aggregate}}: condition for which the data 
should be aggregated.}

\item{name}{For \code{getDataTrack}: a valid transcript name. Must be a name
of \code{ranges(x)}}
}
\value{
a \code{ProtectedEndSequenceData} object
}
\description{
\code{ProtectedEndSequenceData} implements
\code{\link[=SequenceData-class]{SequenceData}} to contain and aggregate the
start and ends of reads per position along the transcripts. 
\code{ProtectedEndSequenceData} behaves specially and offsets the start 
position by -1 to align the information on the 5'-3'-phosphate bonds to one
position. The \code{ProtectedEndSequenceData} class is specifically used
for the RiboMethSeq method.

Both classes contain three columns per data file named using the
following naming convention \code{protectedend.condition.replicate}.

\code{aggregate} calculates the mean and sd for samples in the \code{control}
and \code{treated} condition separatly.
}
\examples{
# Construct a ProtectedEndSequenceData object
RNAmodR.files <- RNAmodR.Data::RNAmodR.files
annotation <- rtracklayer::GFF3File(RNAmodR.files[["example_RMS.gff3"]])
sequences <- Rsamtools::FaFile(RNAmodR.files[["example_RMS.fasta"]])
files <- c(control = RNAmodR.files[["example_RMS1.bam"]],
           treated = RNAmodR.files[["example_RMS2.bam"]])
pesd <- ProtectedEndSequenceData(files, annotation = annotation,
                                 sequences = sequences)
# aggregate data
aggregate(pesd)
}
