% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-protected-end-pos.R
\docType{class}
\name{ProtectedEndSequenceData}
\alias{ProtectedEndSequenceData}
\alias{ProtectedEndSequenceData-class}
\title{ProtectedEndSequenceData}
\usage{
ProtectedEndSequenceData(bamfiles, fasta, gff, ...)
}
\description{
title
}
