% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmodR-quant-5end.R, R/class-RNAmodR-quant-allends.R
\docType{methods}
\name{quantifiyReadDataPerTranscript}
\alias{quantifiyReadDataPerTranscript}
\alias{quantifiyReadDataPerTranscript,RNAmodRquant_5end,GAlignments,RNAmodRargs,numeric,GRanges-method}
\alias{quantifiyReadDataPerTranscript,RNAmodRquant_allends,GAlignments,RNAmodRargs,numeric,GRanges-method}
\title{Returns the data for each transcript}
\usage{
quantifiyReadDataPerTranscript(x, bamData, args, counts, gff)


  \S4method{quantifiyReadDataPerTranscript}{RNAmodRquant_5end,GAlignments,RNAmodRargs,numeric,GRanges}(x,
  bamData, args, counts, gff)


  \S4method{quantifiyReadDataPerTranscript}{RNAmodRquant_allends,GAlignments,RNAmodRargs,numeric,GRanges}(x,
  bamData, args, counts, gff)
}
