% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/class-RNAmodR-quant.R
\docType{methods}
\name{quantifyReadData}
\alias{quantifyReadData}
\alias{quantifyReadData,RNAmodRquant,RNAmodRargs,GRanges,ScanBamParam-method}
\title{Convert reads to data for search for modifications}
\usage{
quantifyReadData(x, args, gff, param)

\S4method{quantifyReadData}{RNAmodRquant,RNAmodRargs,GRanges,ScanBamParam}(x,
  args, gff, param)
}
\arguments{
\item{x}{RNAmodRquant object}

\item{gff}{a GRanges object for the genome}

\item{param}{ScanBamParam to use for loading the bam files}

\item{files}{files which to analyze}
}
\value{
the read data quantified in the flavour of each specific incarnation
}
\description{
load the input files, eg. bam files andpre-processes the information to be 
used in the modification detection.
}
