% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR-data.R
\docType{data}
\name{RDN18}
\alias{RDN18}
\title{Example data for RDN18 gene}
\format{An object of class \code{list} of length 4.}
\usage{
RDN18
}
\description{
contains a list of example data for the RDN18 gene
}
\keyword{RDN18}
\keyword{RNAmod}
\keyword{datasets}
