% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{ReadBED}
\alias{ReadBED}
\title{Read BED-formatted file.}
\usage{
ReadBED(file)
}
\arguments{
\item{file}{Input BED file.}
}
\value{
a GRanges object of features from the BED input file.
}
\description{
Open and read a BED file,  and store the annotation
features in a GRanges object.
}
\examples{

bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR");
sites <- ReadBED(bedFile);

}
\keyword{BED}

