% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{ReadDBN}
\alias{ReadDBN}
\title{Read a DBN file.}
\usage{
ReadDBN(file)
}
\arguments{
\item{file}{A character string; specifies the input DBN file.}
}
\value{
A \code{dataframe} object. See 'Details'.
}
\description{
Read a DBN (dot-bracket) structure file. See 'Details'.
}
\details{
Open and read a DBN file, and return a \code{dataframe}
with the following columns:
\enumerate{
\item Column 1: Sequence ID
\item Column 2: Length of the sequence (in nt)
\item Column 3: Mean free energy (MFE)
}
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

