% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RNAmod.R
\docType{class}
\name{RNAmod-class}
\alias{RNAmod-class}
\alias{RNAmod-class}
\alias{RNAmod}
\alias{show,RNAmod-method}
\title{RNAmod Class}
\usage{
RNAmod(experimentName, experimentLayout, experimentGFF, experimentFasta)

\S4method{show}{RNAmod}(object)
}
\arguments{
\item{object}{a RNAmod object.}
}
\value{
a RNAmod object
}
\description{
\code{RNAmod}
}
