% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RNAmodR.R
\docType{class}
\name{RNAmodR-class}
\alias{RNAmodR-class}
\alias{RNAmodR-class}
\alias{RNAmod}
\alias{show,RNAmodR-method}
\title{RNAmodR Class}
\usage{
RNAmod(experimentName, experimentLayout, experimentGFF, experimentFasta)

\S4method{show}{RNAmodR}(object)
}
\arguments{
\item{object}{a RNAmod object.}
}
\value{
a RNAmodR object
}
\description{
\code{RNAmodR}

title
}
\examples{
\donttest{
unzip(system.file("extdata", package = "RNAmodR", file = "RNAmodR.zip" ))
mod <- RNAmodR("test",
               "test_layout.csv",
               "test_gff.gff3",
               "test_masked.fasta")
}
}
