% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RNAmodR-data-type.R,
%   R/class-RNAmodR-data-5end.R, R/class-RNAmodR-data-allends.R
\docType{class}
\name{RNAmodR-data-class}
\alias{RNAmodR-data-class}
\alias{show,data-method}
\alias{data_5end-class}
\alias{data_allends-class}
\title{RNAmodR data class}
\usage{
\S4method{show}{data}(object)
}
\arguments{
\item{object}{a RNAmodR data object.}
}
\value{
a RNAmodR data class object
}
\description{
The class is virtual and has to be extended from, eg. \code{data_5end}
The data class is the main function for detecting modifications. It works
usually in a three step manner.
\itemize{
  \item{"1."}{\code{convertReadsToPositions}: This function aggregates the
  information of reads per bam file, eg. per replicate and stores them for
  subsequent analysis}
  \item{"2."}{\code{parseMod}: This function searches for a defined pattern. 
  For this \code{mod} class objects will be created on the fly based on the 
  names of the modification one wants to detect and which can understand the
  data aggregated by \code{convertReadsToPositions}.}
  \item{"3."}{\code{mergePositionsOfReplicates}: This function aggregates the
  base data further for storage. This involves usually generating a per
  position mean and sd or equivalent.}
}

\code{data_5end}: the class can be used for analyzing the 5'-end of
reads to detect specific post-transcriptional modifications. This includes
m7G, m3C and Dihydrouridine modifications.

\code{data_allends}: the class can be used for analyzing both the 5'-end 
and 3'-ends of reads to detect specific post-transcriptional modifications.
Methylation of the ribose is detected with this data type.
}
\section{Slots}{

\describe{
\item{\code{plotType}}{}

\item{\code{data}}{}

\item{\code{modifications}}{}
}}

\examples{
\donttest{
ad <- new("data_5end")
}
\donttest{
ad <- new("data_allends")
}
}
