% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.R
\docType{data}
\name{RNAmodR-datasets}
\alias{RNAmodR-datasets}
\alias{msi}
\alias{sds}
\alias{sdl}
\alias{psd}
\alias{e5sd}
\alias{e3sd}
\alias{esd}
\alias{csd}
\alias{ne3sd}
\alias{ne5sd}
\alias{pesd}
\title{Example data in the RNAmodR package}
\format{
\itemize{
\item msi a \code{ModSetInosine} instance
\item sds a \code{SequenceDataSet} instance
\item sdl a \code{SequenceDataList} instance
\item psd a \code{PileupSequenceData} instance
\item e5sd a \code{End5SequenceData} instance
\item e3sd a \code{End3SequenceData} instance
\item esd a \code{EndSequenceData} instance
\item csd a \code{CoverageSequenceData} instance
\item ne3sd a \code{NormEnd3SequenceData} instance
\item ne5sd a \code{NormEnd5SequenceData} instance
\item pesd a \code{ProtectedEndSequenceData} instance
}

An object of class \code{SequenceDataSet} of length 2.

An object of class \code{SequenceDataList} of length 3.

An object of class \code{PileupSequenceData} of dimension 100 x 101 x 15 x 15.

An object of class \code{End5SequenceData} of dimension 100 x 101 x 3 x 3.

An object of class \code{End3SequenceData} of dimension 100 x 101 x 3 x 3.

An object of class \code{EndSequenceData} of dimension 100 x 101 x 3 x 3.

An object of class \code{CoverageSequenceData} of dimension 100 x 101 x 3 x 3.

An object of class \code{NormEnd3SequenceData} of dimension 100 x 101 x 9 x 9.

An object of class \code{NormEnd5SequenceData} of dimension 100 x 101 x 9 x 9.

An object of class \code{ProtectedEndSequenceData} of dimension 100 x 101 x 3 x 3.
}
\usage{
data(msi)

data(sds)

data(sdl)

data(psd)

data(e5sd)

data(e3sd)

data(esd)

data(csd)

data(ne3sd)

data(ne5sd)

data(pesd)
}
\description{
The following datasets are contained in the RNAmodR package. They are used
in the man page examples.
}
\keyword{datasets}
