% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.R, R/AllGenerics.R, R/Modifier-utils.R
\docType{data}
\name{RNAmodR-internals}
\alias{RNAmodR-internals}
\alias{MOD_CONTAINING_FEATURES}
\alias{.dataTrack}
\alias{.constructModRanges}
\alias{.dataTracks}
\alias{.constructModRanges,GRanges,DataFrame-method}
\title{RNAmodR internal functions}
\format{An object of class \code{character} of length 14.}
\usage{
MOD_CONTAINING_FEATURES

.constructModRanges(range, data, modType, scoreFun, source, type)

.dataTracks(x, data, seqdata, sequence, args)

\S4method{.constructModRanges}{GRanges,DataFrame}(range, data, modType,
  scoreFun, source, type)
}
\description{
These functions are not intended for general use, but maybe exported for 
additional package development.
}
\keyword{datasets}
