% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.R
\docType{package}
\name{RNAmodR-package}
\alias{RNAmodR-package}
\title{RNAmodR: Detection of post-transcriptional modifications in high throughput sequencing data}
\description{
RNAmodR provides classes and workflows for loading/aggregation data from high througput sequencing aimed at detecting post-transcriptional modifications through analysis of specific patterns. In addition, utilities are provided to validate and visualize the results. The RNAmodR package provides a core functionality from which specific analysis strategies can be easily implemented as a seperate package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FelixErnst/RNAmodR}
  \item Report bugs at \url{https://github.com/FelixErnst/RNAmodR/issues}
}

}
\author{
\strong{Maintainer}: Felix G.M. Ernst \email{felix.gm.ernst@outlook.com} (\href{https://orcid.org/0000-0001-5064-0928}{ORCID})

Other contributors:
\itemize{
  \item Denis L.J. Lafontaine \email{denis.lafontaine@ulb.ac.be} [contributor, funder]
}

}
\keyword{internal}
