% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/class-RNAmodR-quant.R
\docType{methods}
\name{RNAmodRquant-accessors}
\alias{RNAmodRquant-accessors}
\alias{getPositions}
\alias{getModifications}
\alias{getDataLabel}
\alias{getDataFormat}
\alias{getPlotType,RNAmodRquant-method}
\alias{getPlotType}
\alias{getPositions}
\alias{getModifications}
\alias{getPositions,RNAmodRquant-method}
\alias{getModifications,RNAmodRquant-method}
\alias{getDataLabel,RNAmodRquant-method}
\alias{getDataFormat,RNAmodRquant-method}
\title{Accessor for \code{RNAmodRquant} class objects}
\usage{
getPlotType(x)

getPositions(x)

getModifications(x)

getDataLabel(x)

getDataFormat(x)

\S4method{getPlotType}{RNAmodRquant}(x)

\S4method{getPositions}{RNAmodRquant}(x)

\S4method{getModifications}{RNAmodRquant}(x)

\S4method{getDataLabel}{RNAmodRquant}(x)

\S4method{getDataFormat}{RNAmodRquant}(x)
}
\arguments{
\item{object}{an RNAmodRquant object}
}
\value{
character defining the plot type for this RNAmodRquant class

a list of data as a lists of list(replicate) of DataFrame(transcript)

a list of data as a lists of list(replicate) of DataFrame(transcript)

the description of the type of position data, which can be used eg. 
as a label for printing

a function for formating the label of the position data
}
\description{
The accessor function to \code{RNAmodRquant} class objects can be used to access
the RNAmodRquant saved in slots of the object. See examples for available functions.
}
\examples{
\donttest{
getPlotType(quant)
getPositions(quant)
getModifications(quant)
}
}
