% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gviz-ModifiedSequenceTrack-class.R
\docType{class}
\name{RNASequenceTrack}
\alias{RNASequenceTrack}
\alias{RNASequenceTrack-class}
\title{RNASequenceTrack}
\usage{
RNASequenceTrack(sequence, chromosome, genome, ...)
}
\arguments{
\item{sequence}{A \code{character} vector or \code{RNAString} object of 
length one. The sequence to display.}

\item{...}{Additional items which will all be interpreted as display
parameters.}

\item{name}{A \code{character}. The name of the track used in the title panel
when plotting}
}
\description{
A \code{Gviz} compatible 
\code{\link[Gviz:SequenceTrack-class]{SequenceTrack}} for showing RNA 
sequences.
}
\examples{


}
