% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-roc.R
\docType{methods}
\name{roc}
\alias{roc}
\alias{roc,Modifier-method}
\alias{roc,ModifierSet-method}
\title{Performance}
\usage{
\S4method{roc}{Modifier}(x, coord, redo, redo.args, prediction.args,
  performance.args, plot.args)

\S4method{roc}{ModifierSet}(x, coord, redo, redo.args, prediction.args,
  performance.args, plot.args)
}
\description{
title
}
