% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{SafeLoad}
\alias{SafeLoad}
\title{Safe-loading an R package.}
\usage{
SafeLoad(lib)
}
\arguments{
\item{lib}{A character string; package to be loaded.}
}
\value{
A logical scalar.
}
\description{
Safe-loading an R package. See 'Details'.
}
\details{
The function calls \code{require} to load a package.
If the package can be loaded it returns \code{TRUE},
else \code{FALSE}.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
