% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmodR.R
\docType{methods}
\name{saveModifications}
\alias{saveModifications}
\alias{saveModifications}
\alias{saveModifications,RNAmodR,numeric,GRangesList-method}
\alias{loadModifications}
\alias{saveModifications}
\title{Saves and loads identified modifications}
\usage{
saveModifications(x, number, grl)

loadModifications(x, number, geneNames = NULL)

\S4method{saveModifications}{RNAmodR,numeric,GRangesList}(x, number, grl)
}
\arguments{
\item{x}{RNAmodR}

\item{grl}{GRangesList with modifications found per gene}

\item{geneNames}{optional: gene names for which modifications should be loaded}
}
\value{
GRangesList
}
\examples{
\donttest{
saveModifications(experiment, grl)
loadModifications(experiment, "RNA28S5")
}
}
