% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-RNAmodR.R
\docType{methods}
\name{seGetModifications,SummarizedExperiment-method}
\alias{seGetModifications,SummarizedExperiment-method}
\title{Title}
\usage{
\S4method{seGetModifications}{SummarizedExperiment}(se, genes, modifications)
}
\arguments{
\item{se}{SummarizedExperiment.}
}
\description{
Title
}
