% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmodR.R
\docType{methods}
\name{seGetModifications}
\alias{seGetModifications}
\alias{seGetPositions}
\alias{seGetModifications,SummarizedExperiment-method}
\alias{seGetModifications}
\alias{seGetPositions}
\alias{seGetPositions,SummarizedExperiment-method}
\title{Getting Position and Modification data from SummarizedExperiment 
objects}
\usage{
seGetModifications(se, genes, modifications)

seGetPositions(se, genes)

\S4method{seGetModifications}{SummarizedExperiment}(se, genes, modifications)

\S4method{seGetPositions}{SummarizedExperiment}(se, genes)
}
\arguments{
\item{se}{a SummarizedExperiment object}

\item{genes}{gene names used for subsetting the results}

\item{modifications}{\code{seGetModifications} modification types used for 
subsetting the results}
}
\value{
DataFrame containing the information on saved modifications

a list of DataFrames containing the position data
}
\description{
\code{seGetModifications} and \code{seGetPositions} returns modification and
position data, respectively, from a SummarizedExperiment object. The function
expects the data to be stored as provided and defined by the RNAmodR package 
as rowData and assay data.
}
\examples{
\donttest{
seGetModifications(se)
seGetModifications(se,"RDN18-1")
seGetModifications(se,"RDN18-1","m7G")
}
\donttest{
seGetPositions(se)
seGetPositions(se,"RDN18-1")
}
}
