% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-RNAmodR.R
\docType{methods}
\name{seGetPositions,SummarizedExperiment-method}
\alias{seGetPositions,SummarizedExperiment-method}
\title{Title}
\usage{
\S4method{seGetPositions}{SummarizedExperiment}(se, genes)
}
\arguments{
\item{se}{SummarizedExperiment.}
}
\description{
Title
}
