% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceDataFrame-class.R
\docType{class}
\name{SequenceDataFrame}
\alias{SequenceDataFrame}
\alias{SequenceDataFrame-class}
\alias{sequences,SequenceDataFrame-method}
\alias{ranges,SequenceDataFrame-method}
\title{SequenceDataFrame}
\usage{
SequenceDataFrame(df, ranges, sequence, replicate, conditions)

\S4method{sequences}{SequenceDataFrame}(x)

\S4method{ranges}{SequenceDataFrame}(x)
}
\description{
title
}
