% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceDataList-class.R
\docType{class}
\name{SequenceDataList-class}
\alias{SequenceDataList-class}
\alias{SequenceDataList}
\title{The SequenceDataList class}
\usage{
SequenceDataList(...)
}
\arguments{
\item{...}{The elements to be included in the \code{SequenceDataList}.}
}
\value{
a \code{SequenceDataList}
}
\description{
The \code{SequenceDataList} class is used to hold \code{SequenceData} objects
as its elements. It is derived from the 
\code{\link[S4Vectors:List-class]{List}}.
}
\examples{
data(psd,package="RNAmodR")
data(e5sd,package="RNAmodR")
sdl <- SequenceDataList(psd,e5sd)
}
