% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceDataList-class.R
\docType{class}
\name{SequenceDataList}
\alias{SequenceDataList}
\alias{SequenceDataList-class}
\title{SequenceDataList}
\usage{
SequenceDataList(...)

\S4method{seqinfo}{SequenceDataList}(x)

\S4method{sequences}{SequenceDataList}(x)

\S4method{ranges}{SequenceDataList}(x)

\S4method{bamfiles}{SequenceDataList}(x)

\S4method{aggregate}{SequenceDataList}(x, condition = "Treated")
}
\description{
title

default names are class names
}
