% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceDataSet-class.R
\docType{class}
\name{SequenceDataSet-class}
\alias{SequenceDataSet-class}
\alias{SequenceDataSet}
\title{The SequenceDataSet class}
\usage{
SequenceDataSet(...)
}
\arguments{
\item{...}{The elements to be included in the \code{SequenceDataSet}.}
}
\value{
a \code{SequenceDataSet}
}
\description{
The \code{SequenceDataSet} class is used to hold \code{SequenceData} objects
as its elements. It is derived from the 
\code{\link[S4Vectors:List-class]{List}} class.

The \code{SequenceDataSet} is used to hold different data types from the of
same aligned reads. The same dataset can be used to generate multiple sets of
data types. Bam files, annotation and sequence data must be the same for all 
elements.
}
\examples{
data(psd,package="RNAmodR")
data(e5sd,package="RNAmodR")
sdl <- SequenceDataSet(psd,e5sd)
}
