% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmod.R
\docType{methods}
\name{setGff}
\alias{setGff}
\alias{setGff,RNAmod,GRanges,numeric,character-method}
\title{setGff}
\usage{
setGff(.Object, gff, number, modification)

\S4method{setGff}{RNAmod,GRanges,numeric,character}(.Object, gff, number,
  modification)
}
\arguments{
\item{.Object}{RNAmod.}

\item{gff}{GRanges.}

\item{number}{numeric.}

\item{modification}{character.}
}
\value{
GRanges object
}
\description{
setGff
}
\examples{
\donttest{
setGff(experiment, gff, 1, c("m7G"))
}
}
