% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmodR.R
\docType{methods}
\name{setGffResult}
\alias{setGffResult}
\alias{setGffResult,RNAmodR,GRanges,numeric,character-method}
\title{setGffResult}
\usage{
setGffResult(.Object, gff, number, modification)

\S4method{setGffResult}{RNAmodR,GRanges,numeric,character}(.Object, gff, number,
  modification)
}
\arguments{
\item{.Object}{RNAmod.}

\item{gff}{GRanges.}

\item{number}{numeric.}

\item{modification}{character.}
}
\value{
GRanges object
}
\description{
setGff
}
\examples{
\donttest{
setGffResult(experiment, gff, 1, c("m7G"))
}
}
