% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{SetId}
\alias{SetId}
\alias{SetId,txLoc-method}
\title{Method "SetId" for S4 object txLoc.}
\usage{
SetId(x, id)

\S4method{SetId}{txLoc}(x, id)
}
\arguments{
\item{x}{A \code{txLoc} object.}

\item{id}{A character string.}
}
\description{
Set the id string of a \code{txLoc} object.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

