% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmodR.R
\docType{methods}
\name{setSummarizedExperiment}
\alias{setSummarizedExperiment}
\alias{setSummarizedExperiment,RNAmodR,RangedSummarizedExperiment,numeric-method}
\title{sets a SummarizedExperiment object}
\usage{
setSummarizedExperiment(.Object, se, number)


  \S4method{setSummarizedExperiment}{RNAmodR,RangedSummarizedExperiment,numeric}(.Object,
  se, number)
}
\arguments{
\item{.Object}{a RNAmod object}

\item{se}{a RangedSummarizedExperiment object}

\item{number}{a number defining the experiment}
}
\value{
the RpfSummarizedExperiment
}
\description{
Saves/overwrites a SummarizedExperiment object for certain experiment and 
passes the SummarizedExperiment object on to be saved to disk as .RData file
in the \code{results\\SE} folder
}
\examples{
\donttest{
setSummarizedExperiment(experiment, se, 1, c("m7G"))
}
}
