% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/RNAmodR-utils.R
\docType{methods}
\name{setupWorkEnvir}
\alias{setupWorkEnvir}
\alias{setupWorkEnvir,character-method}
\title{setupWorkEnvir}
\usage{
setupWorkEnvir(experimentName)

\S4method{setupWorkEnvir}{character}(experimentName = "experiment")
}
\arguments{
\item{experimentName}{a name for the experiment, e.g. "xyz"}
}
\value{
TRUE
}
\description{
conveniance function for setting up folders for an experiment
}
\examples{
experimentName <- "test"
setupWorkEnvir(experimentName)
}
