% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{SetVersion}
\alias{SetVersion}
\alias{SetVersion,txLoc-method}
\title{Method "SetVersion" for S4 object txLoc.}
\usage{
SetVersion(x, version)

\S4method{SetVersion}{txLoc}(x, version)
}
\arguments{
\item{x}{A \code{txLoc} object.}

\item{version}{A character string.}
}
\description{
Set the version string of a \code{txLoc} object.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

