% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{show,txLoc-method}
\alias{show,txLoc-method}
\title{Method "show" for S4 object txLoc.}
\usage{
\S4method{show}{txLoc}(object)
}
\arguments{
\item{x}{A \code{txLoc} object.}
}
\description{
Print summary information about \code{txLoc} object.
}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR")
sites <- ReadBED(bedFile)
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38")
posSites
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
