% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{SmartMap}
\alias{SmartMap}
\title{Map genome coordinates to transcript coordinates.}
\usage{
SmartMap(gr, id = NULL, refGenome = "hg38", ignore.strand = FALSE)
}
\arguments{
\item{gr}{GRanges object of genomic features}

\item{id}{Identifier for loci from \code{gr}. If NULL then id = "".
Default is NULL.}

\item{refGenome}{Reference transcriptome, defaults to hg38.}

\item{ignore.strand}{Ignore strand when mapping genome coordinates to
transcript coordinates. Default is FALSE.}
}
\value{
A \code{txLoc} object with coordinates from \code{gr} mapped to
the transcriptome based on assembly version \code{refGenome}.
}
\description{
Map genome coordinates to transcript coordinates.
}
\examples{

bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR");
sites <- ReadBED(bedFile);
txSites <- SmartMap(sites, id = "m6A", refGenome = "hg38");
}

