% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{SmartMap.ToTx}
\alias{SmartMap.ToTx}
\title{Map genome coordinates to transcript coordinates.}
\usage{
SmartMap.ToTx(locus, txFeatures, txSequences, geneXID, ignore.strand = FALSE,
  noFactors = TRUE)
}
\arguments{
\item{locus}{GRanges object of genomic features.}

\item{txFeatures}{List of GRangesList transcript features.}

\item{txSequences}{List of DNAStringSet sequences.}

\item{geneXID}{Dataframe of cross-referencing gene IDs.}

\item{ignore.strand}{Ignore strand when mapping genome coordinates to
transcript coordinates. Default is FALSE.}

\item{noFactors}{No (character) factors in output. Default is TRUE.}
}
\value{
A list of dataframes with mapped features in transcript region.
}
\description{
Map genome coordinates to transcript coordinates. This function
should not be invoked by the end-user directly. It is called
from within \code{SmartMap}.
}
\examples{

}
\keyword{}

