% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{SubsampleTxLoc}
\alias{SubsampleTxLoc}
\title{Subsample from a \code{txLoc} object.}
\usage{
SubsampleTxLoc(locus, fraction = 0)
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{fraction}{A float scalar;}
}
\value{
A \code{txLoc} object.
}
\description{
Subsample from a \code{txLoc} object.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
