% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{test.MotifSearch}
\alias{test.MotifSearch}
\title{Experimental motif search}
\usage{
test.MotifSearch(locus, flank = 100, filter = NULL, maxMM = 1)
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{flank}{An integer scalar; see 'Details'.}

\item{filter}{A character vector; only plot sequence logos
of sections specified in \code{filter}; if \code{NULL} plot
all sections; default is NULL.}

\item{maxMM}{An integer scalar; specifies the maximum number
of mismatch(es) allowed during the motif search; default is 1.}
}
\description{
Experimental motif search
}

