% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{txLoc-accessors}
\alias{txLoc-accessors}
\alias{show,txLoc-method}
\alias{GetLoci,txLoc-method}
\alias{GetId,txLoc-method}
\alias{GetRef,txLoc-method}
\alias{GetVersion,txLoc-method}
\alias{GetNumberOfLoci,txLoc-method}
\alias{GetRegions,txLoc-method}
\title{\code{txLoc} accessors.}
\usage{
\S4method{show}{txLoc}(object)

\S4method{GetLoci}{txLoc}(object)

\S4method{GetId}{txLoc}(object)

\S4method{GetRef}{txLoc}(object)

\S4method{GetVersion}{txLoc}(object)

\S4method{GetNumberOfLoci}{txLoc}(object)

\S4method{GetRegions}{txLoc}(object)
}
\arguments{
\item{object}{A \code{txLoc} object.}
}
\description{
Various \code{txLoc} accessors.
}
\details{
\itemize{
    \item \code{show(object)}: Print summary information about the 
    \code{txLoc} object.
    \item \code{GetLoci(object)}: Get a \code{list} of \code{DataFrame} 
    objects of the loci of the \code{txLoc} object.
    \item \code{GetId(object)}: Get the identifier string of the \code{txLoc} 
    object.
    \item \code{GetRef(object)}: Get the reference genome string of the
    \code{txLoc} object.
    \item \code{GetVersion(object)}: Get the version string of a \code{txLoc} 
    object.
    \item \code{GetNumberOfLoci(object)}: Get the number of loci of a 
    \code{txLoc} object in every transcript section.
    \item \code{GetRegions(objects)}: Character vector of the transcript 
    regions of a \code{txLoc} object.
}
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
