% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{TxLoc2GRangesList}
\alias{TxLoc2GRangesList}
\title{Convert a \code{txLoc} object to a \code{GRangesList} object.}
\usage{
TxLoc2GRangesList(txLoc, method = c("tx_region", "genome"))
}
\arguments{
\item{txLoc}{A \code{txLoc} object.}

\item{method}{A character string; specifies whether coordinates
are genome (\code{method = "genome"}) or transcriptome coordinates
(\code{method = "tx_region"}).}
}
\value{
A \code{list} of \code{GRanges} objects. See 'Details'.
}
\description{
Convert a \code{txLoc} object to a \code{GRangesList} object.
See 'Details'.
}
\details{
The function converts a \code{txLoc} to a \code{GRangesList}
object. Coordinates can be either genomic coordinates
(\code{method = "genome"}) or transcript region coordinates
(\code{method = "tx_region"}).
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
