% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{Unfactor}
\alias{Unfactor}
\title{Unfactor entries in a \code{data.frame}.}
\usage{
Unfactor(df)
}
\arguments{
\item{df}{A \code{data.frame} object.}
}
\value{
A \code{data.frame} object.
}
\description{
Unfactor entries in a \code{data.frame}.
}
\keyword{internal}

