% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-Inosine-viz.R, R/Modifier-viz.R
\docType{methods}
\name{visualizeDataByCoord,ModInosine,GRanges-method}
\alias{visualizeDataByCoord,ModInosine,GRanges-method}
\alias{visualizeDataByCoord}
\alias{visualizeDataByCoord,Modifier,GRanges-method}
\title{visualizeDataByCoord}
\usage{
\S4method{visualizeDataByCoord}{ModInosine,GRanges}(x, coord, type = NA,
  window.size = 15L, ...)

\S4method{visualizeDataByCoord}{Modifier,GRanges}(x, coord, type = NA,
  window.size = 15L, ...)
}
\arguments{
\item{x}{a \code{Modifier} or \code{ModifierSet} object.}

\item{coord}{coordinates of a single position to subset to as a 
\code{GRanges} object. The Parent column is expected to match the gene or 
transcript name.}

\item{type}{the data type of data show as data tracks.}

\item{window.size}{integer value for the number of positions on the left and 
right site of the selected positions included in the plotting (default: 
\code{window.size = 15L})}

\item{...}{optional parameters:
\itemize{
\item{\code{modified.seq}}{\code{TRUE} or \code{FALSE}. Should the sequence 
shown with modified nucleotide positions? (default: 
\code{modified.seq = FALSE})}
\item{\code{additional.mod}}{other modifications, which should be shown
in the annotation and sequence track.}
\item{\code{annotation.track.pars}}{Parameters passed onto the A
nnotationTrack.}
\item{\code{sequence.track.pars}}{Parameters passed onto the SequenceTrack.}
\item{\code{data.track.pars}}{Parameters passed onto the DataTrack(s).}
}}
}
\description{
title
}
\details{
\code{ModInosine} specific arguments:
\itemize{
\item{colour.bases}{a named character vector of \code{length = 4} for the
colours of the individual bases. The names are expected to be 
\code{c("G","A","U","C")}}
}
}
