% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmod-plot.R
\docType{methods}
\name{visualizeModifications}
\alias{visualizeModifications}
\alias{visualizeModifications}
\title{visualizeModifications}
\usage{
visualizeModifications(.Object, number, modifications, genes)

\S4method{visualizeModifications}{RNAmod,numeric,character,character}(.Object,
  number, modifications, genes)
}
\arguments{
\item{.Object}{a RNAmod object.}

\item{number}{a number defining an experiment.}

\item{modifications}{name of modification to be used for analysis.}
}
