% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{WriteBED}
\alias{WriteBED}
\title{Write \code{txLoc} object to a BED file.}
\usage{
WriteBED(txLoc, file = NULL)
}
\arguments{
\item{txLoc}{A \code{txLoc} object.}

\item{file}{A character string; specifies the filename of the output 
BED file. If \code{NULL}, then \code{file = "sites.bed"}; default is
\code{NULL}.}
}
\description{
Write \code{txLoc} object to a BED file. See 'Details'.
}
\details{
The function writes entries from a \code{txLoc} object to a 6-column BED 
file (BED6). Note that this process is not "splice-aware", i.e. if an entry 
spans an intron the BED entry gives the left and right-most genomic 
coordinate of the feature. If \code{file = NULL}, entries will be written to  
\code{sites.bed}.
}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR")
sites <- ReadBED(bedFile)
txSites <- SmartMap(sites, id = "m6A", refGenome = "hg38")
WriteBED(txSites)
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
