% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{WriteCSV}
\alias{WriteCSV}
\title{Write txLoc object to CSV file.}
\usage{
WriteCSV(txLoc, file = NULL, withSeq = FALSE, withGC = FALSE)
}
\arguments{
\item{txLoc}{A \code{txLoc} object.}

\item{file}{Filename of output CSV file. If NULL then file = "sites.csv".}

\item{withSeq}{If TRUE then include full sequence. Default is FALSE.}

\item{withGC}{If TRUE then include GC content. Default is FALSE. Unused.}
}
\description{
Write a \code{txLoc} object to a comma-separated-values file.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
