% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{WriteFeatToBED}
\alias{WriteFeatToBED}
\title{Write a list of GRangesList to BED file}
\usage{
WriteFeatToBED(txFeatures)
}
\arguments{
\item{txFeatures}{List of GRangesList.}
}
\description{
Write a list of GRangesList transcript features to a BED file.
One file per transcript region.
}

