% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{WriteLocusToBED}
\alias{WriteLocusToBED}
\title{Write txLoc object to BED file.}
\usage{
WriteLocusToBED(locus, file = NULL, formatChr = "noChrName")
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{file}{Filename of output BED file. If NULL then file = "sites.bed".}

\item{formatChr}{Set output format of chromosome column. Default is
"noChrName".}
}
\description{
Write a txLoc object to a/multiple BED file(s).
}
\examples{

}
\keyword{BED}

