% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{WriteLocusToCSV}
\alias{WriteLocusToCSV}
\title{Write txLoc object to CSV file.}
\usage{
WriteLocusToCSV(locus, file = NULL, withSeq = FALSE, withGC = FALSE)
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{file}{Filename of output CSV file. If NULL then file = "sites.csv".}

\item{withSeq}{If TRUE then include full sequence. Default is FALSE.}

\item{withGC}{If TRUE then include GC content. Default is FALSE.}
}
\description{
Write a txLoc object to a/multiple comma-separated-values file(s).
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

