% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR-data.R
\docType{data}
\name{YAL030W}
\alias{YAL030W}
\title{Example data for YAL030W gene}
\format{An object of class \code{list} of length 4.}
\usage{
YAL030W
}
\keyword{RNAmod}
\keyword{YAL030W}
\keyword{datasets}
