% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmod-data.R
\docType{data}
\name{YBR041W}
\alias{YBR041W}
\title{Example data for YBR041W gene}
\format{An object of class \code{list} of length 4.}
\usage{
YBR041W
}
\keyword{RNAmod}
\keyword{YBR041W}
\keyword{datasets}
