% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{AddAlpha}
\alias{AddAlpha}
\title{Add transparency to a list of hex colors}
\usage{
AddAlpha(hexList, alpha = 0.5)
}
\arguments{
\item{hexList}{A vector or list of character strings.}

\item{alpha}{A real scalar; specifies the transparency; default
is \code{alpha = 0.5}.}
}
\description{
Add transparency to a list of colors
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}

