% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Modifier-class.R
\docType{methods}
\name{aggregate}
\alias{aggregate}
\alias{aggregateData}
\alias{hasAggregateData}
\alias{aggregateData,Modifier-method}
\alias{hasAggregateData,Modifier-method}
\title{Aggreagte data per positions}
\usage{
aggregate(x, ...)

aggregateData(x)

hasAggregateData(x, ...)

\S4method{aggregate}{Modifier}(x, force = FALSE)

\S4method{aggregateData}{Modifier}(x)

\S4method{hasAggregateData}{Modifier}(x)
}
\arguments{
\item{x}{a \code{\link{SequenceData}} or \code{\link{Modifier}} object.}

\item{force}{whether to recreate the aggregated data, if it is already stored
inside the \code{Modifier} object.}
}
\value{
\itemize{
\item{\code{aggregate}: }{for \code{SequenceData} the aggregated data is
returned as a \code{SplitDataFrameList} with an element per transcript, 
whereas for a \code{Modifier} the modified input object is returned, 
containing the aggregated data, which can be accessed using 
\code{aggregateData}.}
\item{\code{aggregateData}: }{only for \code{Modifier}: a 
\code{SplitDataFrameList} with an element per transcript is returned. If the 
aggregated data is not stored in the object, it is generated on the fly, but 
does not persist}
\item{\code{hasAggregateData}: }{TRUE or FALSE. Does the \code{Modifier} 
object already contain aggregated data?}
}
}
\description{
The aggregate function is defined per \code{\link{SequenceData}} object and
can be triggered by either using the a \code{\link{SequenceData}} or 
\code{\link{Modifier}}. For the letter it will just redirect to the 
\code{\link{SequenceData}} object, but will store the result. The data
is then used for subsequent tasks, such as search for modifications and 
visualization of the results.
}
