% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Modifier-class.R
\docType{methods}
\name{aggregateData}
\alias{aggregateData}
\alias{hasAggregateData}
\alias{aggregate}
\alias{hasAggregateData,Modifier-method}
\title{aggregate}
\usage{
aggregateData(x)

hasAggregateData(x, ...)

aggregate(x, ...)

\S4method{aggregateData}{Modifier}(x)

\S4method{aggregate}{Modifier}(x)

\S4method{hasAggregateData}{Modifier}(x)
}
\description{
title
}
