% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmod-utils-preparations.R
\name{appendFastaFiles}
\alias{appendFastaFiles}
\alias{appendGFF}
\alias{appendGFF}
\title{appendFastaFiles}
\usage{
appendFastaFiles(..., ident, msg)

appendGFF(..., ident, msg)
}
\arguments{
\item{...}{}

\item{ident}{}

\item{msg}{}
}
\value{
file name of the saved file
}
\description{
append one or more fasta or gff files
}
\examples{
\donttest{
appendFastaFiles(fafile, 
                 fafile2, 
                 ident = "test", 
                 msg = "Test fasta file saved: ")
appendGFF(gfffile, 
          gfffile2, 
          ident = "test", 
          msg = "Test gff3 file saved: ")
}
}
