% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTxMethods.R
\name{BuildTx}
\alias{BuildTx}
\title{Build a custom transcriptome.}
\usage{
BuildTx(genomeVersion = c("hg38", "hg19", "hg18", "mm10", "mm9", "mm8",
  "dm6", "dm3", "dm2", "sacCer3", "sacCer2", "sacCer1"), force = FALSE)
}
\arguments{
\item{genomeVersion}{A character string; refers to a specific
reference genome assembly version; default is \code{"hg38"}.}

\item{force}{A logical scalar; if \code{TRUE} force rebuild of
transcriptome; this will overwrite existing data.}
}
\description{
Build a custom, organism-specific transcriptome. See 'Details'.
}
\details{
The function builds an organism-specific transcriptome containing
one transcript per unique Entrez ID; the transcript is selected
from all UCSC RefSeq annotation-based isoforms as the transcript
with the longest CDS, and longest upstream/downstream adjoining
UTRs. Transcript segments are stored per transcript section, and
written into a \code{.RData} file.
For most operations, the user will run this function once, and
continue with further downstream analyses. Various RNAModR
routines will automatically load the transcriptome data to e.g.
map sites to and from the transcriptome.
Currently, RNAModR supports analyses of human, mouse, fruitfly
and yeast data, based on different reference genome versions:
\itemize{
\item Homo sapiens: hg38, hg19, hg18
\item Mus musculus: mm10, mm9, mm8
\item Drosophila melanogaster: dm6, dm3, dm2
\item Cerevisiae saccharomyces: sacCer3, sacCer2, sacCer1
}
Reconstruction of existing transcriptome data can be achieved
by running \code{BuildTx} with \code{force = TRUE}. Note that
this will overwrite the existing RData file.
Running \code{BuildTx} with \code{sanityCheck = TRUE} performs
additional checks of the various transcriptome components, and
is intended for debugging purposes only. It is usually safe to
run with the default \code{sanityCheck = FALSE}.
}
\examples{
\dontrun{
# Build the human hg38-based reference transcriptome
BuildTx("hg38");
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
