% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTxMethods.R
\name{BuildTx}
\alias{BuildTx}
\title{Build a custom transcriptome.}
\usage{
BuildTx(genomeVersion = c("hg38", "hg19", "mm10", "mm9"), force = FALSE,
  sanityCheck = FALSE)
}
\arguments{
\item{genomeVersion}{A character string; refers to a specific
reference genome assembly version; default is \code{"hg38"}.}

\item{force}{A logical scalar; if \code{TRUE} force rebuild of
transcriptome; this will overwrite existing data.}

\item{sanityCheck}{A logical scalar; if \code{TRUE} perform
sanity checks.}
}
\description{
Build a custom transcriptome. See 'Details'.
}
\details{
The function builds a custom transcriptome that consists of
one transcript per unique Entrez ID; the transcript is
determined from all RefSeq-based isoforms as the transcript
with the longest CDS, and adjoining upstream/downstream longest
UTRs. Transcript segments are stored per transcript section, and
written to a \code{.RData} file.
}

