% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{CheckClass}
\alias{CheckClass}
\title{Check if object has class classType.}
\usage{
CheckClass(object, classType = NULL)
}
\arguments{
\item{object}{An R object.}

\item{classType}{String of class type.}
}
\value{
Logical. \code{TRUE} if object is of class classType.
}
\description{
Check if object has class classType.
}
\keyword{internal}

