% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{CheckClassTxLocConsistency}
\alias{CheckClassTxLocConsistency}
\title{Check if two txLoc objects are consistent.}
\usage{
CheckClassTxLocConsistency(obj1, obj2)
}
\arguments{
\item{obj1}{A txLoc object.}

\item{obj2}{A txLoc object.}
}
\value{
Logical. \code{TRUE} if obj1 and obj2 are consistent.
}
\description{
Check if two txLoc objects are consistent, i.e. are based on the same
reference genome and contain the same transcript regions.
}
\keyword{internal}

