% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/class-RNAmodR-mod-D.R,
%   R/class-RNAmodR-mod-m3C.R, R/class-RNAmodR-mod-m7G.R
\docType{methods}
\name{checkForModification}
\alias{checkForModification}
\alias{checkForModification,mod_D,numeric,numeric,list,numeric-method}
\alias{checkForModification,mod_m3C,numeric,numeric,list,numeric-method}
\alias{checkForModification,mod_m7G,numeric,numeric,list,numeric-method}
\title{checkForModification}
\usage{
checkForModification(object, location, locations, data, nreads)

\S4method{checkForModification}{mod_D,numeric,numeric,list,numeric}(object,
  location, locations, data, nreads)

\S4method{checkForModification}{mod_m3C,numeric,numeric,list,numeric}(object,
  location, locations, data, nreads)

\S4method{checkForModification}{mod_m7G,numeric,numeric,list,numeric}(object,
  location, locations, data, nreads)
}
\description{
\code{mod_D}

\code{mod_m3C}

\code{mod_m7G}
}
