% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ModifierSet-comparison.R
\docType{methods}
\name{compare}
\alias{compare}
\alias{compareByCoord}
\alias{plotCompare}
\alias{plotCompareByCoord}
\alias{compareByCoord,ModifierSet,GRanges-method}
\alias{compareByCoord,ModifierSet,GRangesList-method}
\alias{plotCompareByCoord,ModifierSet,GRanges-method}
\alias{plotCompareByCoord,ModifierSet,GRangesList-method}
\title{Comparison of Samples}
\usage{
compare(x, name, from = 1L, to = 30L, ...)

compareByCoord(x, coord, ...)

plotCompare(x, name, from = 1L, to = 30L, normalize, ...)

plotCompareByCoord(x, coord, normalize, ...)

\S4method{compareByCoord}{ModifierSet,GRanges}(x, coord, ...)

\S4method{compareByCoord}{ModifierSet,GRangesList}(x, coord, ...)

\S4method{plotCompareByCoord}{ModifierSet,GRanges}(x, coord, normalize,
  ...)

\S4method{plotCompareByCoord}{ModifierSet,GRangesList}(x, coord, normalize,
  ...)
}
\arguments{
\item{x}{a \code{Modifier} or \code{ModifierSet} object.}

\item{name}{Only for \code{compare}: the transcript name}

\item{from}{Only for \code{compare}: start position}

\item{to}{Only for \code{compare}: end position}

\item{...}{optional parameters:
\itemize{
\item{\code{alias}} {a data.frame with two columns, \code{tx_id} and 
\code{name}, to convert transcipt ids to another identifier}
\item{\code{name}} {Limit results to one specific gene or transcript}
\item{...} {passed on to \code{\link{subsetByCoord}}}
}}

\item{coord}{coordinates of position to subset to. Either a \code{GRanges} or
a \code{GRangesList} object. For both types the Parent column is expected to
match the gene or transcript name.}

\item{normalize}{either a single logical or character value. If it is a 
character, it must match one of the names in the \code{ModifierSet}.}
}
\value{
\code{compareByCoord} returns a 
\code{\link[IRanges:DataFrameList-class]{DataFrameList}} and
\code{plotCompareByCoord} returns a \code{ggplot} object, which can be
modified further. The \code{\link{DataFrameList}} contains columns per sample
as well as \code{names}, \code{positions} and \code{mod} incorporated from
the \code{coord} input. If \code{coord} contains a column \code{Activity}
this is included in the results as well.
}
\description{
To compare data of different samples, a
\code{\link[=ModifierSet-class]{ModifierSet}} is used. To select the data
alongside the transcripts and their positions a
\code{\link[GenomicRanges:GRanges-class]{GRanges}} or a
\code{\link[GenomicRanges:GRanges-class]{GRangesList}} needs to be supplied.
In case of a \code{GRanges} object, the parent column must match the
transcript names as defined by the out put of \code{ranges(x)}, whereas in
case of a \code{GRangesList} the element names must match the transcript
names.
}
