% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifierSet-comparison.R
\docType{methods}
\name{compareByCoord}
\alias{compareByCoord}
\alias{compareByCoord,ModifierSet,GRanges-method}
\alias{compareByCoord,ModifierSet,GRangesList-method}
\alias{plotCompareByCoord,ModifierSet,GRanges-method}
\alias{plotCompareByCoord,ModifierSet,GRangesList-method}
\title{compareByCoord}
\usage{
\S4method{compareByCoord}{ModifierSet,GRanges}(x, coord, ...)

\S4method{compareByCoord}{ModifierSet,GRangesList}(x, coord, ...)

\S4method{plotCompareByCoord}{ModifierSet,GRanges}(x, coord, ...)

\S4method{plotCompareByCoord}{ModifierSet,GRangesList}(x, coord, ...)
}
\arguments{
\item{...}{optional parameters:
\itemize{
\item{\code{name}}{Limit results to one specific gene or transcript}
\item{...}{passed on to \code{\link{subsetByCoord}}}
}}
}
\description{
title
}
