% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmod-utils-preparations.R
\name{convertGeneToChrom}
\alias{convertGeneToChrom}
\alias{convertGeneTypeToChrom}
\alias{convertGeneToChrom}
\alias{converttRNAscanToChrom}
\title{convert gene sequences to individual chromosomes}
\usage{
convertGeneTypeToChrom(gfffile, fafile, geneTypes, ident,
  saveTrimmedFile = FALSE, appendToOriginal = FALSE,
  removeDuplicateSeqs = FALSE)

convertGeneToChrom(gfffile, fafile, genes, ident, saveTrimmedFile = FALSE,
  appendToOriginal = FALSE, removeDuplicateSeqs = FALSE)

converttRNAscanToChrom(gfffile, fafile, tRNAscanfile, ident,
  saveTrimmedFile = FALSE, appendToOriginal = FALSE)
}
\arguments{
\item{gfffile}{}

\item{fafile}{}

\item{geneTypes}{}

\item{ident}{}

\item{saveTrimmedFile}{}

\item{appendToOriginal}{}

\item{removeDuplicateSeqs}{}

\item{genes}{}

\item{tRNAscanfile}{}
}
\examples{
\donttest{
gfffile <-
fafile <-
convertGeneTypeToChrom(gfffile,
                       fafile,
                       "tRNA_gene",
                       "tRNA")
convertGeneToChrom(gfffile,
                   fafile,
                   "RDN18-1",
                   "rRNA")
}
}
