% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmod-m7G.R
\docType{methods}
\name{convertReadsToPositions}
\alias{convertReadsToPositions}
\alias{convertReadsToPositions,mod_m7G,numeric,GRanges,DNAString,DataFrame-method}
\title{convertReadsToPositions}
\usage{
convertReadsToPositions(object, counts, gff, seq, data)


  \S4method{convertReadsToPositions}{mod_m7G,numeric,GRanges,DNAString,DataFrame}(object,
  counts, gff, seq, data)
}
\arguments{
\item{object}{mod object}

\item{counts}{total counts in the BAM file containing the data}

\item{gff}{a GRanges object for a single gene}

\item{data}{the read data as DataFrame for a single gene}
}
\description{
\code{mod_m7G}
}
