% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmod-analysis-default.R
\docType{methods}
\name{convertReadsToPositions}
\alias{convertReadsToPositions}
\alias{convertReadsToPositions,analysis_default,character,GRanges,ScanBamParam-method}
\title{convertReadsToPositions}
\usage{
convertReadsToPositions(object, files, gff, param)


  \S4method{convertReadsToPositions}{analysis_default,character,GRanges,ScanBamParam}(object,
  files, gff, param)
}
\arguments{
\item{object}{analysis object}

\item{files}{files which to analyze}

\item{gff}{a GRanges object for the genome}

\item{param}{ScanBamParam to use for loading the bam files}
}
\description{
\code{analysis_default}: calls the default method
}
