% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmodR-analysis-5end.R
\docType{methods}
\name{convertReadsToPositions}
\alias{convertReadsToPositions}
\alias{convertReadsToPositions,analysis_5end,character,character,GRanges,ScanBamParam-method}
\title{convertReadsToPositions}
\usage{
convertReadsToPositions(object, files, conditions, gff, param)


  \S4method{convertReadsToPositions}{analysis_5end,character,character,GRanges,ScanBamParam}(object,
  files, conditions, gff, param)
}
\arguments{
\item{object}{analysis object}

\item{files}{files which to analyze}

\item{gff}{a GRanges object for the genome}

\item{param}{ScanBamParam to use for loading the bam files}
}
\value{
a modifified \code{analysis} class object containing now the read 
data required for modification detection.
}
\description{
\code{analysis_5end}: load the input files, eg. bam files and 
pre-processes the information to be used in the modification detection.
}
\examples{
\donttest{
ad <- new("analysis_5end")
convertReadsToPositions(ad,
                       "test.bam",
                       "test.gff",
                       test_param)
}
}
