% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/class-RNAmod-D.R,
%   R/class-RNAmod-default.R, R/class-RNAmod-m3C.R, R/class-RNAmod-m7G.R
\docType{methods}
\name{convertReadsToPositions}
\alias{convertReadsToPositions}
\alias{convertReadsToPositions,mod_D,numeric,GRanges,DataFrame-method}
\alias{convertReadsToPositions,mod_default,numeric,GRanges,DataFrame-method}
\alias{convertReadsToPositions,mod_m3C,numeric,GRanges,DataFrame-method}
\alias{convertReadsToPositions,mod_m7G,numeric,GRanges,DataFrame-method}
\title{convertReadsToPositions}
\usage{
convertReadsToPositions(object, counts, gff, data)

\S4method{convertReadsToPositions}{mod_D,numeric,GRanges,DataFrame}(object,
  counts, gff, data)


  \S4method{convertReadsToPositions}{mod_default,numeric,GRanges,DataFrame}(object,
  counts, gff, data)

\S4method{convertReadsToPositions}{mod_m3C,numeric,GRanges,DataFrame}(object,
  counts, gff, data)

\S4method{convertReadsToPositions}{mod_m7G,numeric,GRanges,DataFrame}(object,
  counts, gff, data)
}
\arguments{
\item{object}{mod object}

\item{counts}{total counts in the BAM file containing the data}

\item{gff}{a GRanges object for a single gene}

\item{data}{the read data as DataFrame for a single gene}
}
\description{
\code{mod_D}: calls the default method

\code{mod_m7G}: calls the default method

\code{mod_m3C}: calls the default method

\code{mod_m7G}: calls the default method
}
