% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-coverage.R
\docType{class}
\name{CoverageSequenceData-class}
\alias{CoverageSequenceData-class}
\alias{CoverageSequenceData}
\alias{getData,CoverageSequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{aggregate,CoverageSequenceData-method}
\alias{getDataTrack,CoverageSequenceData-method}
\title{CoverageSequenceData}
\usage{
CoverageSequenceData(bamfiles, annotation, sequences, seqinfo, ...)


  \S4method{getData}{CoverageSequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)

\S4method{aggregate}{CoverageSequenceData}(x, condition = c("Both",
  "Treated", "Control"))

\S4method{getDataTrack}{CoverageSequenceData}(x, name, ...)
}
\arguments{
\item{bamfiles, annotation, seqinfo, grl, sequences, param, args, ...}{See 
\code{\link[=SequenceData-class]{SequenceData}}}

\item{x}{a \code{CoverageSequenceData}}

\item{condition}{For \code{\link{aggregate}}: condition for which the data 
should be aggregated.}

\item{name}{For \code{getDataTrack}: a valid transcript name. Must be a name
of \code{ranges(x)}}
}
\value{
a \code{CoverageSequenceData} object
}
\description{
\code{CoverageSequenceData} implements
\code{\link[=SequenceData-class]{SequenceData}} to contain and aggregate the
coverage of reads per position along the transcripts.

\code{aggregate} calculates the mean and sd for samples in the \code{control}
and \code{treated} condition separatly.
}
\examples{
# Construct a CoverageSequenceData object
RNAmodR.files <- RNAmodR.Data::RNAmodR.files
annotation <- rtracklayer::GFF3File(RNAmodR.files[["example1.gff3"]])
sequences <- Rsamtools::FaFile(RNAmodR.files[["example1.fasta"]])
files <- c(control = RNAmodR.files[["example_wt_1.bam"]],
           treated = RNAmodR.files[["example_wt_2.bam"]])
csd <- CoverageSequenceData(files, annotation = annotation,
                            sequences = sequences)
# aggregate data
aggregate(csd)
}
