% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-coverage.R
\docType{class}
\name{CoverageSequenceData}
\alias{CoverageSequenceData}
\alias{CoverageSequenceData-class}
\alias{aggregate,CoverageSequenceData-method}
\title{CoverageSequenceData}
\usage{
CoverageSequenceData(bamfiles, annotation, sequences, seqinfo, ...)

\S4method{aggregate}{CoverageSequenceData}(x, condition = c("Both",
  "Treated", "Control"))
}
\description{
\code{CoverageSequenceData} implements \code{\link{SequenceData}} to contain
and aggregate the coverage of reads per position along the transcripts.

\code{aggregate} calculates the mean and sd for samples in the \code{control}
and \code{treated} condition serparatly.
}
\examples{
# Construct a CoverageSequenceData object
annotation <- system.file("extdata","example1.gff3",package = "RNAmodR.Data")
sequences <- system.file("extdata","example1.fasta",package = "RNAmodR.Data")
files <- c(control = system.file("extdata","example_wt_1.bam",
                                 package = "RNAmodR.Data"),
           treated = system.file("extdata","example_wt_2.bam",
                                 package = "RNAmodR.Data"))
csd <- CoverageSequenceData(files, annotation = annotation,
                            sequences = sequences)
# aggregate data
aggregate(csd)
}
