% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-end-pos.R
\docType{class}
\name{EndSequenceData}
\alias{EndSequenceData}
\alias{End5SequenceData}
\alias{End3SequenceData}
\alias{End5SequenceData-class}
\alias{End3SequenceData-class}
\alias{EndSequenceData-class}
\alias{aggregate,End5SequenceData-method}
\alias{aggregate,End3SequenceData-method}
\alias{aggregate,EndSequenceData-method}
\alias{getDataTrack,EndSequenceData-method}
\alias{getDataTrack,End5SequenceData-method}
\alias{getDataTrack,End3SequenceData-method}
\title{End5SequenceData/End3SequenceData/EndSequenceData}
\usage{
End5SequenceData(bamfiles, annotation, sequences, seqinfo, ...)

End3SequenceData(bamfiles, annotation, sequences, seqinfo, ...)

EndSequenceData(bamfiles, annotation, sequences, seqinfo, ...)

\S4method{aggregate}{End5SequenceData}(x, condition = c("Both",
  "Treated", "Control"))

\S4method{aggregate}{End3SequenceData}(x, condition = c("Both",
  "Treated", "Control"))

\S4method{aggregate}{EndSequenceData}(x, condition = c("Both", "Treated",
  "Control"))

\S4method{getDataTrack}{EndSequenceData}(x, name, ...)

\S4method{getDataTrack}{End5SequenceData}(x, name, ...)

\S4method{getDataTrack}{End3SequenceData}(x, name, ...)
}
\arguments{
\item{bamfiles, annotation, sequences, seqinfo, ...}{See 
\code{\link[=SequenceData-class]{SequenceData}}}

\item{x}{a \code{CoverageSequenceData}}

\item{condition}{For \code{\link{aggregate}}: condition for which the data 
should be aggregated.}

\item{name}{For \code{getDataTrack}: a valid transcript name. Must be a name
of \code{ranges(x)}}
}
\value{
a \code{End5SequenceData}, a \code{End3SequenceData} or a 
\code{EndSequenceData} object
}
\description{
The \code{End5SequenceData}/\code{End3SequenceData}/\code{EndSequenceData}
aggregate the counts of read ends at each position along transcript. Whereas
the first aggregate either the 5'-end or 3'-end, the \code{EndSequenceData}
aggregates both.

\code{aggregate} calculates the mean and sd for samples in the \code{control}
and \code{treated} condition separatly.
}
\examples{
# Construct a End5SequenceData object
RNAmodR.files <- RNAmodR.Data::RNAmodR.files
annotation <- rtracklayer::GFF3File(RNAmodR.files[["example1.gff3"]])
sequences <- Rsamtools::FaFile(RNAmodR.files[["example1.fasta"]])
files <- c(control = RNAmodR.files[["example_wt_1.bam"]],
           treated = RNAmodR.files[["example_wt_2.bam"]])
e5sd <- End5SequenceData(files, annotation = annotation,
                        sequences = sequences)
# aggregate data
aggregate(e5sd)
}
