% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-end-pos.R,
%   R/SequenceData-protected-end-pos.R
\docType{class}
\name{EndSequenceData}
\alias{EndSequenceData}
\alias{End5SequenceData-class}
\alias{End3SequenceData-class}
\alias{EndSequenceData-class}
\alias{End5SequenceData}
\alias{End3SequenceData}
\title{EndSequenceData}
\usage{
End5SequenceData(bamfiles, fasta, gff, ...)

End3SequenceData(bamfiles, fasta, gff, ...)

EndSequenceData(bamfiles, fasta, gff, ...)

\S4method{aggregate}{EndSequenceData}(x, condition = c("Both", "Treated",
  "Control"))

\S4method{aggregate}{End5SequenceData}(x, condition = c("Both",
  "Treated", "Control"))

\S4method{aggregate}{End3SequenceData}(x, condition = c("Both",
  "Treated", "Control"))

\S4method{aggregate}{ProtectedEndSequenceData}(x, condition = c("Both",
  "Treated", "Control"))
}
\description{
title
}
