% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{FilterTxLoc}
\alias{FilterTxLoc}
\title{Filter sections of a \code{txLoc} object.}
\usage{
FilterTxLoc(locus, filter = NULL)
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{filter}{A character vector; only keep transcript sections
specified in \code{filter}; if \code{NULL} consider all sections.}
}
\value{
A \code{txLoc} object.
}
\description{
Filter sections of a \code{txLoc} object.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}

