% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GenerateNull}
\alias{GenerateNull}
\title{Generate a list of null sites.}
\usage{
GenerateNull(txLoc, id = NULL, method = c("ntAbund", "perm"),
  nt = "C", showPb = TRUE)
}
\arguments{
\item{txLoc}{A \code{txLoc} object.}

\item{id}{A character string; identifier for null sites; if \code{NULL} then
\code{id = "null"}; default is \code{NULL}.}

\item{method}{A character string; specifies the method used to the generate
null distribution; if \code{method == "ntAbund"} the position of all
nucleotides specified by \code{nt} will be used as null sites; if
\code{method == "perm"} (DEPRECATED, see 'Details') then null sites will be
generated by uniform-randomly shuffling candidatepositions from \code{locus}
within the corresponding transcript region; default is \code{"ntAbund"}.}

\item{nt}{A single character; if \code{method == "ntAbund"}, use \code{nt}
to derive distribution of null sites.}

\item{showPb}{A logical scalar; if \code{TRUE} show a progress
bar; default is \code{TRUE}.}
}
\value{
A \code{txLoc} object. See 'Details'.
}
\description{
Generate a list of null sites. See 'Details'.
}
\details{
The function generates a null distribution of single-nucleotide
sites across different transcript sections, and returns a
\code{txLoc} object.
Two different methods can be employed:
\enumerate{
\item \code{method = "ntAbund"}: Null sites are generated based on the
position of all non-modified nucleotides of type \code{nt} in those
transcript sections that also contain a modified site of the same type and
as specified in \code{locus}.
For example, if \code{locus} contains a list of m$^{6}$A sites, the list of
null sites consists of all non-methylated adenosines in those transcripts
that contain at least one m$^{6}$A site.
\item \code{method = "perm"}: Null sites are generated by randomly permuting
the position of sites from \code{locus} uniformly within the corresponding
transcript section. Note that this will generate a list of null sites with
the same abundance ratios across  transcript sections as the list of sites
from \code{locus}. It is therefore not useful for assessing an enrichment of
sites within a particular transcript section. In fact, this method should
not be used and is included purely for paedagogical purposes (to demonstrate
the importance of a sensible null distribution). It is likely that this
method will be removed from future RNAModR versions.
}
It is import to emphasise that any downstream enrichment analysis may depend
critically on the choice of the null distribution. For example, a position
permution-based null distribution may not be a valid null distribution, if
the distribution of nucleotides is highly non-uniform across a transcript
section. This is the case e.g. for the spatial distribution of cytosines
within and across the 5'UTR, CDS and/or 3'UTR. In this case, a permutation-
based distribution of cytosines will not give a sensible null distribution.
Instead, a sensible null distribution can be derived from the position of
all cytosines in the relevant transcript region containing the methylated
cytosine site in \code{locus}. \code{method = "ntAbund"} generates a list of
null sites using this approach.
}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR")
sites <- ReadBED(bedFile)
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38")
negSites <- GenerateNull(posSites, method = "ntAbund", nt = "A")
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
