% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GenerateNull}
\alias{GenerateNull}
\title{Generate a list of null sites.}
\usage{
GenerateNull(txLoc, id = NULL, method = c("ntAbund", "perm"),
  nt = "C", showPb = TRUE)
}
\arguments{
\item{txLoc}{A \code{txLoc} object.}

\item{id}{A character string; identifier for null sites; if \code{NULL} then 
\code{id = "null"}; default is \code{NULL}.}

\item{method}{A character string; specifies the method used to the generate 
null distribution; if \code{method == "ntAbund"} the position of all 
nucleotides specified by \code{nt} will be used as null sites; if 
\code{method == "perm"} (DEPRECATED, see 'Details') then null sites will be 
generated by uniform-randomly shuffling candidatepositions from \code{locus} 
within the corresponding transcript region; default is \code{"ntAbund"}.}

\item{nt}{A single character; if \code{method == "ntAbund"}, use \code{nt} 
to derive distribution of null sites.}

\item{showPb}{A logical scalar; if \code{TRUE} show a progress
bar; default is \code{TRUE}.}
}
\value{
A \code{txLoc} object. See 'Details'.
}
\description{
Generate a list of null sites. See 'Details'.
}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR")
sites <- ReadBED(bedFile)
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38")
negSites <- GenerateNull(posSites, method = "ntAbund", nt = "A")
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
