% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetEEJunct}
\alias{GetEEJunct}
\title{Get exon-exon boundary sites from reference transcriptome}
\usage{
GetEEJunct(refGenome = "hg38", filter = c("CDS", "5'UTR"))
}
\arguments{
\item{refGenome}{A character string; specifies a specific
reference genome assembly version based on which the matching
transcriptome is loaded; default is \code{"hg38"}.}

\item{filter}{A character vector; only consider transcript
sections specified in \code{filter}; default is
\code{c("CDS", "5'UTR")}.}
}
\value{
A \code{GRanges} object. See 'Details'.
}
\description{
Get exon-exon boundary sites from transcriptome. See 'Details'.
}
\details{
The function extracts exon-exon boundary sites (EEBS) from a
reference trancriptome specified by \code{refGenome}, and returns
a \code{GRanges} object. Boundary sites are defined as the
location of those exonic single nucleotides that are closest
upstream to the intronic donor site.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
