% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetEEJunct}
\alias{GetEEJunct}
\title{Get exon-exon junctions from transcriptome.}
\usage{
GetEEJunct(refGenome = "hg38", filter = "CDS")
}
\arguments{
\item{refGenome}{A character string; specifies a specific
reference genome assembly version based on which the matching
transcriptome is loaded; default is \code{"hg38"}.}
}
\value{
A \code{txLoc} object. See 'Details'.
}
\description{
Get exon-exon junctions from transcriptome. See 'Details'.
}
\details{
This function extracts exon-exon junction positions from a
reference transcriptome, and returns a \code{txLoc} object.
}

