% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetEEJunct}
\alias{GetEEJunct}
\title{Get exon-exon junctions from transcriptome.}
\usage{
GetEEJunct(refGenome = "hg38", filter = "CDS")
}
\arguments{
\item{refGenome}{A character string; specifies a specific
reference genome assembly version based on which the matching
transcriptome is loaded; default is \code{"hg38"}.}

\item{filter}{A character vector; only consider transcript
sections specified in \code{filter}; \code{filter = NULL}
corresponds to \code{filter = c("5'UTR", "CDS", "3'UTR")}.}
}
\value{
A \code{txLoc} object. See 'Details'.
}
\description{
Get exon-exon junctions from transcriptome. See 'Details'.
}
\details{
The function extracts exon-exon junction positions from a
reference transcriptome, and returns a \code{txLoc} object.
Here an exon-exon junction corresponds to the position of the
3'-last nucleotide of the preceding exon.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

