% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTxMethods.R
\name{GetGeneIds}
\alias{GetGeneIds}
\title{Get different gene IDs.}
\usage{
GetGeneIds(txdb)
}
\arguments{
\item{txdb}{A \code{TxDb} object.}
}
\value{
A \code{data.frame}. See 'Details'.
}
\description{
Get different gene IDs for mapping between different gene
annotations (Ensembl, Entrez, etc.). See 'Details'. 
This is a low-level function that is being called from
\code{BuildTx}.
}
\details{
The function extracts the genome assembly version from
the \code{TxDb} object, and loads the suitable genome wide
annotation package. For example, if \code{TxDb} is based
on \code{"hg38"}, the function loads \code{org.Hs.eg.db}.
Various IDs (Ensembl, Unigene, gene symbols) are extracted
from the annotation package by mapping RefSeq IDs from
\code{TxDb} via \code{mapIds} from the \code{AnnotationDbi}
package. Please note that column entries are not necessarily
unique: For example, two different Ensembl IDs might have the
same Entrez ID. The return object is a \code{data.frame} with
the following columns:
\itemize{
  \item REFSEQ: RefSeq Accession number; these entries are unique
  and refer to the mRNA transcript
  \item ENTREZID: Entrez gene ID
  \item SYMBOL: Gene symbol
  \item ENSEMBL: Ensembl gene ID
  \item UNIGENE: UniGene gene ID
  \item GENENAME: Gene name
}
}

