% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{GetLoci}
\alias{GetLoci}
\title{Method \code{GetLoci} for S4 object \code{txLoc}.}
\usage{
GetLoci(object)
}
\arguments{
\item{object}{A \code{txLoc} object.}
}
\value{
A \code{list} of \code{DataFrame} objects with loci for every
transcript region.
}
\description{
Get a \code{list} of \code{DataFrame} objects of the loci of a \code{txLoc} 
object.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
