% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmod-plot.R
\docType{methods}
\name{getModPlot}
\alias{getModPlot}
\alias{saveModPlot}
\alias{getModPlot,RNAmod,SummarizedExperiment,character,character-method}
\alias{getModPlot}
\alias{saveModPlots}
\alias{saveModPlot,RNAmod,SummarizedExperiment,character,character-method}
\title{getModPlot}
\usage{
getModPlot(.Object, se, modifications, gene, focus = FALSE)

saveModPlot(.Object, se, modifications, genes, focus = FALSE,
  filetype = "pdf")

\S4method{getModPlot}{RNAmod,SummarizedExperiment,character,character}(.Object,
  se, modifications, gene, focus = FALSE)


  \S4method{saveModPlot}{RNAmod,SummarizedExperiment,character,character}(.Object,
  se, modifications, genes, focus = FALSE, filetype = "pdf")
}
\arguments{
\item{.Object}{a RNAmod object.}

\item{se}{a SummarizedExperiment containg the experimental data.}

\item{modifications}{name of modification to be used for analysis.}

\item{gene}{a single gene name}
}
\value{
a grid of layer plot and any mod plot

TRUE if successful
}
\examples{
\donttest{
getModPlot(mod,1,"m7G","RDN18-1")
saveModPlot(mod,1,"m7G",c("RDN18-1","tS\\(CGA\\)P"))
}
}
