% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetMotifLoc}
\alias{GetMotifLoc}
\title{Get loci of motif(s).}
\usage{
GetMotifLoc(motif = NULL, refGenome = "hg38", id = NULL, maxMM = 0,
  showPb = TRUE)
}
\arguments{
\item{motif}{A \code{character} vector; specifies the query sequences that
will be matched against the reference transcriptome.}

\item{refGenome}{A character string; specifies the reference genome version;
default is \code{"hg38"}.}

\item{id}{A character string; specifies an identifier for \code{motif};
default is \code{NULL}}

\item{maxMM}{An integer scalar; specifies the maximum number of mismatches
that are allowed during the motif matching; default is 0.}

\item{showPb}{A logical scalar; if \code{TRUE} show a progress bar; default
is \code{TRUE}.}
}
\value{
A \code{txLoc} object.
}
\description{
Get loci of motif(s) from transcriptome. See 'Details'.
}
\details{
Based on a set of query sequences \code{motif}, extract the loci of matching
sequences within different regions of a reference transcriptome
\code{refGenome}. The function returns a \code{txLoc} object including
transcriptomic and genomic loci of the query motif(s). The maximum number
of mismatches allowed in the motif search can be adjusted through
\code{maxMM}.
Note that the motif search may take a few minutes, depending on the number
of motif(s), hardware, and size of the transcriptome.
}
\examples{
\dontrun{
PAS <- GetMotifLoc(id = "PAS")
PAS <- FilterTxLoc(PAS, c("3'UTR", "CDS", "5'UTR"))
PlotSpatialDistribution(PAS, absolute = FALSE)
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
