% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetMotifLoc}
\alias{GetMotifLoc}
\title{Get loci of motif(s) from transcriptome.}
\usage{
GetMotifLoc(motif, refGenome = "hg38", maxMM = 0, filter = NULL,
  showPb = TRUE)
}
\arguments{
\item{motif}{A character vector; specifies the motif(s)
that will be matched against the transcriptome.}

\item{refGenome}{A character string; specifies the reference
genome version; default is \code{"hg38"}.}

\item{maxMM}{An integer scalar; specifies the maximum number
of mismatches that are allowed during the motif matching;
default is 0.}

\item{filter}{A character vector; only consider transcript
sections specified in \code{filter}; \code{filter = NULL}
corresponds to \code{filter = c("5'UTR", "CDS", "3'UTR")}.}

\item{showPb}{A logical scalar; if \code{TRUE} show a progress
bar; default is \code{TRUE}.}
}
\description{
Get loci of motif(s) from transcriptome.
}
\details{
This function returns a \code{txLoc} object of loci of
the specified motif(s) within transcript sections. Motif
sequence matching allows for mismatches through the
parameter \code{maxMM}.
}

