% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{GetMotifLoc}
\alias{GetMotifLoc}
\title{Get loci of motif(s) from transcriptome.}
\usage{
GetMotifLoc(motif, refGenome = "hg38", filter = NULL, maxMM = 0,
  showPb = TRUE)
}
\arguments{
\item{motif}{A character vector; specifies the motif(s)
that will be matched against the transcriptome.}

\item{refGenome}{A character string; specifies the reference
genome version; default is \code{"hg38"}.}

\item{filter}{A character vector; only consider transcript
sections specified in \code{filter}; \code{filter = NULL}
corresponds to \code{filter = c("5'UTR", "CDS", "3'UTR")}.}

\item{maxMM}{An integer scalar; specifies the maximum number
of mismatches that are allowed during the motif matching;
default is 0.}

\item{showPb}{A logical scalar; if \code{TRUE} show a progress
bar; default is \code{TRUE}.}
}
\description{
Get loci of motif(s) from transcriptome. See 'Details'.
}
\details{
The function searches for one or multiple motifs within
sequences of a reference transcriptome, and returns a
\code{txLoc} object of the motif loci within transcript
the different sections. The maximum number of mismatches 
allowed in the motif search can be adjusted through
\code{maxMM}. By default a text progressbar is shown
\code{showPb = TRUE}.
Note that the motif search may take a few minutes, depending
on the size of the transcriptome and number of motifs.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

