% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{GetNumberOfLoci}
\alias{GetNumberOfLoci}
\title{Generic method "GetNumberOfLoci" for S4 object txLoc.}
\usage{
GetNumberOfLoci(x)
}
\arguments{
\item{x}{A \code{txLoc} object.}
}
\description{
Generic method "GetNumberOfLoci" for S4 object txLoc.
}

