% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{GetNumberOfLoci}
\alias{GetNumberOfLoci}
\alias{GetNumberOfLoci,txLoc-method}
\title{Method "GetNumberOfLoci" for S4 object txLoc.}
\usage{
GetNumberOfLoci(x)

\S4method{GetNumberOfLoci}{txLoc}(x)
}
\arguments{
\item{x}{A \code{txLoc} object.}
}
\value{
A named integer vector with the number of sites per
transcript section.
}
\description{
Get the number of loci of a \code{txLoc} object in every
transcript section.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
