% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{GetRef}
\alias{GetRef}
\alias{GetRef,txLoc-method}
\title{Method "GetRef" for S4 object txLoc.}
\usage{
GetRef(x)

\S4method{GetRef}{txLoc}(x)
}
\arguments{
\item{x}{A \code{txLoc} object.}
}
\value{
A string.
}
\description{
Get the reference genome string of a \code{txLoc} object.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

