% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{GetRef}
\alias{GetRef}
\title{Method \code{GetRef} for S4 object \code{txLoc}.}
\usage{
GetRef(object)
}
\arguments{
\item{object}{A \code{txLoc} object.}
}
\value{
A string.
}
\description{
Get the reference genome string of a \code{txLoc} object.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
