% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{GetRelDistNearest}
\alias{GetRelDistNearest}
\title{Return list of nearest distances between entries from two \code{list}s of
\code{GRanges} objects.}
\usage{
GetRelDistNearest(lst1, lst2)
}
\arguments{
\item{lst1}{A \code{list} of \code{GRanges} object.}

\item{lst2}{A \code{list} of \code{GRanges} object.}
}
\value{
A list of \code{integer} vectors. See 'Details'.
}
\description{
Return list of nearest distances between entries from two \code{list}s of
\code{GRanges} objects. See 'Details'.
}
\details{
The function uses \code{GenomicRanges::distanceToNearest} to return the
nearest distances between the start positions of a \code{GRanges} object 
from \code{lst1} and the corresponding \code{GRanges} object from 
\code{lst2} with the same name (based on field \code{seqnames}).
Note that distances are given as signed \code{integer}s: Negative
distances correspond to pos(gr1) < pos(gr2), positive
distances correspond to pos(gr1) > pos(gr2).
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
