% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmodR.R
\docType{methods}
\name{getSummarizedExperiment}
\alias{getSummarizedExperiment}
\alias{getSummarizedExperiments}
\alias{getSummarizedExperiment,RNAmodR,numeric-method}
\alias{getSummarizedExperiments,RNAmodR,numeric-method}
\title{returns one or more SummarizedExperiment}
\usage{
getSummarizedExperiment(.Object, number)

getSummarizedExperiments(.Object, number)

\S4method{getSummarizedExperiment}{RNAmodR,numeric}(.Object, number)

\S4method{getSummarizedExperiments}{RNAmodR,numeric}(.Object, number)
}
\arguments{
\item{.Object}{an RNAmod object}

\item{number}{a number defining the experiment. For 
getSummarizedExperiments() more than number can be defined as numeric vector.}
}
\value{
\code{getSummarizedExperiment()}: SummarizedExperiment

\code{getSummarizedExperiments()}: list of SummarizedExperiments
}
\description{
Global access to SummarizedExperiments stored by RpfExperiment. 
\code{getSummarizedExperiment()} returns the result of experiment, whereas 
\code{getSummarizedExperiments()}, is the vectorized version returning a
list of experiment results.
}
\section{Methods (by class)}{
\itemize{
\item \code{.Object = RNAmodR,number = numeric}: 
}}

\examples{
\donttest{
getSummarizedExperiment(experiment, 1)
getSummarizedExperiments(experiment, c(1, 2))
}
}
