% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{GetVersion}
\alias{GetVersion}
\alias{GetVersion,txLoc-method}
\title{Method "GetVersion" for S4 object txLoc.}
\usage{
GetVersion(x)

\S4method{GetVersion}{txLoc}(x)
}
\arguments{
\item{x}{A \code{txLoc} object.}
}
\value{
A string.
}
\description{
Get the version string of a \code{txLoc} object.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}

