% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/methods-RNAmodR-compare.R
\docType{methods}
\name{heatmapModifications}
\alias{heatmapModifications}
\alias{saveHeatmapModifications}
\alias{heatmapModifications,RNAmodR,list,character-method}
\alias{heatmapModifications}
\alias{saveHeatmapModifications}
\alias{saveHeatmapModifications,RNAmodR,list,character-method}
\title{Comparative visualization of modifications}
\usage{
heatmapModifications(.Object, ses, modifications)

saveHeatmapModifications(.Object, ses, modifications)

\S4method{heatmapModifications}{RNAmodR,list,character}(.Object, ses,
  modifications)

\S4method{saveHeatmapModifications}{RNAmodR,list,character}(.Object, ses,
  modifications)
}
\arguments{
\item{.Object}{a RNAmod object.}

\item{modifications}{name of modification to be used for analysis.}

\item{se}{a SummarizedExperiment containg the experimental data.}

\item{gene}{a single gene name}
}
\value{
a heatmap plot for given experiments and modifications positions
}
\examples{
\donttest{
}
}
