% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/methods-RNAmodR-compare.R
\docType{methods}
\name{heatmapModifications}
\alias{heatmapModifications}
\alias{saveHeatmapModifications}
\alias{heatmapModifications,RNAmodR,SummarizedExperiment,character,character-method}
\alias{heatmapModifications}
\alias{saveHeatmapModifications}
\alias{saveHeatmapModifications,RNAmodR,SummarizedExperiment,character,character-method}
\title{Comparative visualization of modifications}
\usage{
heatmapModifications(.Object, se, modifications, genes)

saveHeatmapModifications(.Object, se, modifications, genes)


  \S4method{heatmapModifications}{RNAmodR,SummarizedExperiment,character,character}(.Object,
  se, modifications, genes)


  \S4method{saveHeatmapModifications}{RNAmodR,SummarizedExperiment,character,character}(.Object,
  se, modifications, genes)
}
\arguments{
\item{.Object}{a RNAmod object.}

\item{se}{a SummarizedExperiment containg the experimental data.}

\item{modifications}{name of modification to be used for analysis.}

\item{gene}{a single gene name}
}
\value{
a heatmap plot for given experiments and modifications positions
}
\examples{
\donttest{
}
}
