% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/methods-RNAmodR-compare.R
\docType{methods}
\name{heatmapModifications}
\alias{heatmapModifications}
\alias{saveHeatmapModifications}
\alias{heatmapModifications,RNAmodR,numeric,character,character,character-method}
\alias{heatmapModifications}
\alias{saveHeatmapModifications}
\alias{heatmapModifications,GRangesList,missing,missing,character,character-method}
\alias{heatmapModifications,list,missing,missing,character,character-method}
\alias{saveHeatmapModifications,RNAmodR,numeric,character,character,character,missing-method}
\alias{saveHeatmapModifications,GRangesList,missing,missing,character,character,character-method}
\alias{saveHeatmapModifications,list,missing,missing,character,character,character-method}
\title{Comparative visualization of modifications}
\usage{
heatmapModifications(x, numbers, sampleNames, modifications, genes)

saveHeatmapModifications(x, numbers, sampleNames, modifications, genes, folder)


  \S4method{heatmapModifications}{RNAmodR,numeric,character,character,character}(x,
  numbers, sampleNames, modifications, genes)


  \S4method{heatmapModifications}{GRangesList,missing,missing,character,character}(x,
  modifications, genes)

\S4method{heatmapModifications}{list,missing,missing,character,character}(x,
  modifications, genes)


  \S4method{saveHeatmapModifications}{RNAmodR,numeric,character,character,character,missing}(x,
  numbers, sampleNames, modifications, genes)


  \S4method{saveHeatmapModifications}{GRangesList,missing,missing,character,character,character}(x,
  modifications, genes, folder)


  \S4method{saveHeatmapModifications}{list,missing,missing,character,character,character}(x,
  modifications, genes, folder)
}
\arguments{
\item{x}{a RNAmod object or a named list of SummarizedExperiment objects containg the 
experimental data to be used for comparison or a named GRangesList containg the experimental data.}

\item{numbers}{two or more experiment numbers}

\item{sampleNames}{two or more names. This length of \code{numbers} and 
\code{sampleNames} must match}

\item{modifications}{name of modification to be used for analysis. These
should be defined as the format gene-name_nucleotide_position.}

\item{genes}{gene names to be included in the analysis}
}
\value{
a heatmap plot for given experiments and modifications positions
}
\description{
compares different experiments as a heatmap for selected genes and/or 
modifications
}
\examples{
\donttest{
heatmapModifications(grl,"RDN18-1_G_1575")
}
}
