% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/methods-RNAmodR-compare.R
\docType{methods}
\name{heatmapModifications}
\alias{heatmapModifications}
\alias{saveHeatmapModifications}
\alias{heatmapModifications,list,missing,character-method}
\alias{heatmapModifications}
\alias{saveHeatmapModifications}
\alias{heatmapModifications,missing,GRangesList,character-method}
\alias{saveHeatmapModifications,list,missing,character-method}
\alias{saveHeatmapModifications,missing,GRangesList,character-method}
\title{Comparative visualization of modifications}
\usage{
heatmapModifications(ses, grl, modifications)

saveHeatmapModifications(ses, grl, modifications)

\S4method{heatmapModifications}{list,missing,character}(ses, modifications)

\S4method{heatmapModifications}{missing,GRangesList,character}(grl,
  modifications)

\S4method{saveHeatmapModifications}{list,missing,character}(ses, modifications)

\S4method{saveHeatmapModifications}{missing,GRangesList,character}(grl,
  modifications)
}
\arguments{
\item{modifications}{name of modification to be used for analysis.}

\item{.Object}{a RNAmod object.}

\item{se}{a SummarizedExperiment containg the experimental data.}

\item{gene}{a single gene name}
}
\value{
a heatmap plot for given experiments and modifications positions
}
\examples{
\donttest{
}
}
