% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{LoadRefTx}
\alias{LoadRefTx}
\title{Load reference transcriptome.}
\usage{
LoadRefTx(refGenome = "hg38", env = .GlobalEnv)
}
\arguments{
\item{refGenome}{A character string; specifies a specific
reference genome assembly version based on which a transcriptome
is loaded; default is \code{"hg38"}.}

\item{env}{An \code{environment} object; default is the user's
workspace, i.e. \code{env = .GlobalEnv}.}
}
\description{
Load reference transcriptome. See 'Details'.
}
\details{
The function loads transcriptome data stored in a \code{.RData}
file, and makes the objects assessible in the user's workspace.
}
\keyword{internal}

