% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmod-m7G.R
\docType{methods}
\name{mergePositionsOfReplicates}
\alias{mergePositionsOfReplicates}
\alias{mergePositionsOfReplicates,mod_m7G,GRanges,FaFile,list-method}
\title{mergePositionsOfReplicates}
\usage{
mergePositionsOfReplicates(object, gff, seq, data)

\S4method{mergePositionsOfReplicates}{mod_m7G,GRanges,FaFile,list}(object, gff,
  seq, data)
}
\arguments{
\item{object}{mod object}

\item{gff}{a GRanges object for a single gene}

\item{seq}{a DNAString object for a single gene}

\item{data}{the read data as DataFrame for a single gene}
}
\description{
\code{mod_m7G}
}
