% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmodR-analysis-default.R
\docType{methods}
\name{mergePositionsOfReplicates}
\alias{mergePositionsOfReplicates}
\alias{mergePositionsOfReplicates,analysis_default-method}
\title{mergePositionsOfReplicates}
\usage{
mergePositionsOfReplicates(object)

\S4method{mergePositionsOfReplicates}{analysis_default}(object)
}
\arguments{
\item{object}{a mod object}
}
\value{
a modified \code{analysis} class object now containing the aggregated
pre-processed data of the input. This data is used for saving.
}
\description{
\code{analysis_default}
}
\examples{
\donttest{
load(system.file("data", file = "test_example.RData", package = "RNAmodR"))
mergePositionsOfReplicates(ad)
}
}
