% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmod-analysis-default.R
\docType{methods}
\name{mergePositionsOfReplicates}
\alias{mergePositionsOfReplicates}
\alias{mergePositionsOfReplicates,analysis_default-method}
\title{mergePositionsOfReplicates}
\usage{
mergePositionsOfReplicates(object)

\S4method{mergePositionsOfReplicates}{analysis_default}(object)
}
\arguments{
\item{object}{mod object}
}
\description{
\code{analysis_default}
}
