% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmodR-mod-type.R
\docType{methods}
\name{mod-accessors}
\alias{mod-accessors}
\alias{getDataType}
\alias{getModType}
\alias{getDataType}
\alias{getPositionOffset}
\alias{getModType,mod-method}
\alias{getDataType}
\alias{getModType}
\alias{getDataType,mod-method}
\alias{getPositionOffset,mod-method}
\title{accessors for \code{mod} class objects}
\usage{
getModType(object)

getDataType(object)

getPositionOffset(object)

\S4method{getModType}{mod}(object)

\S4method{getDataType}{mod}(object)

\S4method{getPositionOffset}{mod}(object)
}
\arguments{
\item{object}{a RNAmodR mod class object}
}
\value{
character defining the modification type for this modification
class

character defining the plot type for this modification class

a position offset for scoring a modified positions, e.g. +1 for
a 5'-read stop at tlast unmodified position
}
\description{
title
}
\examples{
\donttest{
getDataType(mod)
getModType(mod)
}
}
