% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmod-analysis-type.R, R/class-RNAmod-mod-type.R
\docType{methods}
\name{mod-accessors}
\alias{mod-accessors}
\alias{getPositions}
\alias{getModifications}
\alias{getNumberOfModifications}
\alias{getModType}
\alias{getAnalysisType}
\alias{checkForModification}
\alias{getPlotType,analysis-method}
\alias{getPositions,analysis-method}
\alias{getModifications,analysis-method}
\alias{getNumberOfModifications,analysis-method}
\alias{getAnalysisType,mod-method}
\alias{getModType,mod-method}
\title{Accessor for 'analysis' class}
\usage{
getPlotType(object)

getPositions(object)

getModifications(object)

getNumberOfModifications(object)

getModType(object)

getAnalysisType(object)

checkForModification(object, data, locations)

\S4method{getPlotType}{analysis}(object)

\S4method{getPositions}{analysis}(object)

\S4method{getModifications}{analysis}(object)

\S4method{getNumberOfModifications}{analysis}(object)

\S4method{getAnalysisType}{mod}(object)

\S4method{getModType}{mod}(object)
}
\arguments{
\item{object}{analysis object}
}
\value{
character defining the plot type for this analysis class

a list of data as a lists of list(replicate) of DataFrame(transcript)

a list of data as a lists of list(replicate) of DataFrame(transcript)

returns the total number of modifications detected independent of the
type of modification

character defining the plot type for this modification class

character defining the modification type for this modification
class
}
\examples{
\donttest{
getPlotType(analysis)
}
\donttest{
getPositions(analysis)
}
\donttest{
getModifications(analysis)
}
\donttest{
getNumberOfModifications(analysis)
}
\donttest{
getAnalysisType(mod)
getModType(mod)
}
}
