% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmodR-mod-type.R
\docType{methods}
\name{mod-accessors}
\alias{mod-accessors}
\alias{getAnalysisType}
\alias{getModType}
\alias{getAnalysisType}
\alias{getAnalysisType,mod-method}
\alias{getAnalysisType}
\alias{getModType}
\alias{getModType,mod-method}
\title{accessors for \code{mod} class objects}
\usage{
getModType(object)

getAnalysisType(object)

\S4method{getAnalysisType}{mod}(object)

\S4method{getModType}{mod}(object)
}
\arguments{
\item{object}{a RNAmodR mod class object}
}
\value{
character defining the plot type for this modification class

character defining the modification type for this modification
class
}
\description{
title
}
\examples{
\donttest{
getAnalysisType(mod)
getModType(mod)
}
}
