% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RNAmod-analysis-type.R,
%   R/class-RNAmod-mod-type.R, R/class-RNAmod-analysis-default.R,
%   R/class-RNAmod-mod-D.R, R/class-RNAmod-mod-m3C.R, R/class-RNAmod-mod-m7G.R
\docType{class}
\name{mod}
\alias{mod}
\alias{mod}
\alias{analysis_default-class}
\alias{mod_D-class}
\alias{mod_m3C-class}
\alias{mod_m7G-class}
\title{mod}
\description{
\code{analysis_default}

\code{mod_D}

\code{mod_m3C}

\code{mod_m7G}
}
