% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RNAmodR-mod-type.R,
%   R/class-RNAmodR-mod-D.R, R/class-RNAmodR-mod-m3C.R,
%   R/class-RNAmodR-mod-m7G.R
\docType{class}
\name{mod}
\alias{mod}
\alias{mod_D-class}
\alias{mod_m3C-class}
\alias{mod_m7G-class}
\title{mod}
\description{
Virtual class for modification detection.

\code{mod_D}

\code{mod_m3C}

\code{mod_m7G}
}
