% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModExperiment-class.R
\docType{class}
\name{ModExperiment}
\alias{ModExperiment}
\alias{ModExperiment-class}
\alias{show,ModExperiment-method}
\title{ModExperiment}
\usage{
\S4method{show}{ModExperiment}(object)
}
\description{
title
}
