% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Modifier-class.R,
%   R/ModifierSet-class.R
\docType{methods}
\name{modifierType}
\alias{modifierType}
\alias{modType}
\alias{mainScore}
\alias{settings}
\alias{settings<-}
\alias{seqinfo}
\alias{sequences}
\alias{bamfiles}
\alias{seqData}
\alias{Modifier-functions}
\alias{show,Modifier-method}
\alias{modifierType,Modifier-method}
\alias{modType,Modifier-method}
\alias{mainScore,Modifier-method}
\alias{settings,Modifier-method}
\alias{settings<-,Modifier-method}
\alias{seqData,Modifier-method}
\alias{names,Modifier-method}
\alias{seqinfo,Modifier-method}
\alias{sequences,Modifier-method}
\alias{ranges,Modifier-method}
\alias{bamfiles,Modifier-method}
\alias{show,ModifierSet-method}
\alias{modifierType,ModifierSet-method}
\alias{modType,ModifierSet-method}
\alias{mainScore,ModifierSet-method}
\alias{settings,ModifierSet-method}
\alias{settings<-,ModifierSet-method}
\alias{bamfiles,ModifierSet-method}
\alias{sequences,ModifierSet-method}
\alias{ranges,ModifierSet-method}
\alias{seqinfo,ModifierSet-method}
\alias{modifications,ModifierSet-method}
\title{Modifier/ModifierSet functions}
\usage{
modifierType(x)

modType(x)

mainScore(x)

settings(x, name = NULL)

settings(x, name) <- value

seqinfo(x)

sequences(x, ...)

bamfiles(x)

seqData(x)

\S4method{show}{Modifier}(object)

\S4method{modifierType}{Modifier}(x)

\S4method{modType}{Modifier}(x)

\S4method{mainScore}{Modifier}(x)

\S4method{settings}{Modifier}(x, name = NULL)

\S4method{settings}{Modifier}(x) <- value

\S4method{seqData}{Modifier}(x)

\S4method{names}{Modifier}(x)

\S4method{seqinfo}{Modifier}(x)

\S4method{sequences}{Modifier}(x, modified = FALSE)

\S4method{ranges}{Modifier}(x)

\S4method{bamfiles}{Modifier}(x)

\S4method{show}{ModifierSet}(object)

\S4method{modifierType}{ModifierSet}(x)

\S4method{modType}{ModifierSet}(x)

\S4method{mainScore}{ModifierSet}(x)

\S4method{settings}{ModifierSet}(x, name = NULL)

\S4method{settings}{ModifierSet}(x) <- value

\S4method{bamfiles}{ModifierSet}(x)

\S4method{sequences}{ModifierSet}(x, modified = FALSE)

\S4method{ranges}{ModifierSet}(x)

\S4method{seqinfo}{ModifierSet}(x)

\S4method{modifications}{ModifierSet}(x, perTranscript = FALSE)
}
\arguments{
\item{x, object}{a \code{Modifier} or \code{ModifierSet} class}

\item{name}{For \code{settings}: name of the setting to be returned or set}

\item{value}{For \code{settings}: value of the setting to be set}

\item{..., modified}{For \code{sequences}: \code{TRUE} or \code{FALSE}: Should
the sequences be returned as a \code{ModRNAString} with the found 
modifications added on top of the \code{RNAString}? See 
\code{\link[Modstrings:modifyNucleotides]{modifyNucleotides}}.}
}
\description{
For the \code{Modifier} and  \code{ModifierSet} classes a number of functions
are implemented to access the data stored by the object.
}
