% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-class.R
\docType{class}
\name{Modifier}
\alias{Modifier}
\alias{Modifier-class}
\alias{show,Modifier-method}
\title{Modifier}
\usage{
\S4method{show}{Modifier}(object)

\S4method{gff}{Modifier}(x)

\S4method{fasta}{Modifier}(x)

\S4method{sequences}{Modifier}(x, modified = FALSE,
  with.qualities = FALSE)

\S4method{ranges}{Modifier}(x)

\S4method{bamfiles}{Modifier}(x)

\S4method{seqdata}{Modifier}(x)

\S4method{modifications}{Modifier}(x, perTranscript = FALSE)

\S4method{modify}{Modifier}(x)
}
\description{
title
}
