% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Modifier-class.R
\docType{class}
\name{modifierType}
\alias{modifierType}
\alias{modType}
\alias{mainScore}
\alias{settings}
\alias{settings<-}
\alias{seqinfo}
\alias{sequences}
\alias{bamfiles}
\alias{seqData}
\alias{Modifier}
\alias{modify}
\alias{Modifier-class}
\alias{show,Modifier-method}
\title{Modifier}
\usage{
modifierType(x)

modType(x)

mainScore(x)

settings(x, name = NULL)

settings(x, name) <- value

seqinfo(x)

sequences(x, ...)

bamfiles(x)

seqData(x)

\S4method{show}{Modifier}(object)

\S4method{modifierType}{Modifier}(x)

\S4method{modType}{Modifier}(x)

\S4method{mainScore}{Modifier}(x)

\S4method{settings}{Modifier}(x, name = NULL)

\S4method{settings}{Modifier}(x) <- value

\S4method{seqinfo}{Modifier}(x)

\S4method{sequences}{Modifier}(x, modified = FALSE)

\S4method{ranges}{Modifier}(x)

\S4method{bamfiles}{Modifier}(x)

\S4method{seqData}{Modifier}(x)
}
\description{
title
}
