% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-class.R, R/Modifier-Inosine-class.R,
%   R/Modifier-RiboMethSeq-class.R
\docType{class}
\name{Modifier}
\alias{Modifier}
\alias{modify}
\alias{Modifier-class}
\alias{show,Modifier-method}
\title{Modifier}
\usage{
\S4method{show}{Modifier}(object)

\S4method{modifiertype}{Modifier}(x)

\S4method{gff}{Modifier}(x)

\S4method{fasta}{Modifier}(x)

\S4method{sequences}{Modifier}(x, modified = FALSE,
  with.qualities = FALSE)

\S4method{ranges}{Modifier}(x)

\S4method{bamfiles}{Modifier}(x)

\S4method{seqData}{Modifier}(x)

\S4method{aggregateData}{Modifier}(x)

\S4method{modifications}{Modifier}(x, perTranscript = FALSE)

\S4method{modify}{Modifier}(x, ...)

\S4method{aggregate}{Modifier}(x, ...)

\S4method{hasAggregateData}{Modifier}(x)

\S4method{aggregate}{ModInosine}(x)

\S4method{aggregate}{ModRiboMethSeq}(x, ...)
}
\description{
title
}
