% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifierSet-export.R
\docType{methods}
\name{ModifierSet-export}
\alias{ModifierSet-export}
\alias{SummarizedExperiment,ModifierSet-method}
\title{Exporting data to other classes and files}
\usage{
\S4method{SummarizedExperiment}{ModifierSet}(assays, ...)
}
\arguments{
\item{assays}{a \code{\link[=ModifierSet-class]{ModifierSet}} object}

\item{...}{See \code{\link[rtracklayer:BigWigFile]{export.bw}} or 
optional arguments for \code{SummarizedExperiment}:
\itemize{
\item{modificationsOnly} {\code{TRUE} or \code{FALSE}: Should only data on
positions with a modification found be included in the 
\code{RangedSummarizedExperiment}? (default = 
\code{modificationsOnly = TRUE})}
\item{flanking} {If \code{modificationsOnly = FALSE} how many neighboring
positions should be included in the \code{RangedSummarizedExperiment}?
(default = \code{flanking = 10L})}
\item{allPositions} {\code{TRUE} or \code{FALSE}: Should data on
*all* positions of the transcripts be included in the 
\code{RangedSummarizedExperiment}? This overrides the previous two options.
It is potentially very resource hungry. Be careful. (default = 
\code{allPositions = FALSE})}
\item{sequenceData} {\code{TRUE} or \code{FALSE}: Should data from the 
\code{SequenceData} objects (utilized in the \code{ModifierSet}) instead of 
aggregated data from the \code{ModifierSet} object be included in the 
\code{RangedSummarizedExperiment}? (default = 
\code{sequenceData = FALSE})}
}}

\item{object}{a \code{\link[=Modifier-class]{Modifier}}, a 
\code{\link[=SequenceData-class]{SequenceData}}, 
\code{\link[=SequenceDataSet-class]{SequenceDataSet}} or a
\code{\link[=SequenceDataList-class]{SequenceDataList}} object}

\item{con}{See \code{\link[rtracklayer:BigWigFile]{export.bw}}}
}
\description{
\code{Modifier}, \code{SequenceData}, \code{SequenceDataSet} and 
\code{SequenceDataList} can be exported into several other formats.

\code{export.bw} exports to a bigWig file as implemented by the
\code{rtracklayer} package.

\code{SummarizedExperiment} converts the results in a \code{ModifierSet} into
a \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment}}
object.
}
