% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Modifier-class.R
\docType{methods}
\name{modifications}
\alias{modifications}
\alias{modify}
\alias{modifications,Modifier-method}
\title{modify}
\usage{
modifications(x, ...)

modify(x, force = FALSE)

\S4method{modifications}{Modifier}(x, perTranscript = FALSE)

\S4method{modify}{Modifier}(x)
}
\description{
title
}
