% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Modifier-class.R
\docType{methods}
\name{modify}
\alias{modify}
\alias{modifications,Modifier-method}
\alias{modifications}
\title{Searching for modifications in \code{SequenceData}}
\usage{
modifications(x, ...)

modify(x, force = FALSE)

\S4method{modifications}{Modifier}(x, perTranscript = FALSE)

\S4method{modify}{Modifier}(x)
}
\arguments{
\item{x}{a \code{Modifier} object.}

\item{perTranscript}{\code{TRUE} or \code{FALSE}: Should the coordinates be
returned as local per transcript coordinates?}
}
\value{
\itemize{
\item{\code{modify}: }{the updated \code{Modifier} object.}
\item{\code{modifications}: }{the modifications found as a \code{GRanges}
object.}
}
}
\description{
\code{modify} triggers the search for modifications for a \code{Modifier} 
class. Usually this is done automatically during construction of a 
\code{Modifier} object.

\code{modifications} is the accessor for the found modifications.
}
