% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Modifier-class.R,
%   R/ModifierSet-class.R
\docType{methods}
\name{modify}
\alias{modify}
\alias{modifications,Modifier-method}
\alias{modifications}
\alias{modify,Modifier-method}
\alias{modify,ModifierSet-method}
\title{Searching for modifications in \code{SequenceData}}
\usage{
modifications(x, ...)

modify(x, force = FALSE)

\S4method{modifications}{Modifier}(x, perTranscript = FALSE)

\S4method{modify}{Modifier}(x, force = FALSE)

\S4method{modify}{ModifierSet}(x, force = FALSE)
}
\arguments{
\item{x}{a \code{Modifier} object.}

\item{...}{additional arguments}

\item{force}{force to run \code{aggregate} again, if data is already stored
in \code{x}.}

\item{perTranscript}{For \code{modifications>} \code{TRUE} or \code{FALSE}:
Should the coordinates be returned as local per transcript coordinates?}
}
\value{
\itemize{
\item{\code{modify}: }{the updated \code{Modifier} object.}
\item{\code{modifications}: }{the modifications found as a \code{GRanges}
object.}
}
}
\description{
\code{modify} triggers the search for modifications for a 
\code{\link[=Modifier-class]{Modifier}} class. Usually this is done 
automatically during construction of a \code{Modifier} object.

\code{modifications} is the accessor for the found modifications.
}
\examples{
data(msi,package="RNAmodR")
# modify() triggers the search for modifications in the data contained in
# the Modifier or ModifierSet object
msi <- modify(msi)
mi <- modify(msi[[1]])
}
