% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-Inosine-class.R,
%   R/Modifier-Inosine-viz.R
\docType{methods}
\name{ModInosine-functions}
\alias{ModInosine-functions}
\alias{settings<-,ModInosine-method}
\alias{aggregate,ModInosine-method}
\alias{modify,ModInosine-method}
\alias{visualizeDataByCoord,ModInosine,GRanges-method}
\alias{visualizeData,ModInosine-method}
\alias{visualizeDataByCoord,ModSetInosine,GRanges-method}
\alias{visualizeData,ModSetInosine-method}
\title{Functions for ModInosine}
\usage{
\S4method{settings}{ModInosine}(x) <- value

\S4method{aggregate}{ModInosine}(x, force = FALSE)

\S4method{modify}{ModInosine}(x, force = FALSE)

\S4method{visualizeDataByCoord}{ModInosine,GRanges}(x, coord,
  type = "score", window.size = 15L, ...)

\S4method{visualizeData}{ModInosine}(x, name, from = 1L, to = 30L,
  type = "score", ...)

\S4method{visualizeDataByCoord}{ModSetInosine,GRanges}(x, coord,
  type = "score", window.size = 15L, ...)

\S4method{visualizeData}{ModSetInosine}(x, name, from = 1L, to = 30L,
  type = "score", ...)
}
\arguments{
\item{x}{a \code{\link[RNAmodR:Modifier]{Modifier}} or a 
\code{\link[RNAmodR:ModifierSet]{ModifierSet}} object. For more details see 
also the man pages for the functions mentioned below.}

\item{value}{See \code{\link[=Modifier]{settings}}}

\item{force}{See \code{\link{aggregate}}}

\item{coord, name, from, to, type, window.size, ...}{See 
\code{\link{visualizeData}}}
}
\description{
All of the functions of \code{\link[RNAmodR:Modifier]{Modifier}} and the
\code{\link[RNAmodR:Modifier]{ModifierSet}} classes are inherited by the 
\code{ModInosine} and \code{ModSetInosine} classes.

Check below for the specifically implemented functions.
}
\details{
\code{ModInosine} specific arguments for \link{visualizeData}:
\itemize{
\item{\code{colour.bases} - }{a named character vector of \code{length = 4} 
for the colours of the individual bases. The names are expected to be 
\code{c("G","A","U","C")}}
}
}
