% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-Inosine-class.R
\docType{class}
\name{ModInosine}
\alias{ModInosine}
\alias{Inosine}
\alias{ModifierInosine}
\alias{ICE-Seq}
\alias{ModInosine-class}
\alias{ModSetInosine-class}
\alias{ModSetInosine}
\title{ModInosine}
\usage{
ModInosine(x, annotation, sequences, seqinfo, ...)

ModSetInosine(x, annotation = NA, sequences = NA, seqinfo = NA, ...)
}
\arguments{
\item{x}{the input which can be of the different types depending on whether
a \code{ModRiboMethSeq} or a \code{ModSetRiboMethSeq} object is to be 
constructed. For more information have a look at the documentation of
the \code{\link[RNAmodR:Modifier-class]{Modifier}} and 
\code{\link[RNAmodR:ModifierSet-class]{ModifierSet}} classes.}

\item{annotation}{annotation data, which must match the information contained
in the BAM files. This is parameter is only required if \code{x} if not a 
\code{Modifier} object.}

\item{sequences}{sequences matching the target sequences the reads were 
mapped onto. This must match the information contained in the BAM files. This
is parameter is only required if \code{x} if not a \code{Modifier} object.}

\item{seqinfo}{An optional \code{\link[GenomeInfoDb:Seqinfo-class]{Seqinfo}} 
argument or character vector, which can be coerced to one, to subset the 
sequences to be analyzed on a per chromosome basis.}

\item{...}{Optional arguments overwriting default values, which are
\itemize{
\item{minCoverage:} {The minimal coverage at the position as integer value 
(default: \code{minCoverage = 10L}).}
\item{minReplicate:} {minimum number of replicates needed for the analysis 
(default: \code{minReplicate = 1L}).}
\item{minScore:} {minimum score to identify Inosine positions de novo 
(default: \code{minScore = 0.4}).}
}}
}
\value{
a \code{ModInosine} or \code{ModSetInosine} object
}
\description{
Inosine can be detected in RNA-Seq data by the conversion of A positions to 
G. This conversion is detected by \code{ModInosine} and used to search for 
Inosine positions.

Only samples named \code{treated} are used for this analysis, since the
A to G conversion is common feature among the reverse transcriptases usually
emploied. Let us know, if that is not the case, and the class needs to be
modified.

Further information on \code{\link[=ModInosine-functions]{Functions}} of 
\code{ModInosine}.
}
\details{
\code{ModInosine} score: the scores for reported Inosine positions are 
between 0 and 1. They are calculated as the relative amount of called G bases 
(\code{(G / N)}) and only kept for A positions.
}
\examples{
# construction of ModInosine object
RNAmodR.files <- RNAmodR.Data::RNAmodR.files
annotation <- rtracklayer::GFF3File(RNAmodR.files[["example1.gff3"]])
sequences <- Rsamtools::FaFile(RNAmodR.files[["example1.fasta"]])
files <- c(Treated = RNAmodR.files[["example_wt_1.bam"]],
           Treated = RNAmodR.files[["example_wt_2.bam"]],
           Treated = RNAmodR.files[["example_wt_3.bam"]])
mi <- ModInosine(files,annotation = annotation ,sequences = sequences)
# construction of ModSetInosine object
files <- list("SampleSet1" = c(treated = RNAmodR.files[["example_wt_1.bam"]],
                               treated = RNAmodR.files[["example_wt_2.bam"]],
                               treated = RNAmodR.files[["example_wt_3.bam"]]),
              "SampleSet2" = c(treated = RNAmodR.files[["example_bud23_1.bam"]],
                               treated = RNAmodR.files[["example_bud23_2.bam"]]),
              "SampleSet3" = c(treated = RNAmodR.files[["example_trm8_1.bam"]],
                               treated = RNAmodR.files[["example_trm8_2.bam"]]))
msi <- ModSetInosine(files, annotation = annotation, sequences = sequences)
}
\author{
Felix G.M. Ernst [aut]
}
