% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-Inosine-class.R
\docType{class}
\name{ModInosine}
\alias{ModInosine}
\alias{Inosine}
\alias{ModifierInosine}
\alias{ICE-Seq}
\alias{ModInosine-class}
\alias{ModSetInosine-class}
\alias{ModSetInosine}
\title{ModInosine}
\usage{
ModInosine(x, annotation, sequences, seqinfo, ...)

ModSetInosine(x, annotation, sequences, seqinfo, ...)
}
\description{
Inosine can be detected in sequence by the conversion of A positions to G.
\code{ModInosine} uses to search for Inosine positions.

\code{\link{ModInosine-functions}}
}
\details{
\code{ModInosine} score: the score for reported Inosine positions are between
1 and 100. It is calculated as \code{min(log2(percent(G) /
(percent(A)+percent(T)+percent(C))) * 10,100)}.
}
\author{
Felix G.M. Ernst [aut]
}
