% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-Inosine-class.R,
%   R/Modifier-Inosine-viz.R
\docType{class}
\name{ModInosine}
\alias{ModInosine}
\alias{Inosine}
\alias{ModifierInosine}
\alias{ICE-Seq}
\alias{ModInosine-class}
\alias{ModInosine,character-method}
\alias{ModInosine,BamFileList-method}
\alias{ModInosine,SequenceData-method}
\alias{modify,ModInosine-method}
\alias{ModSetInosine-class}
\alias{ModSetInosine}
\alias{visualizeData}
\alias{visualizeDataByCoord,ModInosine,GRanges-method}
\alias{visualizeData,ModInosine-method}
\title{ModInosine}
\usage{
\S4method{settings}{ModInosine}(x) <- value

\S4method{ModInosine}{character}(x, fasta, gff, modifications = NULL,
  ...)

\S4method{ModInosine}{BamFileList}(x, fasta, gff, modifications = NULL,
  ...)

\S4method{ModInosine}{SequenceData}(x, modifications = NULL, ...)

\S4method{aggregate}{ModInosine}(x, force = FALSE)

\S4method{modify}{ModInosine}(x, force = FALSE)

ModSetInosine(x, fasta = NA, gff = NA)

\S4method{visualizeDataByCoord}{ModInosine,GRanges}(x, coord,
  type = "score", window.size = 15L, ...)

\S4method{visualizeData}{ModInosine}(x, name, from = 1L, to = 30L,
  type = "score", ...)
}
\description{
title
}
\details{
\code{ModInosine} score: the score for reported Inosine positions are between
1 and 100. It is calculated as \code{min(log2(percent(G) /
(percent(A)+percent(T)+percent(C))) * 10,100)}.

\code{ModInosine} specific arguments for \link{visualizeData}:
\itemize{
\item{\code{colour.bases} - }{a named character vector of \code{length = 4} 
for the colours of the individual bases. The names are expected to be 
\code{c("G","A","U","C")}}
}
}
\author{
Felix G.M. Ernst \email{felix.gm.ernst@outlook.com}
}
