% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-Inosine-class.R
\docType{class}
\name{ModInosine}
\alias{ModInosine}
\alias{Inosine}
\alias{ModifierInosine}
\alias{ModInosine-class}
\alias{ModInosine,character-method}
\alias{ModInosine,BamFileList-method}
\alias{ModInosine,PosData-method}
\title{ModInosine}
\usage{
\S4method{ModInosine}{character}(x, fasta, gff, modifications = NULL,
  ...)

\S4method{ModInosine}{BamFileList}(x, fasta, gff, modifications = NULL,
  ...)

\S4method{ModInosine}{PosData}(x, modifications = NULL, ...)
}
\description{
title
}
