% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-Inosine-class.R
\docType{class}
\name{ModInosine}
\alias{ModInosine}
\alias{Inosine}
\alias{ModifierInosine}
\alias{ModInosine-class}
\alias{ModInosine,character-method}
\alias{ModInosine,BamFileList-method}
\alias{ModInosine,SequenceData-method}
\alias{modify,ModInosine-method}
\title{ModInosine}
\usage{
\S4method{ModInosine}{character}(x, fasta, gff, modifications = NULL,
  ...)

\S4method{ModInosine}{BamFileList}(x, fasta, gff, modifications = NULL,
  ...)

\S4method{ModInosine}{SequenceData}(x, modifications = NULL, ...)

\S4method{modify}{ModInosine}(x, ...)
}
\description{
title

\code{score}: the score for reported Inosine positions are between 1 and 100.
It is calculated as \code{min(log2(percent(G)/percent(A)) * 10,100)}.
}
