% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-normalized-end-pos.R
\docType{class}
\name{NormEndSequenceData}
\alias{NormEndSequenceData}
\alias{NormEnd5SequenceData}
\alias{NormEnd3SequenceData}
\alias{NormEnd5SequenceData-class}
\alias{NormEnd3SequenceData-class}
\title{NormEndSequenceData}
\usage{
NormEnd5SequenceData(bamfiles, fasta, gff, ...)

NormEnd3SequenceData(bamfiles, fasta, gff, ...)

\S4method{aggregate}{NormEnd5SequenceData}(x, condition = c("Both",
  "Treated", "Control"))

\S4method{aggregate}{NormEnd3SequenceData}(x, condition = c("Both",
  "Treated", "Control"))
}
\description{
title
}
