% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmodR-parse.R
\docType{methods}
\name{parseForModifications}
\alias{parseForModifications}
\alias{parseForModifications}
\alias{parseForModifications,RNAmodR,numeric,character-method}
\alias{parseForModificationsWithArgs,RNAmodRargs,character,GRanges,FaFile-method}
\title{parseForModifications}
\usage{
parseForModifications(x, number, param)

parseForModificationsWithArgs(x, name, gff, fasta)

\S4method{parseForModifications}{RNAmodR,numeric,character}(x, number, param)


  \S4method{parseForModificationsWithArgs}{RNAmodRargs,character,GRanges,FaFile}(x,
  name, gff, fasta)
}
\arguments{
\item{x}{a RNAmod object.}

\item{number}{a number defining an experiment.}

\item{name}{a character string for identification}

\item{gff}{a gff3 file name}

\item{fasta}{a fasta file name}

\item{args}{arguments for modification parsing as an RNAmodRargs object}

\item{files}{file names for the bam input files}

\item{conditions}{treated or control}
}
\value{
TRUE if the analysis does finish without error
}
\description{
\code{parseForModifications} uses the input information from GRanges, Fafile
and bam input files to look for known patterns of RNA post-transcriptional
modifications.
}
\examples{
\donttest{
unzip(system.file("extdata", package = "RNAmodR", file = "RNAmodR.zip" ))
mod <- RNAmodR("test",
               "test_layout.csv",
               "test_gff.gff3",
               "test_masked.fasta")
parseForModifications(mod,1)               
}
}
