% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmodR-parse.R
\docType{methods}
\name{parseForModifications}
\alias{parseForModifications}
\alias{parseForModifications,RNAmodR,numeric,character-method}
\title{parseForModifications}
\usage{
parseForModifications(.Object, number, modifications)

\S4method{parseForModifications}{RNAmodR,numeric,character}(.Object, number,
  modifications)
}
\arguments{
\item{.Object}{a RNAmod object.}

\item{number}{a number defining an experiment.}

\item{modifications}{name of modification to be used for analysis.}
}
\value{
TRUE if the analysis does finish without error
}
\examples{
\donttest{
unzip(system.file("extdata", package = "RNAmodR", file = "RNAmodR.zip" ))
mod <- RNAmodR("test",
               "test_layout.csv",
               "test_gff.gff3",
               "test_masked.fasta")
parseForModifications(mod,1,"m7G")               
}
}
