% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmod-m7G-D.R
\docType{methods}
\name{parseMod}
\alias{parseMod}
\alias{parseMod,mod_m7G-method}
\alias{parseMod,mod_D-method}
\title{parseMod}
\usage{
parseMod(object, data)

\S4method{parseMod}{mod_m7G}(object, data)

\S4method{parseMod}{mod_D}(object, data)
}
\arguments{
\item{object}{mod object}
}
\description{
\code{mod_m7G}

\code{mod_D}
}
