% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/class-RNAmod-D.R,
%   R/class-RNAmod-default.R, R/class-RNAmod-m3C.R, R/class-RNAmod-m7G.R
\docType{methods}
\name{parseMod}
\alias{parseMod}
\alias{parseMod,mod_D,GRanges,FaFile,list-method}
\alias{parseMod,mod_default,GRanges,FaFile,list-method}
\alias{parseMod,mod_m3C,GRanges,FaFile,list-method}
\alias{parseMod,mod_m7G,GRanges,FaFile,list-method}
\title{parseMod}
\usage{
parseMod(object, gff, fafile, data)

\S4method{parseMod}{mod_D,GRanges,FaFile,list}(object, gff, fafile, data)

\S4method{parseMod}{mod_default,GRanges,FaFile,list}(object, gff, fafile, data)

\S4method{parseMod}{mod_m3C,GRanges,FaFile,list}(object, gff, fafile, data)

\S4method{parseMod}{mod_m7G,GRanges,FaFile,list}(object, gff, fafile, data)
}
\arguments{
\item{object}{mod object}

\item{gff}{a GRanges object for the genome}

\item{fafile}{a FaFile object for the genome}

\item{data}{list of DataFrame containg the read data}

\item{counts}{total counts in the BAM file containing the data}
}
\description{
\code{mod_D}

\code{mod_m7G}

\code{mod_m3C}

\code{mod_m7G}
}
