% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmod-analysis-default.R, R/class-RNAmod-mod-D.R,
%   R/class-RNAmod-mod-m3C.R, R/class-RNAmod-mod-m7G.R
\docType{methods}
\name{parseMod}
\alias{parseMod}
\alias{parseMod,analysis_default,GRanges,FaFile,list-method}
\alias{checkForModification,mod_D,DataFrame,numeric-method}
\alias{checkForModification,mod_D,DataFrame,numeric-method}
\alias{checkForModification,mod_D,DataFrame,numeric-method}
\title{parseMod}
\usage{
parseMod(object, gff, fafile, modifications)

\S4method{parseMod}{analysis_default,GRanges,FaFile,list}(object, gff, fafile,
  modifications)

\S4method{checkForModification}{mod_D,DataFrame,numeric}(object, data,
  locations)

\S4method{checkForModification}{mod_D,DataFrame,numeric}(object, data,
  locations)

\S4method{checkForModification}{mod_D,DataFrame,numeric}(object, data,
  locations)
}
\arguments{
\item{object}{mod object}

\item{gff}{a GRanges object for the genome}

\item{fafile}{a FaFile object for the genome}

\item{modifications}{list of modification used for parsing}
}
\description{
\code{analysis_default}

\code{mod_D}

\code{mod_m3C}

\code{mod_m7G}
}
