% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/methods-RNAmod-D.R,
%   R/methods-RNAmod-m7G.R
\docType{methods}
\name{parseMod}
\alias{parseMod}
\alias{parseMod,mod_D,numeric,GRanges,DNAString,DataFrame-method}
\alias{parseMod,mod_m7G,numeric,GRanges,DNAString,DataFrame-method}
\title{parseMod}
\usage{
parseMod(object, counts, gff, seq, data)

\S4method{parseMod}{mod_D,numeric,GRanges,DNAString,DataFrame}(object, counts,
  gff, seq, data)

\S4method{parseMod}{mod_m7G,numeric,GRanges,DNAString,DataFrame}(object, counts,
  gff, seq, data)
}
\arguments{
\item{object}{mod object}

\item{counts}{total counts in the BAM file containing the data}

\item{gff}{a GRanges object for a single gene}

\item{seq}{a DNAString object for a single gene}

\item{data}{the read data as DataFrame for a single gene}
}
\description{
\code{mod_D}

\code{mod_m7G}
}
