% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmod-analysis-default.R
\docType{methods}
\name{parseMod}
\alias{parseMod}
\alias{parseMod,analysis_default,GRanges,FaFile,list-method}
\title{parseMod}
\usage{
parseMod(object, gff, fafile, modClasses)

\S4method{parseMod}{analysis_default,GRanges,FaFile,list}(object, gff, fafile,
  modClasses)
}
\arguments{
\item{object}{mod object}

\item{gff}{a GRanges object for the genome}

\item{fafile}{a FaFile object for the genome}

\item{modClasses}{list of modification classes used for parsing}
}
\description{
\code{analysis_default}
}
