% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-pileup.R
\docType{class}
\name{PileupSequenceData}
\alias{PileupSequenceData}
\alias{PileupSequenceData-class}
\alias{aggregate,PileupSequenceData-method}
\title{PileupSequenceData}
\usage{
PileupSequenceData(bamfiles, annotation, sequences, seqinfo, ...)

\S4method{aggregate}{PileupSequenceData}(x, condition = c("Both",
  "Treated", "Control"))
}
\description{
title
}
