% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-pileup.R
\docType{class}
\name{PileupSequenceData}
\alias{PileupSequenceData}
\alias{PileupSequenceData-class}
\alias{aggregate,PileupSequenceData-method}
\alias{getDataTrack,PileupSequenceData-method}
\title{PileupSequenceData}
\usage{
PileupSequenceData(bamfiles, annotation, sequences, seqinfo, ...)

\S4method{aggregate}{PileupSequenceData}(x, condition = c("Both",
  "Treated", "Control"))

\S4method{getDataTrack}{PileupSequenceData}(x, name, ...)
}
\arguments{
\item{bamfiles, annotation, sequences, seqinfo, ...}{See 
\code{\link[=SequenceData-class]{SequenceData}}}

\item{x}{a \code{PileupSequenceData}}

\item{condition}{For \code{\link{aggregate}}: condition for which the data 
should be aggregated.}

\item{name}{For \code{getDataTrack}: a valid transcript name. Must be a name
of \code{ranges(x)}}
}
\value{
a \code{PileupSequenceData} object
}
\description{
title
}
\examples{
# Construct a End5SequenceData object
annotation <- system.file("extdata","example1.gff3",package = "RNAmodR.Data")
sequences <- system.file("extdata","example1.fasta",package = "RNAmodR.Data")
files <- c(control = system.file("extdata","example_wt_1.bam",
                                 package = "RNAmodR.Data"),
           treated = system.file("extdata","example_wt_2.bam",
                                 package = "RNAmodR.Data"))
psd <- PileupSequenceData(files, annotation = annotation,
                          sequences = sequences)
# aggregate data
aggregate(psd)
}
