% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotEnrichment.Generic}
\alias{PlotEnrichment.Generic}
\title{Generic function to perform enrichment analysis and plot results.}
\usage{
PlotEnrichment.Generic(mat, title = "", xlab = "", x.las = 1,
  x.cex = 1, x.padj = 1, plotType = "l", revXaxis = FALSE,
  xAxisLblFmt = 0)
}
\arguments{
\item{mat}{A \code{dataframe}; specifies the contingency table.}

\item{title}{A character string; specifies plot title.}

\item{xlab}{A character string; specifies the x-axis label.}

\item{x.las}{An integer scalar; specifies the orientation of
x-axis labels; Default is 1.}

\item{x.cex}{A float scalar; specifies a scaling factor for
x-axis labels; default is 1.}

\item{x.padj}{A float scalar; specifies the vertical adjustment
of x-axis labels; default is 1.}

\item{plotType}{A character string; specifies the plot style;
default is "l" (for line).}

\item{revXaxis}{A logical scalar; if \code{TRUE}, the order
of columns from \code{mat} is reversed; default is \code{FALSE}.}

\item{xAxisLblFmt}{An integer scalar; specifies the
format of the x-axis label.}
}
\value{
A list of \code{fisher.test} return objects and \code{mat}.
}
\description{
Generic function to perform enrichment analysis and plot results.
Enrichment/depletion is evaluated using (multiple) Fisher's exact test(s).
Multiple hypothesis testing correction is applied following the method of
Bejamini and Hochberg.
Note: This function should not be invoked directly by the user.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
\keyword{internal}
