% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotGC}
\alias{PlotGC}
\title{Plot GC content.}
\usage{
PlotGC(locPos, locNeg, filter = NULL, geneNorm = FALSE)
}
\arguments{
\item{locPos}{A \code{txLoc} object.}

\item{locNeg}{A \code{txLoc} object.}

\item{filter}{A logical scalar; only consider loci in transcript
regions specified in filter; Default is \code{NULL}.}

\item{geneNorm}{A logical scalar; if \code{TRUE} normalise GC
content in window to GC content of transcript section; default
is \code{FALSE}.}
}
\description{
Plot GC content.
}

