% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotGC}
\alias{PlotGC}
\title{Plot GC content.}
\usage{
PlotGC(locPos, locNeg, flank = 10, filter = NULL, geneNorm = FALSE,
  subsample = TRUE)
}
\arguments{
\item{locPos}{A \code{txLoc} object.}

\item{locNeg}{A \code{txLoc} object.}

\item{flank}{An integer scalar; see 'Details'.}

\item{filter}{A logical scalar; only consider loci in transcript
regions specified in filter; Default is \code{NULL}.}

\item{geneNorm}{A logical scalar; if \code{TRUE} normalise GC
content in window to GC content of transcript section; default
is \code{FALSE}.}

\item{subsample}{A logical scalar; if \code{TRUE} a subsample
of \code{locNeg} is used instead of the full set; the subsample 
size is dynamically determined based on the total number of sites
in \code{locPos}; default is \code{TRUE}.}
}
\description{
Plot GC content.
}
\details{
The function calculates GC contents within a region around every
site from two \code{txLoc} objects. The window is defined
by extending the position of every transcript locus upstream and
downstream by \code{flank} nucleotides (if possible).
The means of resulting GC content distributions are compared and
assessed using a two-tailed t-test.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

