% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotOverlap}
\alias{PlotOverlap}
\title{Plot overlap of two txLoc objects.}
\usage{
PlotOverlap(loc1, loc2)
}
\arguments{
\item{loc1}{A \code{txLoc} object.}

\item{loc2}{A \code{txLoc} object.}
}
\description{
Plot overlap of two txLoc objects.
}
\details{
This function plots the overlap of entries from two
\code{txLoc} objects as Venn diagrams.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

