% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotPieNumberOfLoci}
\alias{PlotPieNumberOfLoci}
\title{Plot piechart of the number of loci in every transcript section.}
\usage{
PlotPieNumberOfLoci(locus)
}
\arguments{
\item{locus}{A \code{txLoc} object.}
}
\description{
Plot piechart of the number of loci in every transcript section.
}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR");
sites <- ReadBED(bedFile);
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38");
PlotPieNumberOfLoci(posSites);
}

}

