% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotRelDistDistribution}
\alias{PlotRelDistDistribution}
\title{Plot distribution of relative distances between sites
from two \code{txLoc} objects.}
\usage{
PlotRelDistDistribution(loc1, loc2, flank = 1000, binWidth = 20,
  doBootstrap = TRUE)
}
\arguments{
\item{loc1}{A \code{txLoc} object.}

\item{loc2}{A \code{txLoc} object.}

\item{flank}{An integer scalar; specifies the absolute maximum
relative distance used as a cutoff; default is 1000.}

\item{binWidth}{An integer scalar; specifies the spatial width
by which distances will be binned; default is 20.}

\item{doBootstrap}{A logical scalar; if \code{YES} calculate
95% CI based on empirical bootstrap of sites within transcript
region; default is \code{TRUE}.}
}
\description{
Plot distribution of relative distances between sites
from two \code{txLoc} objects.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

