% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotRelDistEnrichment}
\alias{PlotRelDistEnrichment}
\title{Perform enrichment analysis of relative distances.}
\usage{
PlotRelDistEnrichment(locPos, locNeg, locRef, flank = 1000, binWidth = 20)
}
\arguments{
\item{locPos}{A \code{txLoc} object.}

\item{locNeg}{A \code{txLoc} object.}

\item{locRef}{A \code{txLoc} object.}

\item{flank}{An integer scalar; specifies the absolute maximum
relative distance used as a cutoff; default is 1000.}

\item{binWidth}{An integer scalar; specifies the spatial width
by which distances will be binned; default is 20.}
}
\description{
Perform enrichment analysis and plot results of two relative
distance distributions. See 'Details'.
}
\details{
The function calculates minimum distances between entries from
\code{locPos} relative to \code{locRef}, and \code{locNeg}
relative to \code{locRef}. Enrichment/depletion is assessed
using multiple Fisher's exact tests on the counts per distance
bin relative to the counts in all other bins within the window
defined by (-\code{flank}, \code{flank}). Resulting enrichment
plots show odds-ratios (including 95\% confidence intervals) and
associated p-values as a function of relative distance bins.
Negative distances correspond to an enrichment/depletion of
sites from \code{locPos} relative to sites from \code{locNeg}
upstream of the closest site from \code{locRef}. Positive
distances correspond to a downstream enrichment/depletion
relative to sites from \code{locRef}.
The bin width and window size can be adjusted with
\code{flank} and \code{binWidth}.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

