% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotRelDistEnrichment}
\alias{PlotRelDistEnrichment}
\title{Perform enrichment analysis of relative distances.}
\usage{
PlotRelDistEnrichment(txLoc1, txLoc2, txLocRef, flank = 1000,
  binWidth = 20)
}
\arguments{
\item{txLoc1}{A \code{txLoc} object.}

\item{txLoc2}{A \code{txLoc} object.}

\item{txLocRef}{A \code{txLoc} object.}

\item{flank}{An integer scalar or an integer vector of length 2; specifies 
the absolute maximum relative distance(s) used as a cutoff; by specifying 
\code{flank} as a vector, a non-symmetric window can be  defined; e.g. 
`flank = c(1000, 2000)` corresponds to a window defined by 1000 nt 
downstream and 2000 nt upstream of the reference site. Default is 1000.}

\item{binWidth}{An \code{integer} scalar; specifies the spatial width in nt 
by which distances will be binned; default is \code{20}.}
}
\value{
\code{NULL}.
}
\description{
Perform enrichment analysis and plot results of two relative
distance distributions. See 'Details'.
}
\details{
The function calculates minimum distances per transcript region, between
entries from \code{txLoc1} relative to \code{txLocRef}, and \code{txLoc2}
relative to \code{txLocRef}.
Enrichment/depletion is assessed using multiple Fisher's exact tests on the
counts per distance bin relative to the counts in all other bins within the
window defined by (-\code{flank}, \code{flank}). Resulting enrichment plots
show odds-ratios (including 95\% confidence intervals) and associated
p-values as a function of relative distance bins. Negative distances
indicate sites from \code{txLoc1} and \code{txLoc2} that are \emph{upstream}
of sites from \code{txLocRef}; positive distances correspond to sites from
\code{txLoc1} and \code{txLoc2} that are \emph{downstream} of
\code{txLocRef}. The bin width and window size can be adjusted with
\code{flank} and \code{binWidth}.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
