% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-roc.R
\docType{methods}
\name{plotROC}
\alias{plotROC}
\alias{plotROC,Modifier-method}
\alias{plotROC,ModifierSet-method}
\title{Performance}
\usage{
\S4method{plotROC}{Modifier}(x, coord, redo, redo.args, prediction.args,
  performance.args, plot.args)

\S4method{plotROC}{ModifierSet}(x, coord, redo, redo.args, prediction.args,
  performance.args, plot.args)
}
\description{
title
}
