% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotSectionDistribution}
\alias{PlotSectionDistribution}
\title{Plot piechart of the number of loci in every transcript region}
\usage{
PlotSectionDistribution(txLoc)
}
\arguments{
\item{txLoc}{A \code{txLoc} object.}
}
\value{
\code{NULL}.
}
\description{
Plot piechart of the number of loci in every transcript region.
}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR")
sites <- ReadBED(bedFile)
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38")
PlotSectionDistribution(posSites)
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
