% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotSectionEnrichment}
\alias{PlotSectionEnrichment}
\title{Perform transcript section enrichment analysis and plot results.}
\usage{
PlotSectionEnrichment(locPos, locNeg, filter = NULL, xAxisLblFmt = 2)
}
\arguments{
\item{locPos}{A \code{txLoc} object. These should be the positive control sites.}

\item{locNeg}{A \code{txLoc} object. These should be the negative control sites.}

\item{filter}{Only plot loci in transcript regions specified in filter.  Default is NULL.}

\item{xAxisLblFmt}{Plot extended axis labels. Default is 2. See ??? for details.}
}
\description{
Perform transcript section enrichment analysis and plot results.
Enrichment/depletion is evaluated using (multiple) Fisher's exact test(s).
Multiple hypothesis testing correction is applied following the method of
Bejamini and Hochberg.
}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR");
sites <- ReadBED(bedFile);
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38");
negSites <- GenerateNull(posSites, method = "nuclAbundance");
PlotSectionEnrichment(posSites, negSites,
                      filter = c("5'UTR", "CDS", "3'UTR"));
}

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
