% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotSeqLogo}
\alias{PlotSeqLogo}
\title{Plot sequence logo.}
\usage{
PlotSeqLogo(locus, flank = 5, filter = NULL, ylim = c(0, 2))
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{flank}{An integer scalar; see 'Details'.}

\item{filter}{A character vector; only plot sequence logos
of sections specified in \code{filter}; if \code{NULL} plot
all sections; default is NULL.}

\item{ylim}{An integer vector; specifies limits for the
y-axis; automatically determined if \code{ymin = NULL};
default is \code{c(0, 2)}.}
}
\description{
Plot sequence logo.
}
\details{
The function determines the sequence logo within a window
defined by extending sites from \code{locus} upstream and
downstream by \code{flank} nucleotides. By default logos
are shown for every transcript section from \code{locus}.
Use \code{filter} to specify specific transcript sections.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

