% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotSpatialEnrichment}
\alias{PlotSpatialEnrichment}
\title{Perform spatial enrichment analysis and plot results.}
\usage{
PlotSpatialEnrichment(locPos, locNeg, filter = NULL, binWidth = 20,
  posMax = 1000)
}
\arguments{
\item{locPos}{A txLoc object. These should be the positive control sites.}

\item{locNeg}{A txLoc object. These should be the negative control sites.}

\item{filter}{Only plot loci in transcript regions specified in filter.}

\item{binWidth}{Spatial bin width. Default is 20 nt.}

\item{posMax}{Evaluate enrichment within a window given by \code{posMax}.
Default is 1000 nt.}
}
\description{
Perform spatial enrichment analysis and plot results.
Enrichment/depletion is evaluated using (multiple) Fisher's exact test(s).
Multiple hypothesis testing correction is applied following the method of
Bejamini and Hochberg.
}
\examples{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR");
sites <- ReadBED(bedFile);
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38");
negSites <- GenerateSNMNull(posSites, method = "permutation");
PlotSpatialEnrichment(posSites, negSites,
                      filter = c("5'UTR", "CDS", "3'UTR"));

}
\keyword{}

