% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosDataFrame-class.R
\docType{class}
\name{PosDataFrame}
\alias{PosDataFrame}
\alias{PosDataFrame-class}
\alias{getSeq,PosDataFrame-method}
\alias{ranges,PosDataFrame-method}
\title{PosDataFrame}
\usage{
PosDataFrame(df, ranges, sequence)

\S4method{getSeq}{PosDataFrame}(x)

\S4method{ranges}{PosDataFrame}(x)
}
\description{
title
}
