% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/class-RNAmodR-mod-D.R,
%   R/class-RNAmodR-mod-m3C.R, R/class-RNAmodR-mod-m7G.R
\docType{methods}
\name{preTest}
\alias{preTest}
\alias{preTest,mod_D,numeric,numeric,list-method}
\alias{preTest,mod_m3C,numeric,numeric,list-method}
\alias{preTest,mod_m7G,numeric,numeric,list-method}
\title{Testing data for compatibility with modification detection}
\usage{
preTest(object, location, locations, data)

\S4method{preTest}{mod_D,numeric,numeric,list}(object, location, locations,
  data)

\S4method{preTest}{mod_m3C,numeric,numeric,list}(object, location, locations,
  data)

\S4method{preTest}{mod_m7G,numeric,numeric,list}(object, location, locations,
  data)
}
\arguments{
\item{object}{a mod class}

\item{location}{a local position on a transcript}

\item{locations}{all local positions on the transcript}

\item{data}{the data to be analyzed}
}
\description{
\code{mod_D}

\code{mod_m3C}

\code{mod_m7G}
}
