% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/class-RNAmodR-mod-D.R,
%   R/class-RNAmodR-mod-m3C.R, R/class-RNAmodR-mod-m7G.R
\docType{methods}
\name{preTest}
\alias{preTest}
\alias{preTest,mod_D,numeric,list,numeric,numeric-method}
\alias{preTest,mod_m3C,numeric,list,numeric,numeric-method}
\alias{preTest,mod_m7G,numeric,list,numeric,numeric-method}
\title{preTest}
\usage{
preTest(object, location, data, spmThreshold, locations)

\S4method{preTest}{mod_D,numeric,list,numeric,numeric}(object, location, data,
  spmThreshold, locations)

\S4method{preTest}{mod_m3C,numeric,list,numeric,numeric}(object, location, data,
  spmThreshold, locations)

\S4method{preTest}{mod_m7G,numeric,list,numeric,numeric}(object, location, data,
  spmThreshold, locations)
}
\description{
\code{mod_D}

\code{mod_m3C}

\code{mod_m7G}
}
