% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-protected-end-pos.R
\docType{class}
\name{ProtectedEndSequenceData}
\alias{ProtectedEndSequenceData}
\alias{ProtectedEndSequenceData-class}
\alias{aggregate,ProtectedEndSequenceData-method}
\alias{getDataTrack,ProtectedEndSequenceData-method}
\title{ProtectedEndSequenceData}
\usage{
ProtectedEndSequenceData(bamfiles, annotation, sequences, seqinfo, ...)

\S4method{aggregate}{ProtectedEndSequenceData}(x, condition = c("Both",
  "Treated", "Control"))

\S4method{getDataTrack}{ProtectedEndSequenceData}(x, name, ...)
}
\arguments{
\item{bamfiles, annotation, sequences, seqinfo, ...}{See 
\code{\link[=SequenceData-class]{SequenceData}}}

\item{x}{a \code{ProtectedEndSequenceData}}

\item{condition}{For \code{\link{aggregate}}: condition for which the data 
should be aggregated.}

\item{name}{For \code{getDataTrack}: a valid transcript name. Must be a name
of \code{ranges(x)}}
}
\value{
a \code{ProtectedEndSequenceData} object
}
\description{
title
}
\examples{
# Construct a End5SequenceData object
annotation <- system.file("extdata","example1.gff3",package = "RNAmodR.Data")
sequences <- system.file("extdata","example1.fasta",package = "RNAmodR.Data")
files <- c(control = system.file("extdata","example_wt_1.bam",
                                 package = "RNAmodR.Data"),
           treated = system.file("extdata","example_wt_2.bam",
                                 package = "RNAmodR.Data"))
pesd <- ProtectedEndSequenceData(files, annotation = annotation,
                                 sequences = sequences)
# aggregate data
aggregate(pesd)
}
