% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-protected-end-pos.R
\docType{class}
\name{ProtectedEndSequenceData}
\alias{ProtectedEndSequenceData}
\alias{ProtectedEndSequenceData-class}
\alias{aggregate,ProtectedEndSequenceData-method}
\alias{getDataTrack,ProtectedEndSequenceData-method}
\title{ProtectedEndSequenceData}
\usage{
ProtectedEndSequenceData(bamfiles, annotation, sequences, seqinfo, ...)

\S4method{aggregate}{ProtectedEndSequenceData}(x, condition = c("Both",
  "Treated", "Control"))

\S4method{getDataTrack}{ProtectedEndSequenceData}(x, ...)
}
\arguments{
\item{bamfiles, annotation, sequences, seqinfo, ...}{See 
\code{\link[=SequenceData-class]{SequenceData}}}

\item{x}{a \code{CoverageSequenceData}}

\item{condition}{For \code{\link{aggregate}}: condition for which the data 
should be aggregated.}
}
\description{
title
}
