% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/class-RNAmodR-quant-5end.R, R/class-RNAmodR-quant-allends.R,
%   R/class-RNAmodR-quant-methends.R
\docType{methods}
\name{quantifiyReadDataPerTranscript}
\alias{quantifiyReadDataPerTranscript}
\alias{quantifiyReadDataPerTranscript,RNAmodRquant_5end,GAlignmentsList,RNAmodRargs,numeric,GRanges,FaFile-method}
\alias{quantifiyReadDataPerTranscript,RNAmodRquant_allends,GAlignmentsList,RNAmodRargs,numeric,GRanges,FaFile-method}
\alias{quantifiyReadDataPerTranscript,RNAmodRquant_methends,GAlignmentsList,RNAmodRargs,numeric,GRanges,FaFile-method}
\title{Returns the data for each transcript}
\usage{
quantifiyReadDataPerTranscript(x, bamData, args, counts, gff, fafile)


  \S4method{quantifiyReadDataPerTranscript}{RNAmodRquant_5end,GAlignmentsList,RNAmodRargs,numeric,GRanges,FaFile}(x,
  bamData, args, counts, gff, fafile)


  \S4method{quantifiyReadDataPerTranscript}{RNAmodRquant_allends,GAlignmentsList,RNAmodRargs,numeric,GRanges,FaFile}(x,
  bamData, args, counts, gff, fafile)


  \S4method{quantifiyReadDataPerTranscript}{RNAmodRquant_methends,GAlignmentsList,RNAmodRargs,numeric,GRanges,FaFile}(x,
  bamData, args, counts, gff, fafile)
}
