% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{ReadBED}
\alias{ReadBED}
\title{Read BED-formatted file.}
\usage{
ReadBED(file)
}
\arguments{
\item{file}{A character string; specifies the input BED file.}
}
\value{
A \code{GRanges} object.
}
\description{
Open and read a BED file, and store the annotation
features in a \code{GRanges} object.
}
\examples{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR");
sites <- ReadBED(bedFile);

}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

