% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/class-RNAmod.R
\docType{methods}
\name{RNAmod-Accessors}
\alias{RNAmod-Accessors}
\alias{getInputFolder}
\alias{getOutputFolder}
\alias{getExperimentName}
\alias{getInputFolder,RNAmod-method}
\alias{getOutputFolder,RNAmod-method}
\alias{getExperimentName,RNAmod-method}
\title{Accessors for RNAmod Object}
\usage{
getInputFolder(.Object)

getOutputFolder(.Object)

getExperimentName(.Object)

\S4method{getInputFolder}{RNAmod}(.Object)

\S4method{getOutputFolder}{RNAmod}(.Object)

\S4method{getExperimentName}{RNAmod}(.Object)
}
\arguments{
\item{.Object}{a RNAmod object}
}
\value{
*Folder: the requested folder path

*Name: the title of the experiment collection. This is also the path 
identifier, which is expected to contain the data and source folder.
}
\description{
Accessors for RNAmod Object
}
\examples{
\donttest{
data("RNAmod_test", package = "RNAmod")

getInputFolder(experiment)
getOutputFolder(experiment)
getExperimentName(experiment)
}
}
