% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/class-RNAmodR.R
\docType{methods}
\name{RNAmodR-Accessors}
\alias{RNAmodR-Accessors}
\alias{getInputFolder}
\alias{getOutputFolder}
\alias{getExperimentName}
\alias{getGFFFile}
\alias{getFastaFile}
\alias{getInputFolder,RNAmodR-method}
\alias{getOutputFolder,RNAmodR-method}
\alias{getExperimentName,RNAmodR-method}
\alias{getGFFFile,RNAmodR-method}
\alias{getFastaFile,RNAmodR-method}
\title{Accessors for RNAmodR Object}
\usage{
getInputFolder(.Object)

getOutputFolder(.Object)

getExperimentName(.Object)

getGFFFile(.Object)

getFastaFile(.Object)

\S4method{getInputFolder}{RNAmodR}(.Object)

\S4method{getOutputFolder}{RNAmodR}(.Object)

\S4method{getExperimentName}{RNAmodR}(.Object)

\S4method{getGFFFile}{RNAmodR}(.Object)

\S4method{getFastaFile}{RNAmodR}(.Object)
}
\arguments{
\item{.Object}{a RNAmodR object}
}
\value{
\code{getInputFolder}, \code{getOutputFolder}: the requested folder path

\code{getExperimentName}: the title of the experiment collection. This is 
also the path identifier, which is expected to contain the data and source 
folder.

\code{getGFFFile}: the information of the input gff file as GRanges object

\code{getFastaFile}: the link to the fasta file as FaFile class
}
\description{
Accessors for RNAmodR Object
}
\examples{
\donttest{#' 
unzip(system.file("extdata", package = "RNAmodR", file = "RNAmodR.zip" ))
mod <- RNAmodR("test",
               "test_layout.csv",
               "test_gff.gff3",
               "test_masked.fasta")
getInputFolder(mod)  
getOutputFolder(mod)  
getExperimentName(mod)  
getGFFFile(mod)  
getFastaFile(mod)
}
}
