% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/class-RNAmodR-args.R,
%   R/methods-RNAmodR-args.R
\docType{class}
\name{getInputFiles}
\alias{getInputFiles}
\alias{getConditions}
\alias{getParam}
\alias{loadQuantifier}
\alias{loadIdentifier}
\alias{RNAmodR-args-class}
\alias{RNAmodRargs-class}
\alias{RNAmodRargs}
\alias{getInputFiles,RNAmodRargs-method}
\alias{getConditions,RNAmodRargs-method}
\alias{getParam,RNAmodRargs-method}
\alias{loadQuantifier,RNAmodRargs-method}
\alias{loadIdentifier,RNAmodRargs-method}
\title{RNAmodR arguments class}
\usage{
getInputFiles(x)

getConditions(x)

getParam(x, identifier, param)

loadQuantifier(x)

loadIdentifier(x)

RNAmodRargs(param, files, conditions)

\S4method{getInputFiles}{RNAmodRargs}(x)

\S4method{getConditions}{RNAmodRargs}(x)

\S4method{getParam}{RNAmodRargs}(x, identifier, param)

\S4method{loadQuantifier}{RNAmodRargs}(x)

\S4method{loadIdentifier}{RNAmodRargs}(x)
}
\arguments{
\item{x}{a RNAmodRargs class object}

\item{identifier}{an optional identifier for subsetting the arguments}

\item{x}{a RNAmodRargs class object}

\item{x}{a RNAmodRargs class object}

\item{x}{a RNAmodRargs class object}

\item{x}{a RNAmodRargs class object}
}
\value{
a RNAmodRargs object

a list of input files

the conditions of each input file. Either "Control" or "Treated"

a data.frame with the selected arguments

a list of RNAmodR quantifier classes

a list of RNAmodR identifier class
}
\description{
\code{RNAmodRargs}
}
\section{Slots}{

\describe{
\item{\code{files}}{input files for arguments}

\item{\code{conditions}}{identifier for sample condition. Treated or Control}

\item{\code{args}}{a data.frame containing the arguments}
}}

