% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RNAmodR.R
\docType{class}
\name{RNAmodR-class}
\alias{RNAmodR-class}
\alias{RNAmodR-class}
\alias{RNAmodR}
\alias{show,RNAmodR-method}
\title{RNAmodR Class}
\usage{
RNAmodR(experimentName, experimentLayout, experimentGFF, experimentFasta)

\S4method{show}{RNAmodR}(object)
}
\arguments{
\item{experimentName}{a character string used for identifying an experiment}

\item{experimentLayout}{a csv-file name containing the experimental layout}

\item{experimentGFF}{a file anme for the input gff3 annotation}

\item{experimentFasta}{a file name for the input fasta sequences}

\item{object}{a RNAmod object.}
}
\value{
a RNAmodR object
}
\description{
\code{RNAmodR}
}
\section{Slots}{

\describe{
\item{\code{experimentName}}{}

\item{\code{inputFile}}{}

\item{\code{inputGFF}}{}

\item{\code{inputFasta}}{}

\item{\code{.dataSamples}}{}

\item{\code{.dataFasta}}{}

\item{\code{.dataGFF}}{}

\item{\code{.wd}}{}

\item{\code{.inputFolder}}{}

\item{\code{.outputFolder}}{}

\item{\code{.mapQuality}}{}
}}

\examples{
\donttest{
unzip(system.file("extdata", package = "RNAmodR", file = "RNAmodR.zip" ))
mod <- RNAmodR("test",
               "test_layout.csv",
               "test_gff.gff3",
               "test_masked.fasta")
}
}
