% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.R
\docType{data}
\name{RNAmodR-datasets}
\alias{RNAmodR-datasets}
\alias{msi}
\alias{sds}
\alias{sdl}
\alias{psd}
\alias{e5sd}
\alias{e3sd}
\alias{esd}
\alias{csd}
\alias{ne3sd}
\alias{ne5sd}
\alias{pesd}
\title{Example data in the RNAmodR package}
\format{a \code{ModSetInosine} instance}
\usage{
msi

sds

sdl

psd

e5sd

e3sd

esd

csd

ne3sd

ne5sd

pesd
}
\description{
This contains an example ModifierSet object
}
\keyword{datasets}
