% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.R
\docType{data}
\name{RNAmodR-datasets}
\alias{RNAmodR-datasets}
\alias{msi}
\alias{psd}
\alias{e5sd}
\alias{sds}
\alias{sdl}
\title{Example data in the RNAmodR package}
\format{a \code{ModSetInosine} instance}
\usage{
msi

psd

e5sd

sds

sdl
}
\description{
This contains an example ModifierSet object
}
\keyword{datasets}
