% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.R, R/AllGenerics.R
\name{RNAmodR-internals}
\alias{RNAmodR-internals}
\alias{.dataTrack}
\alias{.dataTracks}
\title{RNAmodR internal functions}
\usage{
.dataTracks(x, data, seqdata, sequence, args)
}
\description{
These functions are not intended for general use, but maybe exported for 
additional package development.
}
