% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.R,
%   R/Gviz-ModifiedSequenceTrack-class.R, R/Gviz-functions.R,
%   R/SequenceDataSet-class.R, R/SequenceDataList-class.R, R/Modifier-subset.R,
%   R/SequenceData-subset.R
\name{RNAmodR-internals}
\alias{RNAmodR-internals}
\alias{.getData}
\alias{show,ModifiedSequenceTrack-method}
\alias{length,ModifiedSequenceTrack-method}
\alias{parallelSlotNames,SequenceDataSet-method}
\alias{getListElement,SequenceDataSet-method}
\alias{parallelSlotNames,SequenceDataList-method}
\alias{getListElement,SequenceDataList-method}
\alias{subset,Modifier-method}
\alias{labelByCoord,SplitDataFrameList,GRanges-method}
\title{RNAmodR internal functions}
\usage{
\S4method{show}{ModifiedSequenceTrack}(object)

\S4method{length}{ModifiedSequenceTrack}(x)

\S4method{parallelSlotNames}{SequenceDataSet}(x)

\S4method{getListElement}{SequenceDataSet}(x, i, exact = TRUE)

\S4method{parallelSlotNames}{SequenceDataList}(x)

\S4method{getListElement}{SequenceDataList}(x, i, exact = TRUE)

\S4method{subset}{Modifier}(x, name, pos = 1L, ...)

\S4method{labelByCoord}{SplitDataFrameList,GRanges}(x, coord, ...)
}
\arguments{
\item{object, range, data, modType, scoreFun, source, type, x, i, j, ..., exact, value}{Internal arguments}
}
\value{
internally used values
}
\description{
These functions are used internally.
}
\keyword{internal}
