% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.R, R/AllGenerics.R, R/Modifier-utils.R,
%   R/SequenceData-class.R, R/Modifier-viz.R, R/SequenceData-coverage.R,
%   R/SequenceData-end-pos.R, R/SequenceData-normalized-end-pos.R,
%   R/SequenceData-pileup.R, R/SequenceData-viz.R,
%   R/SequenceData-protected-end-pos.R
\docType{methods}
\name{RNAmodR-internals}
\alias{RNAmodR-internals}
\alias{.dataTrack}
\alias{.constructModRanges}
\alias{.dataTracks}
\alias{.constructModRanges,GRanges,DataFrame-method}
\alias{.getData,CoverageSequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{.dataTracks,CoverageSequenceData,missing,GRanges,XString-method}
\alias{.getData,End5SequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{.getData,End3SequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{.getData,EndSequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{.dataTracks,EndSequenceData,missing,GRanges,XString-method}
\alias{.dataTracks,End5SequenceData,missing,GRanges,XString-method}
\alias{.dataTracks,End3SequenceData,missing,GRanges,XString-method}
\alias{.getData,NormEnd5SequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{.getData,NormEnd3SequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{.dataTracks,NormEnd5SequenceData,missing,GRanges,XString-method}
\alias{.dataTracks,NormEnd3SequenceData,missing,GRanges,XString-method}
\alias{.getData,PileupSequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{.getData,ProtectedEndSequenceData,GRangesList,XStringSet,ScanBamParam-method}
\title{RNAmodR internal functions}
\usage{
.constructModRanges(range, data, modType, scoreFun, source, type)

.getData(x, grl, sequences, param, args)

.dataTracks(x, data, seqdata, sequence, args)

\S4method{.constructModRanges}{GRanges,DataFrame}(range, data, modType,
  scoreFun, source, type)

\S4method{.dataTracks}{SequenceData,missing,GRanges,XString}(x, seqdata,
  sequence, args)

\S4method{.dataTracks}{Modifier,ANY,ANY,ANY}(x, data, seqdata, sequence,
  args)


  \S4method{.getData}{CoverageSequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)

\S4method{.dataTracks}{CoverageSequenceData,missing,GRanges,XString}(x,
  seqdata, sequence, args)


  \S4method{.getData}{End5SequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)


  \S4method{.getData}{End3SequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)


  \S4method{.getData}{EndSequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)

\S4method{.dataTracks}{EndSequenceData,missing,GRanges,XString}(x, seqdata,
  sequence, args)

\S4method{.dataTracks}{End5SequenceData,missing,GRanges,XString}(x,
  seqdata, sequence, args)

\S4method{.dataTracks}{End3SequenceData,missing,GRanges,XString}(x,
  seqdata, sequence, args)


  \S4method{.getData}{NormEnd5SequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)


  \S4method{.getData}{NormEnd3SequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)

\S4method{.dataTracks}{NormEnd5SequenceData,missing,GRanges,XString}(x,
  seqdata, sequence, args)

\S4method{.dataTracks}{NormEnd3SequenceData,missing,GRanges,XString}(x,
  seqdata, sequence, args)


  \S4method{.getData}{PileupSequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)

\S4method{.dataTracks}{SequenceData,ANY,ANY,ANY}(x, seqdata, sequence,
  args)


  \S4method{.getData}{ProtectedEndSequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)
}
\description{
These functions are not intended for general use, but maybe exported for 
additional package development.
}
