% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.R, R/AllGenerics.R,
%   R/Gviz-ModifiedSequenceTrack-class.R, R/Gviz-functions.R,
%   R/Modifier-utils.R, R/SequenceData-class.R, R/SequenceDataList-class.R,
%   R/SequenceData-coverage.R, R/SequenceData-end-pos.R,
%   R/SequenceData-normalized-end-pos.R, R/SequenceData-pileup.R,
%   R/SequenceData-protected-end-pos.R
\docType{methods}
\name{RNAmodR-internals}
\alias{RNAmodR-internals}
\alias{.getData}
\alias{.constructModRanges}
\alias{show,ModifiedSequenceTrack-method}
\alias{length,ModifiedSequenceTrack-method}
\alias{.constructModRanges,GRanges,DataFrame-method}
\alias{parallelSlotNames,SequenceData-method}
\alias{setListElement,SequenceData-method}
\alias{getListElement,SequenceData-method}
\alias{.getData,SequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{parallelSlotNames,SequenceDataList-method}
\alias{getListElement,SequenceDataList-method}
\alias{.getData,CoverageSequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{.getData,End5SequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{.getData,End3SequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{.getData,EndSequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{.getData,NormEnd5SequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{.getData,NormEnd3SequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{.getData,PileupSequenceData,GRangesList,XStringSet,ScanBamParam-method}
\alias{.getData,ProtectedEndSequenceData,GRangesList,XStringSet,ScanBamParam-method}
\title{RNAmodR internal functions}
\usage{
.constructModRanges(range, data, modType, scoreFun, source, type)

.getData(x, grl, sequences, param, args)

\S4method{show}{ModifiedSequenceTrack}(object)

\S4method{length}{ModifiedSequenceTrack}(x)

\S4method{.constructModRanges}{GRanges,DataFrame}(range, data, modType,
  scoreFun, source, type)

\S4method{parallelSlotNames}{SequenceData}(x)

\S4method{setListElement}{SequenceData}(x, i, value)

\S4method{getListElement}{SequenceData}(x, i, exact = TRUE)

\S4method{.getData}{SequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)

\S4method{parallelSlotNames}{SequenceDataList}(x)

\S4method{getListElement}{SequenceDataList}(x, i, exact = TRUE)


  \S4method{.getData}{CoverageSequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)


  \S4method{.getData}{End5SequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)


  \S4method{.getData}{End3SequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)


  \S4method{.getData}{EndSequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)


  \S4method{.getData}{NormEnd5SequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)


  \S4method{.getData}{NormEnd3SequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)


  \S4method{.getData}{PileupSequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)


  \S4method{.getData}{ProtectedEndSequenceData,GRangesList,XStringSet,ScanBamParam}(x,
  grl, sequences, param, args)
}
\arguments{
\item{object, range, data, modType, scoreFun, source, type, x, grl, sequences, param, args, i, exact, value}{Internal arguments}
}
\value{
internally used values
}
\description{
These functions are not intended for general use, but are used for 
additional package development.
}
