% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.R, R/AllGenerics.R, R/Modifier-utils.R
\docType{methods}
\name{RNAmodR-internals}
\alias{RNAmodR-internals}
\alias{.dataTrack}
\alias{.constructModRanges}
\alias{.dataTracks}
\alias{.constructModRanges,GRanges,DataFrame-method}
\title{RNAmodR internal functions}
\usage{
.constructModRanges(range, data, modType, scoreFun, source, type)

.dataTracks(x, data, seqdata, sequence, args)

\S4method{.constructModRanges}{GRanges,DataFrame}(range, data, modType,
  scoreFun, source, type)
}
\description{
These functions are not intended for general use, but maybe exported for 
additional package development.
}
