% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.R, R/AllGenerics.R,
%   R/Gviz-ModifiedSequenceTrack-class.R, R/Gviz-functions.R,
%   R/Modifier-utils.R, R/SequenceData-class.R, R/SequenceDataSet-class.R,
%   R/SequenceDataList-class.R
\docType{methods}
\name{RNAmodR-internals}
\alias{RNAmodR-internals}
\alias{.getData}
\alias{.constructModRanges}
\alias{show,ModifiedSequenceTrack-method}
\alias{length,ModifiedSequenceTrack-method}
\alias{.constructModRanges,GRanges,DataFrame-method}
\alias{parallelSlotNames,SequenceData-method}
\alias{[<-,SequenceData,ANY,ANY,ANY-method}
\alias{setListElement,SequenceData-method}
\alias{getListElement,SequenceData-method}
\alias{parallelSlotNames,SequenceDataSet-method}
\alias{getListElement,SequenceDataSet-method}
\alias{parallelSlotNames,SequenceDataList-method}
\alias{getListElement,SequenceDataList-method}
\title{RNAmodR internal functions}
\usage{
.constructModRanges(range, data, modType, scoreFun, source, type)

\S4method{show}{ModifiedSequenceTrack}(object)

\S4method{length}{ModifiedSequenceTrack}(x)

\S4method{.constructModRanges}{GRanges,DataFrame}(range, data, modType,
  scoreFun, source, type)

\S4method{parallelSlotNames}{SequenceData}(x)

\S4method{[}{SequenceData,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{setListElement}{SequenceData}(x, i, value)

\S4method{getListElement}{SequenceData}(x, i, exact = TRUE)

\S4method{parallelSlotNames}{SequenceDataSet}(x)

\S4method{getListElement}{SequenceDataSet}(x, i, exact = TRUE)

\S4method{parallelSlotNames}{SequenceDataList}(x)

\S4method{getListElement}{SequenceDataList}(x, i, exact = TRUE)
}
\arguments{
\item{object, range, data, modType, scoreFun, source, type, x, i, j, ..., exact, value}{Internal arguments}
}
\value{
internally used values
}
\description{
These functions are not intended for general use, but are used for 
additional package development.
}
