% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gviz-ModifiedSequenceTrack-class.R
\docType{class}
\name{RNASequenceTrack}
\alias{RNASequenceTrack}
\alias{RNASequenceTrack-class}
\title{RNASequenceTrack}
\usage{
RNASequenceTrack(sequence, chromosome, genome, ...)
}
\arguments{
\item{sequence}{A \code{character} vector or \code{RNAString} object of 
length one. The sequence to display. See 
\code{\link[Gviz:SequenceTrack-class]{SequenceTrack}}.}

\item{chromosome, genome, ...}{See 
\code{\link[Gviz:SequenceTrack-class]{SequenceTrack}}.}
}
\description{
A \code{Gviz} compatible 
\code{\link[Gviz:SequenceTrack-class]{SequenceTrack}} for showing RNA 
sequences.
}
\examples{
seq <- RNAStringSet(c(chr1 = "AGCUAGCUAGCUAGCUAGCUAGCU"))
st <- RNASequenceTrack(seq)
Gviz::plotTracks(st, chromosome = "chr1",from = 1L, to = 20L)
}
