% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-class.R
\name{SequenceData-class}
\alias{SequenceData-class}
\alias{SequenceData}
\title{The SequenceData class}
\arguments{
\item{dataType}{The prefix for construction the class name of the 
\code{SequenceData} subclass to be constructed.}

\item{bamfiles}{the input which can be of the following types
\itemize{
\item{\code{BamFileList}:} {a named \code{BamFileList}}
\item{\code{character}:} {a \code{character} vector, which must be coercible
to a named \code{BamFileList} referencing existing bam files. Valid names are
\code{control} and \code{treated} to define conditions and replicates}
}}

\item{annotation}{annotation data, which must match the information contained
in the BAM files.}

\item{sequences}{sequences matching the target sequences the reads were 
mapped onto. This must match the information contained in the BAM files.}

\item{seqinfo}{optional \code{\link[GenomeInfoDb:Seqinfo]{Seqinfo}} to 
subset the transcripts analyzed on a chromosome basis.}

\item{...}{Optional arguments overwriting default values, which are
\itemize{
\item{minQuality} {class dependent}
}}
}
\description{
The \code{SequenceData} class is a virtual class. It contains data on each
position alongside transcripts and holds the nucleotide sequence of these 
transcripts. To access this data several 
\code{\link[=SequenceData-functions]{functions}} are available.

It is derived from the 
\code{\link[IRanges:DataFrameList-class]{CompressedSplitDataFrameList}} 
class.
}
\section{Slots}{

\describe{
\item{\code{ranges}}{a \code{\link[GenomicRanges:GRangesList-class]{GRangesList}} 
object each element describing a transcript including its element. The 
\code{GRangesList} is constructed from the 
\code{\link[GenomicFeatures:transcriptsBy]{transcriptsBy(x, by="tx")}}.}

\item{\code{sequences}}{a \code{\link[Biostrings:XStringSet-class]{XStringSet}} of 
type \code{sequencesType}.}

\item{\code{sequencesType}}{a \code{character} value for the class name of 
\code{sequences}. Either \code{RNAStringSet} or \code{ModRNAStringSet}.}

\item{\code{bamfiles}}{the input bam files as 
\code{\link[Rsamtools:BamFileList]{BamFileList}}}

\item{\code{condition}}{conditions along the 
\code{\link[Rsamtools:BamFileList]{BamFileList}}: Either \code{control} or 
\code{treated}}

\item{\code{replicate}}{replicate number along the \code{BamFileList} for each of the
condition types.}

\item{\code{seqinfo}}{a \code{\link[GenomeInfoDb:Seqinfo-class]{Seqinfo}} describing
the avialable chromosomes. This is an intersection of}

\item{\code{minQuality}}{a \code{integer} value describing a threshold of the minimum
quality of reads to be used.}
}}

