% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceDataFrame-class.R,
%   R/SequenceData-class.R, R/SequenceDataList-class.R
\docType{methods}
\name{SequenceData-functions}
\alias{SequenceData-functions}
\alias{show,SequenceDataFrame-method}
\alias{sequences,SequenceDataFrame-method}
\alias{ranges,SequenceDataFrame-method}
\alias{show,SequenceData-method}
\alias{seqinfo,SequenceData-method}
\alias{sequences,SequenceData-method}
\alias{ranges,SequenceData-method}
\alias{bamfiles,SequenceData-method}
\alias{show,SequenceDataList-method}
\alias{names,SequenceDataList-method}
\alias{names<-,SequenceDataList,ANY-method}
\alias{seqinfo,SequenceDataList-method}
\alias{sequences,SequenceDataList-method}
\alias{ranges,SequenceDataList-method}
\alias{bamfiles,SequenceDataList-method}
\title{SequenceData/SequenceDataList/SequenceDataFrame functions}
\usage{
\S4method{show}{SequenceDataFrame}(object)

\S4method{sequences}{SequenceDataFrame}(x)

\S4method{ranges}{SequenceDataFrame}(x)

\S4method{show}{SequenceData}(object)

\S4method{seqinfo}{SequenceData}(x)

\S4method{sequences}{SequenceData}(x)

\S4method{ranges}{SequenceData}(x)

\S4method{bamfiles}{SequenceData}(x)

\S4method{show}{SequenceDataList}(object)

\S4method{names}{SequenceDataList}(x)

\S4method{names}{SequenceDataList,ANY}(x) <- value

\S4method{seqinfo}{SequenceDataList}(x)

\S4method{sequences}{SequenceDataList}(x)

\S4method{ranges}{SequenceDataList}(x)

\S4method{bamfiles}{SequenceDataList}(x)
}
\arguments{
\item{x, object, value}{a \code{SequenceData},\code{SequenceDataList},
\code{SequenceDataFrame} object.}
}
\description{
The \code{SequenceData},\code{SequenceDataList} and
\code{SequenceDataFrame} share functionality. Have a look at the elements
listed under Usage.
}
