% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceData-class.R
\docType{methods}
\name{SequenceData}
\alias{SequenceData}
\title{SequenceData}
\usage{
\S4method{parallelSlotNames}{SequenceData}(x)

\S4method{getListElement}{SequenceData}(x, i, exact = TRUE)

\S4method{setListElement}{SequenceData}(x, i, value)

\S4method{gff}{SequenceData}(x)

\S4method{fasta}{SequenceData}(x)

\S4method{sequences}{SequenceData}(x)

\S4method{ranges}{SequenceData}(x)

\S4method{bamfiles}{SequenceData}(x)

\S4method{aggregate}{SequenceData}(x)
}
\arguments{
\item{...}{Optional arguments overwriting default values, which are
\itemize{
\item{minQuality}{class dependent}
}}
}
\description{
title
}
