% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceDataFrame-class.R
\docType{class}
\name{SequenceDataFrame-class}
\alias{SequenceDataFrame-class}
\alias{SequenceDataFrame}
\alias{[,SequenceDataFrame,ANY,ANY,ANY-method}
\title{The SequenceDataFrame class}
\usage{
SequenceDataFrame(df, ranges, sequence, replicate, condition)

\S4method{[}{SequenceDataFrame,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{df}{the data as a \code{DataFrame}.}

\item{ranges}{a \code{GRanges} object containing all annotation elements
for a transcript.}

\item{sequence}{\code{XString} object describing the nucleotide sequence of 
the transcript.}

\item{replicate}{The replicate of each column or set of columns for the 
individual conditions}

\item{condition}{The condition of each column or set of columns. Either 
\code{control} or \code{treated}.}

\item{x, i, j, ..., drop}{arguments used for 
\code{\link[S4Vectors:DataFrame-class]{subsetting}}.}
}
\value{
a \code{SequenceDataFrame} object
}
\description{
The \code{SequenceDataFrame} class contains data for positions along a single
transcript. It is used to describe elements from a \code{SequenceData}
object.

The \code{SequenceDataFrame} class is derived from the
\code{\link[S4Vectors:DataFrame-class]{DataFrame}} class.

Subsetting of a \code{SequenceDataFrame} returns a \code{SequenceDataFrame} or 
\code{DataFrame}, if it is subset by a column or row, respectively. The 
\code{drop} argument is ignored for column subsetting.
}
\examples{
data(e5sd,package="RNAmodR")
# A SequenceDataFrame can is usually constructed by subsetting from 
# a SequenceData object
sdf <- e5sd[[1]]
}
