% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceDataFrame-class.R
\docType{class}
\name{SequenceDataFrame-class}
\alias{SequenceDataFrame-class}
\alias{SequenceDataFrame}
\alias{cbind,SequenceDataFrame-method}
\alias{[,SequenceDataFrame,ANY,ANY,ANY-method}
\title{The SequenceDataFrame class}
\usage{
\S4method{cbind}{SequenceDataFrame}(..., deparse.level = 1)

\S4method{[}{SequenceDataFrame,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x, i, j, ..., drop, deparse.level}{arguments used for 
\code{\link[S4Vectors:DataFrame-class]{subsetting}} or 
\code{\link[base:cbind]{base::cbind}}.}
}
\description{
The \code{SequenceDataFrame} class contains data for positions along a single
transcript. It is used to describe elements from a \code{SequenceData}
object.

The \code{SequenceDataFrame} class is derived from the
\code{\link[S4Vectors:DataFrame-class]{DataFrame}} class.

Subsetting of a \code{SequenceDataFrame} returns a \code{SequenceDataFrame} or 
\code{DataFrame}, if it is subset by a column or row, respectively. The 
\code{drop} argument is ignored for column subsetting.
}
\examples{
data(e5sd,package="RNAmodR")
# A SequenceDataFrame can is usually constructed by subsetting from 
# a SequenceData object
sdf <- e5sd[[1]]
}
