% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{SmartMap.ToTx}
\alias{SmartMap.ToTx}
\title{Map genome coordinates to transcript coordinates.}
\usage{
SmartMap.ToTx(locus, txBySec, seqBySec, geneXID, ignore.strand = FALSE,
  noFactors = TRUE, showPb = FALSE)
}
\arguments{
\item{locus}{A \code{GRanges} object; list of genomic features to
be mapped.}

\item{txBySec}{List of GRangesList transcript features.}

\item{seqBySec}{List of DNAStringSet sequences.}

\item{geneXID}{Dataframe of cross-referencing gene IDs.}

\item{ignore.strand}{Ignore strand when mapping genome coordinates to
transcript coordinates. Default is FALSE.}

\item{noFactors}{No (character) factors in output. Default is TRUE.}

\item{showPb}{A logical scalar; if \code{TRUE} show a progress bar.}
}
\value{
A list of dataframes with mapped features in transcript region.
}
\description{
Map genome coordinates to transcript coordinates. This function
should not be invoked by the end-user directly. It is called
from within \code{SmartMap}.
}
\keyword{internal}

