% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{SubsampleTxLoc}
\alias{SubsampleTxLoc}
\title{Subsample a \code{txLoc} object.}
\usage{
SubsampleTxLoc(txLoc, fractions, seed = NULL)
}
\arguments{
\item{txLoc}{A \code{txLoc} object; this is the \code{txLoc} object from
which the subsampled \code{txLoc} will be created.}

\item{fractions}{A \code{numeric} vector, specifying the fraction of sites
that will be sampled from every transcript region. Note that the length
of \code{fractions} has to match the length of \code{GetRegions(txLoc)}.}

\item{seed}{A single value, interpreted as an \code{integer}, or \code{NULL};
this is to ensure reproducibility when subsampling \code{txLoc2} sites; 
default is \code{NULL}.}
}
\value{
A \code{txLoc} object.
}
\description{
Subsample a \code{txLoc1} object based on a vector of fractions for every
transcript region.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
