% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-subset.R
\docType{methods}
\name{subsetByCoord}
\alias{subsetByCoord}
\alias{subsetByCoord,Modifier,GRanges-method}
\alias{subsetByCoord,Modifier,GRangesList-method}
\alias{subsetByCoord,ModifierSet,GRanges-method}
\alias{subsetByCoord,ModifierSet,GRangesList-method}
\title{subsetByCoord}
\usage{
\S4method{subsetByCoord}{Modifier,GRanges}(x, coord, ...)

\S4method{subsetByCoord}{Modifier,GRangesList}(x, coord, ...)

\S4method{subsetByCoord}{ModifierSet,GRanges}(x, coord, ...)

\S4method{subsetByCoord}{ModifierSet,GRangesList}(x, coord, ...)
}
\arguments{
\item{x}{a \code{Modifier} or \code{ModifierSet} object.}

\item{coord}{coordinates of position to subset to. Either a \code{GRanges} or
a \code{GRangesList} object. For both types the Parent column is expected to
match the gene or transcript name.}

\item{...}{optional parameters:
\itemize{
\item{\code{name}}{Limit results to one specific gene or transcript}
\item{\code{type}}{the modification type used for subsetting. By default this
is derived from the \code{modType(x)}, but it can be overwritten using 
\code{type}. It must be a valid shortName for a modification according to
\code{shortName(ModRNAString())} and of course present in metadata column 
\code{mod} of \code{coord}}
}}
}
\description{
title
}
