% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{txLoc-class}
\alias{txLoc-class}
\title{txLoc object definition}
\description{
txLoc object definition
}
\section{Slots}{

\describe{
\item{\code{loci}}{A list of loci within a transcript section.}

\item{\code{id}}{String. Identifier for sites in \code{txLoc} object.}

\item{\code{refGenome}}{String. Reference genome upon which the transcriptome
and site positions are based.}

\item{\code{version}}{String. Timestamp for version tracking.}
}}

