% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{txLoc-class}
\alias{GetNumberOfLoci}
\alias{GetNumberOfLoci,txLoc-method}
\alias{head,txLoc-method}
\alias{info}
\alias{info,txLoc-method}
\alias{txLoc-class}
\title{txLoc object.
The S4 object stores information about mapped loci per transcript
section. Meta-data such as an identifier and the reference genome
are stored in separate slots.}
\usage{
info(x)

\S4method{info}{txLoc}(x)

\S4method{head}{txLoc}(x, n = 6L, ...)

GetNumberOfLoci(x)

\S4method{GetNumberOfLoci}{txLoc}(x)
}
\arguments{
\item{x}{A \code{txLoc} object.}

\item{n}{Number of rows to be printed.}

\item{...}{Additional parameters passed to \code{head}.}

\item{x}{A \code{txLoc} object.}
}
\value{
A named integer vector with the number of sites per
transcript section.
}
\description{
txLoc object.
The S4 object stores information about mapped loci per transcript
section. Meta-data such as an identifier and the reference genome
are stored in separate slots.

Generic method "info" for S4 object txLoc.

Print general information of a \code{txLoc} object.

Method "head" for S4 object txLoc.

Generic method "GetNumberOfLoci" for S4 object txLoc.

Get the number of loci of a \code{txLoc} object in every
transcript section.
}
\section{Slots}{

\describe{
\item{\code{loci}}{A \code{list} of \code{GRangesList} objects; specifies
the list of loci per transcript section.}

\item{\code{id}}{A character string; identifier for loci in \code{txLoc}
object.}

\item{\code{refGenome}}{A character string; gives the reference genome
upon which transcriptome-derived positions are based.}

\item{\code{version}}{A character string; can be used to flag specific
version, e.g. using the current system time & date.}
}}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR");
sites <- ReadBED(bedFile);
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38");
info(posSites);
}

}

