% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{txLoc-class}
\alias{txLoc-class}
\title{txLoc object.
The S4 object stores information about mapped loci per transcript
section. Meta-data such as an identifier and the reference genome
are stored in separate slots.}
\description{
txLoc object.
The S4 object stores information about mapped loci per transcript
section. Meta-data such as an identifier and the reference genome
are stored in separate slots.
}
\section{Slots}{

\describe{
\item{\code{loci}}{A \code{list} of \code{dataframe} objects; specifies
the list of loci per transcript section.}

\item{\code{id}}{A character string; identifier for loci in \code{txLoc}
object.}

\item{\code{refGenome}}{A character string; gives the reference genome
upon which transcriptome-derived positions are based.}

\item{\code{version}}{A character string; can be used to flag specific
version, e.g. using the current system time & date.}
}}

\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}
