% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{info}
\alias{GetNumberOfLoci,txLoc-method}
\alias{head,txLoc-method}
\alias{info}
\alias{info,txLoc-method}
\title{Generic method "info" for S4 object txLoc.}
\usage{
info(x)

\S4method{info}{txLoc}(x)

\S4method{head}{txLoc}(x, n = 6L, ...)

\S4method{GetNumberOfLoci}{txLoc}(x)
}
\arguments{
\item{x}{A \code{txLoc} object.}

\item{n}{Number of rows to be printed.}

\item{...}{Additional parameters passed to \code{head}.}

\item{x}{A \code{txLoc} object.}
}
\value{
A named integer vector with the number of sites per
transcript section.
}
\description{
Generic method "info" for S4 object txLoc.

Print general information of a \code{txLoc} object.

Method "head" for S4 object txLoc.

Get the number of loci of a \code{txLoc} object in every
transcript section.
}
\examples{
\dontrun{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR");
sites <- ReadBED(bedFile);
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38");
info(posSites);
}

}

