% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{txLoc-class}
\alias{GetNumberOfLoci}
\alias{GetNumberOfLoci,txLoc-method}
\alias{head,txLoc-method}
\alias{info}
\alias{info,txLoc-method}
\alias{txLoc-class}
\title{txLoc object definition}
\usage{
info(x)

\S4method{info}{txLoc}(x)

\S4method{head}{txLoc}(x, n = 6L)

GetNumberOfLoci(x)

\S4method{GetNumberOfLoci}{txLoc}(x)
}
\arguments{
\item{x}{A \code{txLoc} object.}

\item{n}{Number of rows to be printed.}

\item{x}{A \code{txLoc} object.}

\item{x}{A \code{txLoc} object.}

\item{x}{A \code{txLoc} object.}

\item{x}{A \code{txLoc} object.}
}
\value{
A named integer vector with the number of sites per
transcript section.
}
\description{
txLoc object definition

Generic method "info" for S4 object txLoc.

Print general information of a \code{txLoc} object.

Method "head" for S4 object txLoc.

Generic method "GetNumberOfLoci" for S4 object txLoc.

Get the number of loci of a \code{txLoc} object in every
transcript section.
}
\section{Slots}{

\describe{
\item{\code{loci}}{A list of loci within a transcript section.}

\item{\code{id}}{Identifier for sites in \code{txLoc} object.}

\item{\code{refGenome}}{Reference genome upon which the transcriptome
and site positions are based.}

\item{\code{version}}{Timestamp for version tracking.}
}}
\examples{
bedFile <- system.file("extdata",
                       "miCLIP_m6A_Linder2015_hg38.bed",
                       package = "RNAModR");
sites <- ReadBED(bedFile);
posSites <- SmartMap(sites, id = "m6A", refGenome = "hg38");
info(posSites);

}

