% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{TxLoc2GRangesList}
\alias{TxLoc2GRangesList}
\title{Convert a \code{txLoc} object to a \code{GRangesList} object.}
\usage{
TxLoc2GRangesList(locus, filter = NULL)
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{filter}{A character vector; only keep transcript sections
specified in \code{filter}; if \code{NULL} consider all sections.}
}
\value{
A \code{GRangesList} object.
}
\description{
Convert a \code{txLoc} object to a \code{GRangesList} object.
}
\keyword{internal}

