% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Modifier-viz.R,
%   R/ModifierSet-viz.R, R/SequenceData-viz.R
\docType{methods}
\name{visualizeDataByCoord}
\alias{visualizeDataByCoord}
\alias{visualizeData}
\alias{visualizeDataByCoord,Modifier,GRanges-method}
\alias{visualizeData,Modifier-method}
\alias{visualizeDataByCoord,ModifierSet,GRanges-method}
\alias{visualizeData,ModifierSet-method}
\alias{visualizeDataByCoord,SequenceData,GRanges-method}
\alias{visualizeData,SequenceData-method}
\title{visualizeData}
\usage{
visualizeDataByCoord(x, coord, type, window.size = 15L, ...)

visualizeData(x, name, from = 1L, to = 30L, type, ...)

\S4method{visualizeDataByCoord}{Modifier,GRanges}(x, coord, type = NA,
  window.size = 15L, ...)

\S4method{visualizeData}{Modifier}(x, name, from = 1L, to = 30L,
  type = NA, ...)

\S4method{visualizeDataByCoord}{ModifierSet,GRanges}(x, coord, type = NA,
  window.size = 15L, ...)

\S4method{visualizeData}{ModifierSet}(x, name, from = 1L, to = 30L,
  type = NA, ...)

\S4method{visualizeDataByCoord}{SequenceData,GRanges}(x, coord,
  type = NA, window.size = 15L, ...)

\S4method{visualizeData}{SequenceData}(x, name, from = 1L, to = 30L,
  type = NA, ...)
}
\arguments{
\item{x}{a \code{Modifier} or \code{ModifierSet} object.}

\item{coord}{coordinates of a single position to subset to as a 
\code{GRanges} object. The Parent column is expected to match the gene or 
transcript name.}

\item{type}{the data type of data show as data tracks.}

\item{window.size}{integer value for the number of positions on the left and 
right site of the selected positions included in the plotting (default: 
\code{window.size = 15L})}

\item{...}{optional parameters:
\itemize{
\item{\code{modified.seq}}{\code{TRUE} or \code{FALSE}. Should the sequence 
shown with modified nucleotide positions? (default: 
\code{modified.seq = FALSE})}
\item{\code{additional.mod}}{other modifications, which should be shown
in the annotation and sequence track.}
\item{\code{annotation.track.pars}}{Parameters passed onto the A
nnotationTrack.}
\item{\code{sequence.track.pars}}{Parameters passed onto the SequenceTrack.}
\item{\code{data.track.pars}}{Parameters passed onto the DataTrack(s).}
}}
}
\description{
title
}
