% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Modifier-viz.R,
%   R/ModifierSet-viz.R, R/SequenceData-viz.R, R/SequenceDataList-viz.R
\docType{methods}
\name{visualizeDataByCoord}
\alias{visualizeDataByCoord}
\alias{visualizeData}
\alias{getDataTrack}
\alias{visualizeDataByCoord,Modifier,GRanges-method}
\alias{visualizeData,Modifier-method}
\alias{getDataTrack,Modifier-method}
\alias{visualizeDataByCoord,ModifierSet,GRanges-method}
\alias{visualizeData,ModifierSet-method}
\alias{visualizeDataByCoord,SequenceData,GRanges-method}
\alias{visualizeData,SequenceData-method}
\alias{getDataTrack,SequenceData-method}
\alias{getDataTrack,SequenceDataList-method}
\alias{visualizeDataByCoord,SequenceDataList,GRanges-method}
\alias{visualizeData,SequenceDataList-method}
\title{visualizeData}
\usage{
visualizeDataByCoord(x, coord, type, window.size = 15L, ...)

visualizeData(x, name, from = 1L, to = 30L, type, ...)

getDataTrack(x, name, ...)

\S4method{visualizeDataByCoord}{Modifier,GRanges}(x, coord, type = NA,
  seqdata = FALSE, window.size = 15L, ...)

\S4method{visualizeData}{Modifier}(x, name, from, to, type = NA,
  seqdata = FALSE, ...)

\S4method{getDataTrack}{Modifier}(x, name = name, ...)

\S4method{visualizeDataByCoord}{ModifierSet,GRanges}(x, coord, type = NA,
  seqdata = FALSE, window.size = 15L, ...)

\S4method{visualizeData}{ModifierSet}(x, name, from, to, type = NA,
  seqdata = FALSE, ...)

\S4method{visualizeDataByCoord}{SequenceData,GRanges}(x, coord,
  type = NA, window.size = 15L, perTranscript = FALSE, ...)

\S4method{visualizeData}{SequenceData}(x, name, from, to,
  perTranscript = FALSE, ...)

\S4method{getDataTrack}{SequenceData}(x, name = name, ...)

\S4method{getDataTrack}{SequenceDataList}(x, name = name, ...)

\S4method{visualizeDataByCoord}{SequenceDataList,GRanges}(x, coord,
  type = NA, window.size = 15L, perTranscript = FALSE, ...)

\S4method{visualizeData}{SequenceDataList}(x, name, from, to,
  perTranscript = FALSE, ...)
}
\arguments{
\item{x}{a \code{Modifier} or \code{ModifierSet} object.}

\item{coord}{coordinates of a positions to subset to as a 
\code{GRanges} object. The Parent column is expected to match the gene or 
transcript name.}

\item{type}{the data type of data show as data tracks.}

\item{window.size}{integer value for the number of positions on the left and 
right site of the selected positions included in the plotting (default: 
\code{window.size = 15L})}

\item{...}{optional parameters:
\itemize{
\item{\code{modified.seq}} {\code{TRUE} or \code{FALSE}. Should the sequence 
shown with modified nucleotide positions? (default: 
\code{modified.seq = FALSE})}
\item{\code{additional.mod}} {other modifications, which should be shown
in the annotation and sequence track. The must be a \code{GRanges} compatible
with \code{\link[Modstrings:separate]{combineIntoModstrings}}.}
\item{\code{annotation.track.pars}} {Parameters passed onto the 
\code{\link[Gviz:AnnotationTrack-class]{AnnotationTrack}}.}
\item{\code{sequence.track.pars}} {Parameters passed onto the 
\code{\link[Gviz:SequenceTrack-class]{SequenceTrack}}.}
}}

\item{name}{Only for \code{visualizeData}: the transcript name}

\item{from}{Only for \code{visualizeData}: start position}

\item{to}{Only for \code{visualizeData}: end position}

\item{seqdata}{\code{TRUE} or \code{FALSE}: whould the sequence data be 
shown? (default: \code{seqdata = FALSE})}

\item{perTranscript}{\code{TRUE} or \code{FALSE}: Should the positions shown
per transcript? (default: \code{perTranscript = FALSE})}
}
\value{
a plot send to the active graphic device
}
\description{
title
}
\examples{
data(msi,package="RNAmodR")
visualizeData(msi, "2", from = 10L, to = 45L)
}
