% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{WriteFeatToBED}
\alias{WriteFeatToBED}
\title{Write a \code{list} of \code{GRangesList} to a BED file.}
\usage{
WriteFeatToBED(txFeatures)
}
\arguments{
\item{txFeatures}{A \code{list} of \code{GRangesList} objects.}
}
\description{
Write a \code{list} of \code{GRangesList} to a BED file.
One file per transcript region is generated.
}

