% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{WriteTxLocToBED}
\alias{WriteTxLocToBED}
\title{Write \code{txLoc} object to a BED file.}
\usage{
WriteTxLocToBED(locus, file = NULL, formatChr = "noChrName")
}
\arguments{
\item{locus}{A \code{txLoc} object.}

\item{file}{Filename of output BED file. If NULL then file = "sites.bed".}

\item{formatChr}{Set output format of chromosome column. Default is
"noChrName".}
}
\description{
Write \code{txLoc} object to a BED file.
}
\details{
This function writes entries from a \code{txLoc} object to a BED file.
This process is not "splice-aware", i.e. if an entry spans an intron
the BED entry gives the left and right-most genomic coordinate of the
feature.
}
\author{
Maurits Evers, \email{maurits.evers@anu.edu.au}
}

